refactor: split ReportCard.tsx and App.css into per-section modules
Move-only, no behavior change (roadmap item 12).

ReportCard.tsx (3,242 lines) -> src/report/: icons, common, NightTimeline,
Satellites, glow, MilkyWay, Targets, Nearby, LightDomePanel; ReportCard keeps
the caches and the main component (~680 lines). Moved declarations gained
`export`; the module graph is acyclic (verified).

App.css (3,642 lines) -> src/styles/01-base … 11-responsive, imported in
order by an App.css @import shell; the concatenation of the chunks is
byte-identical to the original file (modulo end-of-file newlines), so the
cascade is unchanged.

Verified: tsc -b + vite build clean; instrumented remount in the browser shows
zero console errors; section counts, sticky nav, red-mode palette, and score
render identically to pre-split.
