+++ apps/web/src/ReportCard.tsx
@@ -1,22 +1,18 @@
-import React, { useState, useRef, useEffect, useMemo } from 'react'
-import type { NightReport, SkyEvent, WeatherPoint, VisibleTarget, TargetWindow, MilkyWaySummary, NearbyResult, NearbyPlace, LightDomeSummary, Direction, CalendarResult } from './types'
-import {
-  formatTime, formatDayTime, formatHm, tzAbbr, tzTitle,
-  cardinal, rateConditions, fmtTempValue, tempUnitLabel, fmtWindSpeed, windUnitLabel, fmtDist, lpString,
-  scoreBand, scoreLabel, moonWashSeverity, moonUpAt, tonightIso, availabilityFor,
-  formatAge, formatIssuedUtc, nightVerdict,
-} from './format'
+import { useState, useRef, useEffect, useMemo } from 'react'
+import type { NightReport, NearbyResult, CalendarResult } from './types'
+import { formatTime, formatHm, tzAbbr, tzTitle, fmtDist, lpString, scoreBand, scoreLabel, tonightIso, availabilityFor, nightVerdict } from './format'
+import { MoonPhaseSvg, ScoreBar, InfoTip } from './shared'
 import { fetchNearby, fetchCalendar, fetchNightDateOnly, ApiRequestError } from './api'
 import OutlookTelemetryRibbon from './OutlookTelemetryRibbon'
 import CalendarRangePicker, { type CalendarPickerState } from './CalendarRangePicker'
-import { MoonPhaseSvg, ScoreBar, InfoTip } from './shared'
-import {
-  Star,
-  Navigation,
-} from 'lucide-react'
+import { fmtNightDate, MetaRow } from './report/common'
+import { WeatherTable } from './report/NightTimeline'
+import { SatellitePasses } from './report/Satellites'
+import { MilkyWayAbsent, MilkyWayCard } from './report/MilkyWay'
+import { isPrime, MeteorShowerCard, TargetsTable } from './report/Targets'
+import { NearbyResults } from './report/Nearby'
+import { LightDomePanel } from './report/LightDomePanel'
 
-// Mirrors apps/api/main.py's _MAX_CALENDAR_DAYS — the outlook auto-loads at the
-// full cap so the user never has to open the range picker just to see it.
 const AUTO_CALENDAR_DAYS = 30
 
 // ReportCard fully unmounts and remounts on every /night fetch, including
@@ -65,2566 +61,6 @@ function fetchNearbyCached(lat: number, lon: number, radius: number): Promise<Ne
   }
   _nearbyLastRadius.set(nearbyLocKey(lat, lon), radius)
   return p
-}
-
-// ── WMO weather code icons ───────────────────────────────────────────────────
-
-// ── weather-icons glyphs (erikflowers/weather-icons) ────────────────────────
-// Hand-vendored path data (fill="currentColor", native 0 0 30 30 viewBox) — no new npm
-// dependency, same one-off approach the codebase already used for the CloudStar shape
-// before this migration.
-const WI_PATHS: Record<string, string> = {
-  "cloud": "M4.61,16.88c0-1.15,0.36-2.17,1.08-3.07c0.72-0.9,1.63-1.48,2.74-1.73c0.31-1.37,1.02-2.49,2.11-3.37s2.35-1.32,3.76-1.32 c1.38,0,2.61,0.43,3.69,1.28s1.78,1.95,2.1,3.29h0.33c0.9,0,1.73,0.22,2.49,0.65s1.37,1.03,1.81,1.79c0.44,0.76,0.67,1.58,0.67,2.48 c0,0.88-0.21,1.7-0.63,2.45s-1,1.35-1.73,1.8c-0.73,0.45-1.54,0.69-2.41,0.72H9.41c-1.34-0.06-2.47-0.57-3.4-1.53 C5.08,19.37,4.61,18.22,4.61,16.88z M6.32,16.88c0,0.87,0.3,1.62,0.9,2.26s1.33,0.98,2.19,1.03h11.19c0.86-0.04,1.59-0.39,2.19-1.03 c0.61-0.64,0.91-1.4,0.91-2.26c0-0.88-0.33-1.63-0.98-2.27c-0.65-0.64-1.42-0.96-2.32-0.96H18.8c-0.11,0-0.17-0.06-0.17-0.18 l-0.07-0.57c-0.11-1.08-0.58-1.99-1.4-2.72c-0.82-0.73-1.77-1.1-2.86-1.1c-1.09,0-2.05,0.37-2.85,1.1 c-0.81,0.73-1.27,1.64-1.37,2.72l-0.08,0.57c0,0.12-0.07,0.18-0.2,0.18H9.27c-0.84,0.1-1.54,0.46-2.1,1.07S6.32,16.05,6.32,16.88z",
-  "cloudy": "M3.89,17.6c0-0.99,0.31-1.88,0.93-2.65s1.41-1.27,2.38-1.49c0.26-1.17,0.85-2.14,1.78-2.88c0.93-0.75,2-1.12,3.22-1.12 c1.18,0,2.24,0.36,3.16,1.09c0.93,0.73,1.53,1.66,1.8,2.8h0.27c1.18,0,2.18,0.41,3.01,1.24s1.25,1.83,1.25,3 c0,1.18-0.42,2.18-1.25,3.01s-1.83,1.25-3.01,1.25H8.16c-0.58,0-1.13-0.11-1.65-0.34S5.52,21,5.14,20.62 c-0.38-0.38-0.68-0.84-0.91-1.36S3.89,18.17,3.89,17.6z M5.34,17.6c0,0.76,0.28,1.42,0.82,1.96s1.21,0.82,1.99,0.82h9.28 c0.77,0,1.44-0.27,1.99-0.82c0.55-0.55,0.83-1.2,0.83-1.96c0-0.76-0.27-1.42-0.83-1.96c-0.55-0.54-1.21-0.82-1.99-0.82h-1.39 c-0.1,0-0.15-0.05-0.15-0.15l-0.07-0.49c-0.1-0.94-0.5-1.73-1.19-2.35s-1.51-0.93-2.45-0.93c-0.94,0-1.76,0.31-2.46,0.94 c-0.7,0.62-1.09,1.41-1.18,2.34l-0.07,0.42c0,0.1-0.05,0.15-0.16,0.15l-0.45,0.07c-0.72,0.06-1.32,0.36-1.81,0.89 C5.59,16.24,5.34,16.87,5.34,17.6z M14.19,8.88c-0.1,0.09-0.08,0.16,0.07,0.21c0.43,0.19,0.79,0.37,1.08,0.55 c0.11,0.03,0.19,0.02,0.22-0.03c0.61-0.57,1.31-0.86,2.12-0.86c0.81,0,1.5,0.27,2.1,0.81c0.59,0.54,0.92,1.21,0.99,2l0.09,0.64h1.42 c0.65,0,1.21,0.23,1.68,0.7c0.47,0.47,0.7,1.02,0.7,1.66c0,0.6-0.21,1.12-0.62,1.57s-0.92,0.7-1.53,0.77c-0.1,0-0.15,0.05-0.15,0.16 v1.13c0,0.11,0.05,0.16,0.15,0.16c1.01-0.06,1.86-0.46,2.55-1.19s1.04-1.6,1.04-2.6c0-1.06-0.37-1.96-1.12-2.7 c-0.75-0.75-1.65-1.12-2.7-1.12h-0.15c-0.26-1-0.81-1.82-1.65-2.47c-0.83-0.65-1.77-0.97-2.8-0.97C16.28,7.29,15.11,7.82,14.19,8.88 z",
-  "cloudy-gusts": "M3.62,21.01c0-0.25,0.08-0.46,0.25-0.63c0.17-0.16,0.37-0.24,0.6-0.24h5.42c0.74,0,1.37,0.26,1.89,0.79 c0.52,0.53,0.78,1.16,0.78,1.9c0,0.74-0.26,1.38-0.78,1.9c-0.52,0.52-1.15,0.78-1.89,0.78s-1.38-0.26-1.9-0.79 c-0.16-0.16-0.23-0.36-0.23-0.6c0-0.24,0.08-0.45,0.23-0.6c0.15-0.16,0.35-0.24,0.6-0.24c0.23,0,0.43,0.08,0.61,0.24 c0.2,0.19,0.43,0.29,0.69,0.29s0.49-0.1,0.68-0.29c0.19-0.19,0.29-0.42,0.29-0.7c0-0.26-0.1-0.49-0.29-0.68s-0.42-0.29-0.68-0.29 H4.47c-0.23,0-0.43-0.08-0.6-0.25S3.62,21.25,3.62,21.01z M3.62,17.97c0-0.24,0.08-0.45,0.25-0.62c0.17-0.16,0.37-0.24,0.6-0.24 h10.55c0.26,0,0.49-0.1,0.68-0.29c0.19-0.19,0.29-0.43,0.29-0.69s-0.1-0.5-0.29-0.69c-0.19-0.19-0.42-0.29-0.68-0.29 c-0.28,0-0.5,0.09-0.68,0.28c-0.18,0.15-0.39,0.23-0.64,0.23c-0.24,0-0.44-0.08-0.6-0.23c-0.15-0.15-0.23-0.35-0.23-0.6 c0-0.25,0.07-0.45,0.23-0.61c0.51-0.51,1.15-0.76,1.92-0.76c0.74,0,1.38,0.26,1.9,0.78c0.52,0.52,0.78,1.15,0.78,1.88 s-0.26,1.37-0.78,1.89c-0.52,0.52-1.15,0.78-1.9,0.78H4.47c-0.24,0-0.44-0.08-0.6-0.24C3.7,18.4,3.62,18.2,3.62,17.97z M5.77,15.61 c0,0.08,0.05,0.12,0.16,0.12h1.44c0.08,0,0.15-0.05,0.22-0.15c0.22-0.54,0.58-0.99,1.05-1.35c0.48-0.36,1.01-0.56,1.59-0.6 l0.53-0.08c0.13,0,0.2-0.06,0.2-0.17l0.07-0.52c0.11-1.08,0.56-1.99,1.37-2.72s1.76-1.1,2.86-1.1c1.11,0,2.07,0.36,2.88,1.09 c0.81,0.73,1.27,1.64,1.39,2.73l0.07,0.59c0,0.11,0.06,0.17,0.17,0.17h1.62c0.91,0,1.68,0.32,2.33,0.96c0.65,0.64,0.97,1.4,0.97,2.3 c0,0.89-0.32,1.66-0.97,2.3c-0.65,0.64-1.42,0.96-2.33,0.96h-6.91c-0.12,0-0.19,0.06-0.19,0.17v1.39c0,0.11,0.06,0.17,0.19,0.17 h6.91c0.91,0,1.74-0.22,2.51-0.67c0.77-0.44,1.37-1.05,1.82-1.81c0.45-0.77,0.67-1.6,0.67-2.5c0-0.91-0.22-1.74-0.67-2.5 c-0.45-0.76-1.05-1.37-1.82-1.81c-0.77-0.44-1.6-0.67-2.51-0.67h-0.31c-0.31-1.33-1.01-2.42-2.1-3.27 c-1.08-0.85-2.33-1.27-3.73-1.27c-1.41,0-2.66,0.44-3.75,1.32s-1.78,2-2.07,3.37c-0.86,0.2-1.62,0.61-2.28,1.23 s-1.12,1.36-1.38,2.21v0.04C5.77,15.56,5.77,15.58,5.77,15.61z",
-  "cloudy-windy": "M3.1,21.04c0-0.24,0.08-0.45,0.25-0.61s0.38-0.24,0.63-0.24h8.97c0.24,0,0.43,0.08,0.59,0.24c0.16,0.16,0.23,0.36,0.23,0.61 c0,0.24-0.08,0.44-0.24,0.6c-0.16,0.16-0.35,0.24-0.59,0.24H3.98c-0.25,0-0.46-0.08-0.63-0.24S3.1,21.27,3.1,21.04z M5.73,17.98 c0-0.24,0.09-0.44,0.27-0.6c0.14-0.15,0.34-0.23,0.59-0.23h9c0.23,0,0.42,0.08,0.58,0.23s0.23,0.35,0.23,0.59 c0,0.24-0.08,0.44-0.23,0.61c-0.15,0.17-0.35,0.25-0.58,0.25h-9c-0.23,0-0.43-0.09-0.6-0.26S5.73,18.21,5.73,17.98z M6.35,15.65 c0,0.09,0.06,0.14,0.17,0.14h1.43c0.09,0,0.17-0.05,0.23-0.14c0.23-0.54,0.57-0.99,1.04-1.35s0.99-0.56,1.58-0.6l0.54-0.07 c0.11,0,0.17-0.06,0.17-0.18l0.07-0.52c0.11-1.09,0.58-1.99,1.39-2.72c0.82-0.73,1.77-1.09,2.87-1.09c1.09,0,2.03,0.36,2.83,1.07 c0.8,0.72,1.27,1.62,1.41,2.7l0.07,0.59c0,0.11,0.06,0.16,0.18,0.16h1.6c0.91,0,1.68,0.32,2.32,0.96c0.64,0.64,0.96,1.41,0.96,2.32 c0,0.88-0.33,1.64-0.97,2.28c-0.65,0.65-1.42,0.97-2.31,0.97h-6.89c-0.12,0-0.18,0.06-0.18,0.17v1.34c0,0.12,0.06,0.18,0.18,0.18 h6.89c0.68,0,1.32-0.13,1.94-0.39s1.14-0.61,1.58-1.05s0.79-0.97,1.05-1.58s0.39-1.25,0.39-1.92c0-0.9-0.22-1.73-0.66-2.49 c-0.44-0.76-1.04-1.36-1.8-1.8c-0.76-0.44-1.6-0.66-2.5-0.66h-0.31c-0.33-1.34-1.03-2.44-2.1-3.3c-1.08-0.85-2.3-1.28-3.68-1.28 c-1.42,0-2.67,0.44-3.76,1.33c-1.09,0.88-1.78,2.01-2.08,3.39c-0.86,0.19-1.62,0.6-2.27,1.21s-1.1,1.35-1.36,2.22v0.02 C6.36,15.6,6.35,15.62,6.35,15.65z M7.5,24.13c0-0.24,0.09-0.44,0.26-0.6c0.15-0.16,0.35-0.23,0.59-0.23h8.99 c0.24,0,0.45,0.08,0.61,0.24c0.17,0.16,0.25,0.36,0.25,0.6c0,0.24-0.08,0.44-0.25,0.61c-0.17,0.17-0.37,0.25-0.61,0.25H8.35 c-0.23,0-0.43-0.08-0.6-0.25C7.58,24.57,7.5,24.37,7.5,24.13z",
-  "fog": "M2.62,21.05c0-0.24,0.08-0.45,0.25-0.61c0.17-0.16,0.38-0.24,0.63-0.24h18.67c0.25,0,0.45,0.08,0.61,0.24 c0.16,0.16,0.24,0.36,0.24,0.61c0,0.23-0.08,0.43-0.25,0.58c-0.17,0.16-0.37,0.23-0.6,0.23H3.5c-0.25,0-0.46-0.08-0.63-0.23 C2.7,21.47,2.62,21.28,2.62,21.05z M5.24,17.91c0-0.24,0.09-0.44,0.26-0.6c0.15-0.15,0.35-0.23,0.59-0.23h18.67 c0.23,0,0.42,0.08,0.58,0.24c0.16,0.16,0.23,0.35,0.23,0.59c0,0.24-0.08,0.44-0.23,0.6c-0.16,0.17-0.35,0.25-0.58,0.25H6.09 c-0.24,0-0.44-0.08-0.6-0.25C5.32,18.34,5.24,18.14,5.24,17.91z M5.37,15.52c0,0.09,0.05,0.13,0.15,0.13h1.43 c0.06,0,0.13-0.05,0.2-0.16c0.24-0.52,0.59-0.94,1.06-1.27c0.47-0.33,0.99-0.52,1.55-0.56l0.55-0.07c0.11,0,0.17-0.06,0.17-0.18 l0.07-0.5c0.11-1.08,0.56-1.98,1.37-2.7c0.81-0.72,1.76-1.08,2.85-1.08c1.08,0,2.02,0.36,2.83,1.07c0.8,0.71,1.26,1.61,1.37,2.68 l0.08,0.57c0,0.11,0.07,0.17,0.2,0.17h1.59c0.64,0,1.23,0.17,1.76,0.52s0.92,0.8,1.18,1.37c0.07,0.11,0.14,0.16,0.21,0.16h1.43 c0.12,0,0.17-0.07,0.14-0.23c-0.29-1.02-0.88-1.86-1.74-2.51c-0.87-0.65-1.86-0.97-2.97-0.97h-0.32c-0.33-1.33-1.03-2.42-2.1-3.27 s-2.28-1.27-3.65-1.27c-1.4,0-2.64,0.44-3.73,1.32s-1.78,2-2.09,3.36c-0.85,0.2-1.6,0.6-2.24,1.21c-0.64,0.61-1.09,1.33-1.34,2.18 v-0.04C5.37,15.45,5.37,15.48,5.37,15.52z M6.98,24.11c0-0.24,0.09-0.43,0.26-0.59c0.15-0.15,0.35-0.23,0.6-0.23h18.68 c0.24,0,0.44,0.08,0.6,0.23c0.17,0.16,0.25,0.35,0.25,0.58c0,0.24-0.08,0.44-0.25,0.61c-0.17,0.17-0.37,0.25-0.6,0.25H7.84 c-0.23,0-0.43-0.09-0.6-0.26C7.07,24.55,6.98,24.34,6.98,24.11z",
-  "moonrise": "M7.8,14.86c0-0.98,0.19-1.92,0.58-2.82c0.38-0.9,0.9-1.67,1.55-2.32c0.65-0.65,1.43-1.17,2.32-1.56s1.84-0.58,2.83-0.58 h1.17c0.16,0.04,0.24,0.14,0.24,0.28l0.04,0.9c0.04,1.3,0.51,2.41,1.41,3.33s2,1.41,3.28,1.46l0.85,0.07c0.16,0,0.23,0.08,0.23,0.23 v1.01c0.01,1.03-0.19,2-0.58,2.92h-2.05c0.51-0.74,0.83-1.59,0.97-2.53c-1.67-0.35-3.02-1.07-4.03-2.16s-1.59-2.37-1.75-3.83 c-0.97,0.05-1.85,0.35-2.66,0.9c-0.8,0.55-1.43,1.24-1.87,2.08c-0.44,0.84-0.66,1.72-0.66,2.63c0,1.07,0.28,2.04,0.83,2.92H8.4 C8,16.85,7.8,15.88,7.8,14.86z M8.09,20.87c0-0.29,0.09-0.52,0.28-0.68c0.18-0.18,0.41-0.26,0.69-0.26h2.63L14.8,17 c0.1-0.08,0.22-0.08,0.35,0l3.16,2.92h2.77c0.27,0,0.5,0.09,0.69,0.28c0.19,0.18,0.29,0.41,0.29,0.67c0,0.27-0.1,0.5-0.29,0.69 c-0.19,0.19-0.42,0.29-0.69,0.29h-3.38c-0.1,0-0.2-0.02-0.29-0.07l-2.41-2.27l-2.39,2.27c-0.08,0.05-0.17,0.07-0.28,0.07H9.06 c-0.27,0-0.5-0.1-0.69-0.29S8.09,21.14,8.09,20.87z",
-  "moonset": "M7.74,14.86c0-0.98,0.19-1.92,0.58-2.82c0.38-0.9,0.9-1.67,1.55-2.32c0.65-0.65,1.43-1.17,2.32-1.56 c0.9-0.39,1.84-0.58,2.83-0.58h1.17c0.16,0.04,0.24,0.14,0.24,0.28l0.05,0.9c0.02,0.64,0.15,1.25,0.4,1.83s0.58,1.08,1,1.5 c0.42,0.43,0.91,0.77,1.48,1.03c0.57,0.26,1.17,0.4,1.8,0.43l0.85,0.07c0.16,0,0.24,0.08,0.24,0.23v1.01 c0.01,1.01-0.19,1.98-0.59,2.92h-2.05c0.51-0.74,0.83-1.59,0.97-2.53c-1.68-0.35-3.02-1.07-4.03-2.16s-1.59-2.37-1.75-3.83 c-0.97,0.05-1.85,0.35-2.66,0.9c-0.8,0.55-1.42,1.24-1.87,2.08c-0.44,0.84-0.66,1.72-0.66,2.63c0,1.07,0.28,2.04,0.83,2.92H8.34 C7.94,16.85,7.74,15.88,7.74,14.86z M7.99,20.89c0-0.26,0.1-0.49,0.3-0.69c0.18-0.18,0.41-0.27,0.68-0.27h3.22 c0.11,0,0.2,0.02,0.28,0.08l2.35,2.22L17.21,20c0.07-0.05,0.17-0.08,0.29-0.08h3.3c0.27,0,0.5,0.09,0.69,0.28 c0.19,0.19,0.29,0.42,0.29,0.68c0,0.27-0.1,0.5-0.29,0.69c-0.19,0.19-0.42,0.29-0.69,0.29h-2.68l-3.13,2.84 c-0.12,0.09-0.24,0.09-0.34,0l-3.08-2.84h-2.6c-0.27,0-0.5-0.1-0.69-0.29C8.09,21.39,7.99,21.16,7.99,20.89z",
-  "night-clear": "M7.91,14.48c0-0.96,0.19-1.87,0.56-2.75s0.88-1.63,1.51-2.26c0.63-0.63,1.39-1.14,2.27-1.52c0.88-0.38,1.8-0.57,2.75-0.57 h1.14c0.16,0.04,0.23,0.14,0.23,0.28l0.05,0.88c0.04,1.27,0.49,2.35,1.37,3.24c0.88,0.89,1.94,1.37,3.19,1.42l0.82,0.07 c0.16,0,0.24,0.08,0.24,0.23v0.98c0.01,1.28-0.3,2.47-0.93,3.56c-0.63,1.09-1.48,1.95-2.57,2.59c-1.08,0.63-2.27,0.95-3.55,0.95 c-0.97,0-1.9-0.19-2.78-0.56s-1.63-0.88-2.26-1.51c-0.63-0.63-1.13-1.39-1.5-2.26C8.1,16.37,7.91,15.45,7.91,14.48z M9.74,14.48 c0,0.76,0.15,1.48,0.45,2.16c0.3,0.67,0.7,1.24,1.19,1.7c0.49,0.46,1.05,0.82,1.69,1.08c0.63,0.27,1.28,0.4,1.94,0.4 c0.58,0,1.17-0.11,1.76-0.34c0.59-0.23,1.14-0.55,1.65-0.96c0.51-0.41,0.94-0.93,1.31-1.57c0.37-0.64,0.6-1.33,0.71-2.09 c-1.63-0.34-2.94-1.04-3.92-2.1s-1.55-2.3-1.7-3.74C13.86,9.08,13,9.37,12.21,9.9c-0.78,0.53-1.39,1.2-1.82,2.02 C9.96,12.74,9.74,13.59,9.74,14.48z",
-  "rain": "M4.64,16.91c0-1.15,0.36-2.17,1.08-3.07c0.72-0.9,1.63-1.47,2.73-1.73c0.31-1.36,1.02-2.48,2.11-3.36s2.34-1.31,3.75-1.31 c1.38,0,2.6,0.43,3.68,1.28c1.08,0.85,1.78,1.95,2.1,3.29h0.32c0.89,0,1.72,0.22,2.48,0.65s1.37,1.03,1.81,1.78 c0.44,0.75,0.67,1.58,0.67,2.47c0,0.88-0.21,1.69-0.63,2.44c-0.42,0.75-1,1.35-1.73,1.8c-0.73,0.45-1.53,0.69-2.4,0.71 c-0.13,0-0.2-0.06-0.2-0.17v-1.33c0-0.12,0.07-0.18,0.2-0.18c0.85-0.04,1.58-0.38,2.18-1.02s0.9-1.39,0.9-2.26s-0.33-1.62-0.98-2.26 s-1.42-0.96-2.31-0.96h-1.61c-0.12,0-0.18-0.06-0.18-0.17l-0.08-0.58c-0.11-1.08-0.58-1.99-1.39-2.71 c-0.82-0.73-1.76-1.09-2.85-1.09c-1.09,0-2.05,0.36-2.85,1.09c-0.81,0.73-1.26,1.63-1.36,2.71l-0.07,0.53c0,0.12-0.07,0.19-0.2,0.19 l-0.53,0.03c-0.83,0.1-1.53,0.46-2.1,1.07s-0.85,1.33-0.85,2.16c0,0.87,0.3,1.62,0.9,2.26s1.33,0.98,2.18,1.02 c0.11,0,0.17,0.06,0.17,0.18v1.33c0,0.11-0.06,0.17-0.17,0.17c-1.34-0.06-2.47-0.57-3.4-1.53S4.64,18.24,4.64,16.91z M9.99,23.6 c0-0.04,0.01-0.11,0.04-0.2l1.63-5.77c0.06-0.19,0.17-0.34,0.32-0.44c0.15-0.1,0.31-0.15,0.46-0.15c0.07,0,0.15,0.01,0.24,0.03 c0.24,0.04,0.42,0.17,0.54,0.37c0.12,0.2,0.15,0.42,0.08,0.67l-1.63,5.73c-0.12,0.43-0.4,0.64-0.82,0.64 c-0.04,0-0.07-0.01-0.11-0.02c-0.06-0.02-0.09-0.03-0.1-0.03c-0.22-0.06-0.38-0.17-0.49-0.33C10.04,23.93,9.99,23.77,9.99,23.6z M12.61,26.41l2.44-8.77c0.04-0.19,0.14-0.34,0.3-0.44c0.16-0.1,0.32-0.15,0.49-0.15c0.09,0,0.18,0.01,0.27,0.03 c0.22,0.06,0.38,0.19,0.49,0.39c0.11,0.2,0.13,0.41,0.07,0.64l-2.43,8.78c-0.04,0.17-0.13,0.31-0.29,0.43 c-0.16,0.12-0.32,0.18-0.51,0.18c-0.09,0-0.18-0.02-0.25-0.05c-0.2-0.05-0.37-0.18-0.52-0.39C12.56,26.88,12.54,26.67,12.61,26.41z M16.74,23.62c0-0.04,0.01-0.11,0.04-0.23l1.63-5.77c0.06-0.19,0.16-0.34,0.3-0.44c0.15-0.1,0.3-0.15,0.46-0.15 c0.08,0,0.17,0.01,0.26,0.03c0.21,0.06,0.36,0.16,0.46,0.31c0.1,0.15,0.15,0.31,0.15,0.47c0,0.03-0.01,0.08-0.02,0.14 s-0.02,0.1-0.02,0.12l-1.63,5.73c-0.04,0.19-0.13,0.35-0.28,0.46s-0.32,0.17-0.51,0.17l-0.24-0.05c-0.2-0.06-0.35-0.16-0.46-0.32 C16.79,23.94,16.74,23.78,16.74,23.62z",
-  "rain-mix": "M4.65,16.96c0,1.32,0.47,2.46,1.4,3.41c0.93,0.96,2.06,1.46,3.38,1.5c0.12,0,0.18-0.06,0.18-0.17v-1.33 c0-0.12-0.06-0.18-0.18-0.18c-0.84-0.04-1.57-0.38-2.17-1.02s-0.91-1.37-0.91-2.22c0-0.84,0.28-1.57,0.85-2.19 c0.57-0.62,1.26-0.97,2.1-1.04l0.53-0.07c0.12,0,0.19-0.06,0.19-0.18l0.07-0.5c0.1-1.09,0.55-2.01,1.36-2.75s1.76-1.11,2.86-1.11 c1.08,0,2.03,0.37,2.84,1.1c0.81,0.73,1.28,1.63,1.4,2.71l0.07,0.58c0,0.12,0.06,0.18,0.19,0.18h1.6c0.9,0,1.67,0.32,2.32,0.97 c0.64,0.64,0.97,1.41,0.97,2.3c0,0.84-0.3,1.58-0.9,2.22c-0.6,0.63-1.33,0.97-2.18,1.02c-0.13,0-0.2,0.06-0.2,0.18v1.33 c0,0.11,0.07,0.17,0.2,0.17c1.33-0.04,2.46-0.54,3.38-1.5s1.38-2.09,1.38-3.42c0-0.89-0.22-1.72-0.67-2.48 c-0.44-0.76-1.05-1.36-1.81-1.8c-0.76-0.44-1.59-0.66-2.48-0.66h-0.31c-0.33-1.34-1.03-2.43-2.11-3.29 c-1.07-0.85-2.3-1.28-3.68-1.28c-1.41,0-2.66,0.44-3.75,1.31s-1.79,1.99-2.1,3.35c-1.13,0.29-2.04,0.88-2.75,1.77 S4.65,15.8,4.65,16.96z M10.05,23.98c0,0.17,0.05,0.34,0.16,0.51c0.11,0.17,0.27,0.28,0.47,0.35c0.23,0.07,0.44,0.06,0.64-0.04 c0.19-0.09,0.33-0.28,0.39-0.56l0.14-0.61c0.05-0.23,0.02-0.44-0.09-0.63s-0.28-0.33-0.52-0.4c-0.22-0.07-0.44-0.04-0.64,0.08 s-0.34,0.3-0.4,0.53l-0.14,0.59C10.06,23.83,10.05,23.89,10.05,23.98z M10.81,21.08c0,0.21,0.08,0.4,0.25,0.57 c0.16,0.17,0.34,0.25,0.56,0.25c0.24,0,0.44-0.08,0.6-0.24c0.16-0.16,0.24-0.35,0.24-0.59c0-0.23-0.08-0.43-0.24-0.59 c-0.16-0.16-0.36-0.24-0.6-0.24c-0.23,0-0.42,0.08-0.58,0.23C10.89,20.65,10.81,20.85,10.81,21.08z M11.42,18.81 c-0.01,0.16,0.03,0.31,0.14,0.45c0.1,0.15,0.26,0.25,0.48,0.32c0.21,0.06,0.41,0.04,0.62-0.07S13,19.23,13.07,19l0.28-0.9 c0.07-0.24,0.05-0.46-0.07-0.65c-0.12-0.19-0.3-0.32-0.54-0.39c-0.22-0.07-0.43-0.05-0.63,0.07c-0.2,0.11-0.34,0.28-0.41,0.5 l-0.24,0.92c0,0.02-0.01,0.06-0.02,0.12C11.43,18.72,11.42,18.77,11.42,18.81z M12.59,27.1c0,0.18,0.05,0.34,0.15,0.5 c0.1,0.16,0.26,0.27,0.48,0.33c0.08,0.02,0.17,0.03,0.25,0.03c0.43,0,0.69-0.2,0.79-0.61l0.14-0.59c0.06-0.26,0.03-0.48-0.08-0.68 c-0.12-0.2-0.29-0.32-0.52-0.37c-0.21-0.07-0.42-0.05-0.63,0.07c-0.21,0.12-0.34,0.29-0.41,0.51l-0.14,0.59 C12.6,26.97,12.59,27.04,12.59,27.1z M13.36,24.2c0,0.22,0.08,0.41,0.25,0.58c0.16,0.16,0.35,0.24,0.57,0.24 c0.24,0,0.43-0.08,0.59-0.23c0.16-0.16,0.23-0.35,0.23-0.59c0-0.23-0.08-0.42-0.23-0.58c-0.16-0.16-0.35-0.23-0.59-0.23 c-0.24,0-0.43,0.08-0.59,0.23S13.36,23.97,13.36,24.2z M13.99,21.93c-0.01,0.15,0.03,0.31,0.13,0.47c0.1,0.16,0.25,0.26,0.45,0.3 c0.23,0.06,0.44,0.04,0.64-0.06s0.33-0.29,0.41-0.56l0.27-0.9c0.07-0.22,0.05-0.43-0.07-0.63c-0.12-0.2-0.29-0.33-0.53-0.4 c-0.22-0.07-0.43-0.04-0.64,0.08c-0.21,0.12-0.34,0.3-0.41,0.53l-0.23,0.9C14,21.74,13.99,21.83,13.99,21.93z M16.75,24.08 c0,0.16,0.05,0.32,0.15,0.48c0.1,0.16,0.26,0.27,0.46,0.33c0.03,0,0.08,0.01,0.14,0.02c0.06,0.01,0.1,0.02,0.14,0.02 c0.41,0,0.66-0.22,0.76-0.66l0.14-0.6c0.07-0.21,0.05-0.42-0.07-0.63c-0.11-0.21-0.28-0.34-0.51-0.41 c-0.25-0.06-0.48-0.04-0.68,0.08s-0.34,0.29-0.41,0.53l-0.09,0.59c0,0.02-0.01,0.07-0.02,0.12S16.75,24.04,16.75,24.08z M17.49,21.12c0,0.22,0.08,0.42,0.25,0.57c0.15,0.16,0.34,0.24,0.57,0.24c0.24,0,0.43-0.08,0.59-0.23s0.23-0.35,0.23-0.58 c0-0.24-0.08-0.43-0.23-0.59s-0.35-0.23-0.59-0.23c-0.24,0-0.43,0.08-0.59,0.23C17.57,20.69,17.49,20.88,17.49,21.12z M18.1,18.81 c0,0.17,0.05,0.33,0.16,0.48c0.11,0.15,0.27,0.26,0.49,0.32c0.02,0,0.06,0.01,0.12,0.02s0.11,0.02,0.14,0.02 c0.1,0,0.22-0.03,0.36-0.09c0.21-0.11,0.35-0.29,0.41-0.52l0.24-0.9c0.06-0.23,0.04-0.44-0.08-0.63c-0.11-0.2-0.28-0.33-0.51-0.4 c-0.23-0.07-0.44-0.05-0.64,0.06c-0.19,0.11-0.33,0.27-0.39,0.51l-0.28,0.91c0,0.02-0.01,0.06-0.02,0.12 C18.1,18.74,18.1,18.78,18.1,18.81z",
-  "rain-wind": "M4.63,16.93c0,1.12,0.33,2.11,0.98,2.99c0.65,0.87,1.5,1.47,2.55,1.79c0.09,0.02,0.17-0.01,0.24-0.08l1.16-1.43 c-0.89,0-1.65-0.32-2.28-0.96c-0.63-0.64-0.95-1.41-0.95-2.31c0-0.84,0.28-1.58,0.84-2.2s1.26-0.97,2.1-1.04l0.53-0.07 c0.11,0,0.16-0.04,0.16-0.13l0.08-0.55c0.12-1.1,0.59-2.01,1.39-2.73s1.75-1.08,2.85-1.08c1.1,0,2.06,0.36,2.87,1.09 c0.82,0.73,1.27,1.64,1.37,2.72l0.07,0.58c0.02,0.11,0.1,0.17,0.22,0.17h1.62c0.9,0,1.67,0.32,2.3,0.95s0.95,1.39,0.95,2.29 c0,0.83-0.28,1.56-0.84,2.18s-1.25,0.98-2.07,1.08c-0.12,0-0.28,0.02-0.49,0.06c-0.19,0.02-0.33,0.09-0.41,0.23l-2.36,2.79 c-0.14,0.18-0.2,0.39-0.16,0.63c0.03,0.24,0.14,0.43,0.31,0.57c0.11,0.12,0.29,0.19,0.56,0.19c0.26,0,0.47-0.12,0.61-0.35l2.12-2.44 c1.24-0.13,2.29-0.66,3.15-1.61c0.86-0.95,1.28-2.06,1.28-3.33c0-0.67-0.13-1.32-0.39-1.93c-0.26-0.61-0.61-1.14-1.05-1.58 c-0.44-0.44-0.97-0.79-1.58-1.05c-0.61-0.26-1.25-0.39-1.93-0.39h-0.32c-0.33-1.32-1.04-2.41-2.12-3.26s-2.32-1.27-3.72-1.27 c-0.93,0-1.81,0.2-2.63,0.6c-0.82,0.4-1.51,0.95-2.08,1.66s-0.94,1.52-1.13,2.42c-1.12,0.25-2.04,0.82-2.75,1.72 C4.98,14.74,4.63,15.77,4.63,16.93z M8.01,24.95c0,0.06,0.02,0.16,0.06,0.3c0.09,0.21,0.23,0.36,0.44,0.44 c0.22,0.1,0.44,0.11,0.67,0.02c0.23-0.09,0.38-0.24,0.46-0.45c0.1-0.22,0.11-0.43,0.02-0.65c-0.09-0.21-0.24-0.36-0.46-0.43 c-0.22-0.11-0.44-0.12-0.65-0.03c-0.21,0.09-0.36,0.24-0.46,0.47C8.04,24.72,8.01,24.83,8.01,24.95z M9.86,22.51v0.1 c0.02,0.23,0.12,0.41,0.3,0.56c0.23,0.13,0.43,0.19,0.62,0.19c0.22,0,0.43-0.11,0.61-0.33l2.32-2.77c0.14-0.17,0.21-0.39,0.2-0.66 c-0.02-0.21-0.12-0.39-0.28-0.53c-0.16-0.14-0.33-0.22-0.52-0.22c-0.06,0-0.1,0-0.14,0.01c-0.23,0.04-0.42,0.15-0.56,0.33 l-2.36,2.77C9.92,22.12,9.86,22.3,9.86,22.51z M10.63,27.23c0,0.12,0.03,0.23,0.08,0.32c0.08,0.21,0.23,0.37,0.44,0.47 c0.11,0.05,0.22,0.07,0.33,0.07c0.12,0,0.23-0.02,0.31-0.07c0.23-0.09,0.39-0.23,0.47-0.41c0.1-0.22,0.11-0.44,0.02-0.67 c-0.08-0.23-0.23-0.38-0.45-0.46c-0.22-0.1-0.44-0.11-0.67-0.02c-0.23,0.09-0.38,0.24-0.45,0.45C10.65,27,10.63,27.11,10.63,27.23z M12.3,24.88v0.11c0.02,0.22,0.13,0.4,0.31,0.55c0.18,0.15,0.37,0.22,0.55,0.22c0.23,0,0.43-0.11,0.63-0.33l4.35-5.24 c0.11-0.12,0.17-0.3,0.17-0.52v-0.12c-0.02-0.23-0.12-0.4-0.27-0.53s-0.33-0.2-0.52-0.2h-0.13c-0.23,0.01-0.42,0.12-0.55,0.31 l-4.35,5.2C12.35,24.51,12.3,24.69,12.3,24.88z M15.81,26.03c0,0.09,0.02,0.19,0.06,0.3c0.09,0.22,0.24,0.38,0.46,0.47 c0.14,0.04,0.24,0.06,0.31,0.06c0.14,0,0.26-0.03,0.34-0.08c0.22-0.09,0.38-0.23,0.46-0.42c0.1-0.17,0.11-0.39,0.02-0.67 c-0.08-0.21-0.23-0.35-0.44-0.44l-0.36-0.09c-0.09,0.02-0.19,0.04-0.32,0.07c-0.22,0.08-0.37,0.23-0.45,0.44 C15.84,25.8,15.81,25.92,15.81,26.03z",
-  "raindrops": "M11.01,12.23c0-0.26,0.13-0.59,0.38-1.01c0.25-0.42,0.5-0.77,0.73-1.04c0.06-0.07,0.14-0.17,0.23-0.28s0.15-0.17,0.16-0.18 l0.37,0.43c0.28,0.31,0.53,0.66,0.76,1.07c0.23,0.41,0.35,0.74,0.35,1.01c0,0.41-0.14,0.77-0.43,1.06 c-0.28,0.29-0.63,0.44-1.05,0.44c-0.41,0-0.77-0.15-1.06-0.44C11.16,12.99,11.01,12.64,11.01,12.23z M14.13,16.38 c0-0.29,0.08-0.62,0.24-1.01c0.16-0.38,0.36-0.74,0.6-1.06c0.46-0.61,0.89-1.12,1.31-1.53c0.04-0.03,0.13-0.11,0.26-0.24l0.25,0.24 c0.39,0.37,0.83,0.88,1.32,1.52c0.26,0.34,0.46,0.7,0.62,1.08s0.24,0.71,0.24,1c0,0.69-0.23,1.26-0.7,1.73 c-0.47,0.47-1.05,0.7-1.73,0.7c-0.68,0-1.25-0.24-1.72-0.71S14.13,17.05,14.13,16.38z M15.65,9.48c0-0.43,0.33-1,1-1.7l0.25,0.28 c0.19,0.22,0.36,0.46,0.51,0.74c0.15,0.27,0.23,0.5,0.23,0.68c0,0.28-0.1,0.5-0.29,0.69c-0.19,0.18-0.42,0.28-0.7,0.28 c-0.29,0-0.53-0.09-0.72-0.28C15.75,9.98,15.65,9.75,15.65,9.48z",
-  "showers": "M4.6,16.93c0-1.16,0.36-2.18,1.09-3.08c0.72-0.9,1.65-1.48,2.78-1.73c0.29-1.38,0.98-2.5,2.07-3.39S12.88,7.4,14.3,7.4 c1.39,0,2.63,0.43,3.72,1.28c1.08,0.85,1.79,1.95,2.12,3.3h0.34c0.9,0,1.73,0.22,2.48,0.66c0.76,0.44,1.35,1.04,1.79,1.8 c0.43,0.76,0.65,1.59,0.65,2.49c0,1.34-0.46,2.48-1.37,3.44c-0.92,0.96-2.04,1.46-3.37,1.5c-0.12,0-0.18-0.06-0.18-0.17v-1.34 c0-0.11,0.06-0.17,0.18-0.17c0.84-0.07,1.57-0.42,2.17-1.05s0.9-1.37,0.9-2.22c0-0.89-0.32-1.66-0.96-2.31 c-0.64-0.64-1.4-0.97-2.29-0.97h-1.63c-0.12,0-0.19-0.06-0.22-0.18l-0.07-0.57c-0.07-0.71-0.3-1.36-0.7-1.94s-0.91-1.03-1.53-1.36 c-0.62-0.33-1.3-0.49-2.02-0.49c-1.1,0-2.05,0.36-2.86,1.09c-0.81,0.73-1.27,1.64-1.37,2.72l-0.07,0.54c0,0.09-0.05,0.14-0.16,0.14 L9.31,13.7c-0.84,0.07-1.55,0.41-2.11,1.03c-0.57,0.62-0.85,1.35-0.85,2.2c0,0.87,0.3,1.62,0.89,2.25c0.59,0.63,1.31,0.97,2.17,1.02 c0.12,0,0.18,0.06,0.18,0.17v1.34c0,0.11-0.06,0.17-0.18,0.17c-0.66-0.03-1.28-0.18-1.88-0.45S6.42,20.8,6,20.36 c-0.43-0.44-0.77-0.95-1.02-1.55S4.6,17.59,4.6,16.93z M10.02,23.7c0-0.03,0.01-0.08,0.02-0.13s0.02-0.09,0.02-0.11l0.27-1.03 c0.07-0.22,0.2-0.4,0.4-0.51c0.2-0.12,0.41-0.14,0.64-0.07c0.23,0.07,0.4,0.2,0.52,0.4c0.12,0.2,0.14,0.41,0.07,0.64l-0.24,1.01 c-0.13,0.44-0.38,0.66-0.76,0.66c-0.03,0-0.05,0-0.09,0c-0.03,0-0.07-0.01-0.11-0.01c-0.04-0.01-0.07-0.01-0.1-0.01 c-0.21-0.06-0.37-0.18-0.48-0.34S10.02,23.86,10.02,23.7z M11.34,18.88c0-0.02,0-0.06,0.01-0.11c0.01-0.05,0.01-0.08,0.01-0.09 l0.3-1.05c0.06-0.19,0.17-0.34,0.32-0.45c0.15-0.1,0.31-0.15,0.47-0.15c0.02,0,0.05,0,0.08,0c0.03,0,0.06,0.01,0.09,0.01 c0.03,0.01,0.06,0.01,0.08,0.01c0.23,0.07,0.4,0.2,0.51,0.4c0.12,0.2,0.14,0.41,0.07,0.64l-0.24,1c-0.07,0.28-0.2,0.47-0.4,0.59 s-0.42,0.12-0.65,0.02c-0.22-0.06-0.38-0.17-0.49-0.34S11.34,19.04,11.34,18.88z M12.57,26.83c0-0.03,0.01-0.07,0.02-0.13 s0.02-0.09,0.02-0.12l0.29-0.99c0.06-0.24,0.2-0.42,0.4-0.54c0.2-0.12,0.42-0.15,0.65-0.08c0.23,0.07,0.39,0.2,0.51,0.41 s0.13,0.42,0.07,0.65l-0.25,1.04c-0.11,0.41-0.37,0.61-0.8,0.61c-0.05,0-0.13-0.01-0.24-0.04c-0.22-0.04-0.38-0.14-0.49-0.3 C12.63,27.18,12.57,27.01,12.57,26.83z M13.91,22.06c0-0.06,0.01-0.14,0.04-0.25l0.27-1.03c0.07-0.23,0.2-0.4,0.41-0.51 c0.2-0.12,0.42-0.14,0.65-0.07c0.23,0.06,0.39,0.19,0.51,0.39c0.11,0.2,0.13,0.41,0.06,0.65l-0.24,0.99 c-0.13,0.45-0.37,0.68-0.72,0.68c-0.04,0-0.15-0.02-0.31-0.06c-0.22-0.04-0.38-0.14-0.49-0.3C13.97,22.4,13.91,22.23,13.91,22.06z M16.73,23.74c0-0.07,0.01-0.15,0.03-0.24l0.28-0.99c0.07-0.24,0.2-0.42,0.41-0.54s0.41-0.15,0.63-0.09 c0.23,0.07,0.41,0.2,0.53,0.41c0.12,0.2,0.15,0.41,0.09,0.63l-0.29,1.06c-0.1,0.41-0.36,0.61-0.79,0.61c-0.09,0-0.18-0.01-0.26-0.03 c-0.2-0.04-0.35-0.14-0.46-0.3C16.8,24.08,16.74,23.91,16.73,23.74z M18.11,18.98c0-0.03,0.02-0.12,0.05-0.26l0.3-1.03 c0.04-0.21,0.13-0.37,0.29-0.47c0.16-0.1,0.32-0.15,0.49-0.14c0.04-0.01,0.13,0,0.24,0.03c0.22,0.05,0.39,0.18,0.52,0.38 c0.12,0.17,0.14,0.38,0.07,0.65l-0.24,1.03c-0.13,0.43-0.38,0.65-0.76,0.65c-0.06,0-0.17-0.02-0.34-0.06 c-0.21-0.06-0.36-0.17-0.46-0.31C18.16,19.29,18.11,19.14,18.11,18.98z",
-  "sleet": "M4.09,16.95c0-1.16,0.36-2.19,1.08-3.09s1.64-1.49,2.74-1.74c0.31-1.37,1.01-2.49,2.1-3.37s2.35-1.32,3.77-1.32 c0.81,0,1.55,0.13,2.2,0.4c0,0.01,0,0.01,0.01,0.02c0.84,0.32,1.58,0.84,2.21,1.55h0.03c0.68,0.73,1.13,1.6,1.37,2.62h0.31 c1.08,0,2.02,0.29,2.83,0.86v-0.01c0.53,0.36,0.98,0.8,1.34,1.33c0.36,0.53,0.6,1.11,0.73,1.74c0.04,0.21,0.06,0.38,0.08,0.52v0.06 c0,0.01,0,0.06,0.01,0.17s0.01,0.19,0.01,0.24v0.03c0,0.88-0.21,1.7-0.64,2.45c-0.42,0.75-1,1.36-1.74,1.81 c-0.73,0.45-1.54,0.69-2.42,0.72c-0.13,0-0.2-0.06-0.2-0.17v-1.34c0-0.13,0.07-0.19,0.2-0.19c0.86-0.04,1.58-0.38,2.18-1.02 c0.6-0.64,0.9-1.39,0.9-2.26c0-0.89-0.32-1.65-0.97-2.29c-0.64-0.64-1.41-0.96-2.31-0.96h-1.61c-0.12,0-0.18-0.06-0.18-0.17 l-0.08-0.59c-0.1-1-0.52-1.86-1.27-2.59c-0.01-0.01-0.01-0.02-0.02-0.03s-0.02-0.02-0.02-0.03c-0.01-0.01-0.02-0.02-0.04-0.02 c0-0.02-0.01-0.03-0.02-0.03c-0.73-0.66-1.58-1.04-2.56-1.12c-0.07-0.01-0.18-0.01-0.34-0.01c-1.1,0-2.05,0.37-2.86,1.1 s-1.27,1.64-1.37,2.72l-0.08,0.59c-0.03,0.09-0.11,0.14-0.22,0.14L8.75,13.7c-0.84,0.1-1.55,0.46-2.11,1.07s-0.84,1.34-0.84,2.18 v0.04h0.03c0.01,0.48,0.11,0.93,0.3,1.35c0.2,0.43,0.46,0.79,0.8,1.09c0.21,0.18,0.45,0.34,0.74,0.48v0.01 c0.4,0.19,0.8,0.3,1.21,0.32c0.11,0,0.17,0.06,0.17,0.18v1.34c0,0.11-0.06,0.17-0.17,0.17c-0.52-0.03-1.01-0.13-1.48-0.3v0.01 c-0.83-0.29-1.54-0.77-2.11-1.43s-0.95-1.44-1.11-2.31v-0.03c-0.01-0.01-0.01-0.02-0.01-0.04C4.11,17.57,4.09,17.27,4.09,16.95z M9.59,24.12c0-0.03,0.01-0.07,0.02-0.13c0.01-0.05,0.02-0.09,0.02-0.12l0.09-0.59c0.07-0.24,0.2-0.41,0.41-0.53 s0.43-0.14,0.68-0.08c0.23,0.07,0.39,0.21,0.51,0.41c0.11,0.21,0.13,0.42,0.07,0.63l-0.14,0.6c-0.1,0.44-0.35,0.66-0.76,0.66 c-0.03,0-0.08-0.01-0.14-0.02c-0.06-0.01-0.1-0.02-0.14-0.02c-0.21-0.06-0.36-0.17-0.46-0.33C9.64,24.45,9.59,24.29,9.59,24.12z M10.33,21.18c0-0.24,0.08-0.43,0.23-0.59s0.35-0.23,0.59-0.23s0.43,0.08,0.59,0.23s0.23,0.35,0.23,0.59c0,0.23-0.08,0.42-0.23,0.58 s-0.35,0.23-0.59,0.23c-0.23,0-0.42-0.08-0.57-0.24C10.42,21.59,10.33,21.4,10.33,21.18z M11.97,27.17c0-0.04,0.01-0.11,0.04-0.23 l0.13-0.59c0.07-0.23,0.21-0.4,0.41-0.51c0.21-0.12,0.42-0.14,0.63-0.07c0.23,0.04,0.41,0.17,0.53,0.37 c0.12,0.2,0.15,0.43,0.08,0.68l-0.13,0.59c-0.1,0.41-0.37,0.61-0.8,0.61c-0.07,0-0.16-0.01-0.24-0.03 c-0.22-0.06-0.38-0.17-0.49-0.33C12.03,27.5,11.97,27.34,11.97,27.17z M12.76,24.26c0-0.23,0.08-0.43,0.23-0.58 c0.16-0.16,0.35-0.23,0.59-0.23s0.43,0.08,0.59,0.23c0.16,0.16,0.23,0.35,0.23,0.58c0,0.24-0.08,0.43-0.23,0.59 c-0.16,0.16-0.35,0.23-0.59,0.23c-0.23,0-0.42-0.08-0.58-0.24C12.84,24.68,12.76,24.49,12.76,24.26z M13.38,21.99 c0-0.1,0.01-0.19,0.03-0.27l0.23-0.9c0.07-0.23,0.21-0.41,0.41-0.53c0.21-0.12,0.42-0.15,0.64-0.08c0.24,0.07,0.41,0.2,0.53,0.4 s0.14,0.41,0.07,0.63l-0.26,0.9c-0.08,0.28-0.22,0.46-0.41,0.56c-0.19,0.1-0.41,0.12-0.64,0.06c-0.2-0.04-0.35-0.14-0.45-0.3 C13.41,22.3,13.37,22.14,13.38,21.99z M16.14,24.13c0-0.03,0-0.08,0.01-0.13s0.01-0.09,0.01-0.11l0.09-0.59 c0.07-0.24,0.2-0.41,0.41-0.53s0.43-0.14,0.68-0.08c0.23,0.07,0.4,0.21,0.51,0.41c0.12,0.21,0.14,0.42,0.07,0.63l-0.14,0.6 c-0.1,0.44-0.35,0.66-0.76,0.66c-0.03,0-0.08-0.01-0.14-0.02c-0.06-0.01-0.11-0.02-0.14-0.02c-0.2-0.06-0.35-0.17-0.45-0.33 C16.2,24.46,16.14,24.3,16.14,24.13z M16.88,21.19c0-0.24,0.08-0.43,0.23-0.59c0.16-0.16,0.35-0.23,0.59-0.23 c0.24,0,0.43,0.08,0.59,0.23c0.16,0.16,0.23,0.35,0.23,0.59c0,0.23-0.08,0.43-0.23,0.58C18.14,21.92,17.95,22,17.71,22 c-0.24,0-0.43-0.08-0.58-0.24C16.97,21.61,16.88,21.42,16.88,21.19z",
-  "snow": "M4.64,16.95c0-1.16,0.35-2.18,1.06-3.08s1.62-1.48,2.74-1.76c0.31-1.36,1.01-2.48,2.1-3.36s2.34-1.31,3.75-1.31 c1.38,0,2.6,0.43,3.68,1.28c1.08,0.85,1.78,1.95,2.1,3.29h0.32c0.89,0,1.72,0.22,2.48,0.66c0.76,0.44,1.37,1.04,1.81,1.8 c0.44,0.76,0.67,1.59,0.67,2.48c0,1.32-0.46,2.47-1.39,3.42c-0.92,0.96-2.05,1.46-3.38,1.5c-0.13,0-0.2-0.06-0.2-0.17v-1.33 c0-0.12,0.07-0.18,0.2-0.18c0.85-0.04,1.58-0.38,2.18-1.02s0.9-1.38,0.9-2.23c0-0.89-0.32-1.65-0.97-2.3s-1.42-0.97-2.32-0.97h-1.61 c-0.12,0-0.18-0.06-0.18-0.17l-0.08-0.58c-0.11-1.08-0.58-1.99-1.39-2.72c-0.82-0.73-1.76-1.1-2.85-1.1c-1.1,0-2.05,0.37-2.86,1.11 c-0.81,0.74-1.27,1.65-1.37,2.75l-0.06,0.5c0,0.12-0.07,0.19-0.2,0.19l-0.53,0.07c-0.83,0.07-1.53,0.41-2.1,1.04 s-0.85,1.35-0.85,2.19c0,0.85,0.3,1.59,0.9,2.23s1.33,0.97,2.18,1.02c0.11,0,0.17,0.06,0.17,0.18v1.33c0,0.11-0.06,0.17-0.17,0.17 c-1.34-0.04-2.47-0.54-3.4-1.5C5.1,19.42,4.64,18.27,4.64,16.95z M11,21.02c0-0.22,0.08-0.42,0.24-0.58 c0.16-0.16,0.35-0.24,0.59-0.24c0.23,0,0.43,0.08,0.59,0.24c0.16,0.16,0.24,0.36,0.24,0.58c0,0.24-0.08,0.44-0.24,0.6 c-0.16,0.17-0.35,0.25-0.59,0.25c-0.23,0-0.43-0.08-0.59-0.25C11.08,21.46,11,21.26,11,21.02z M11,24.65c0-0.24,0.08-0.44,0.24-0.6 c0.16-0.15,0.35-0.23,0.58-0.23c0.23,0,0.43,0.08,0.59,0.23c0.16,0.16,0.24,0.35,0.24,0.59c0,0.24-0.08,0.43-0.24,0.59 c-0.16,0.16-0.35,0.23-0.59,0.23c-0.23,0-0.43-0.08-0.59-0.23C11.08,25.08,11,24.88,11,24.65z M14.19,22.95 c0-0.23,0.08-0.44,0.25-0.62c0.16-0.16,0.35-0.24,0.57-0.24c0.23,0,0.43,0.09,0.6,0.26c0.17,0.17,0.26,0.37,0.26,0.6 c0,0.23-0.08,0.43-0.25,0.6c-0.17,0.17-0.37,0.25-0.61,0.25c-0.23,0-0.42-0.08-0.58-0.25S14.19,23.18,14.19,22.95z M14.19,19.33 c0-0.23,0.08-0.43,0.25-0.6c0.18-0.16,0.37-0.24,0.57-0.24c0.24,0,0.44,0.08,0.61,0.25c0.17,0.17,0.25,0.36,0.25,0.6 c0,0.23-0.08,0.43-0.25,0.59c-0.17,0.16-0.37,0.24-0.61,0.24c-0.23,0-0.42-0.08-0.58-0.24C14.27,19.76,14.19,19.56,14.19,19.33z M14.19,26.61c0-0.23,0.08-0.43,0.25-0.61c0.16-0.16,0.35-0.24,0.57-0.24c0.24,0,0.44,0.08,0.61,0.25c0.17,0.17,0.25,0.37,0.25,0.6 s-0.08,0.43-0.25,0.59c-0.17,0.16-0.37,0.24-0.61,0.24c-0.23,0-0.42-0.08-0.58-0.24C14.27,27.03,14.19,26.84,14.19,26.61z M17.41,21.02c0-0.22,0.08-0.41,0.25-0.58c0.17-0.17,0.37-0.25,0.6-0.25c0.23,0,0.43,0.08,0.59,0.24c0.16,0.16,0.24,0.36,0.24,0.58 c0,0.24-0.08,0.44-0.24,0.6c-0.16,0.17-0.35,0.25-0.59,0.25c-0.24,0-0.44-0.08-0.6-0.25C17.5,21.45,17.41,21.25,17.41,21.02z M17.41,24.65c0-0.22,0.08-0.42,0.25-0.6c0.16-0.15,0.36-0.23,0.6-0.23c0.24,0,0.43,0.08,0.59,0.23s0.23,0.35,0.23,0.59 c0,0.24-0.08,0.43-0.23,0.59c-0.16,0.16-0.35,0.23-0.59,0.23c-0.24,0-0.44-0.08-0.6-0.24C17.5,25.07,17.41,24.88,17.41,24.65z",
-  "stars": "M5.37,16.18c0.65-0.03,1.2-0.28,1.65-0.75c0.45-0.47,0.68-1.03,0.68-1.68c0,0.65,0.22,1.21,0.67,1.68 c0.45,0.47,1,0.72,1.65,0.75c-0.65,0.03-1.2,0.28-1.65,0.75c-0.45,0.47-0.67,1.03-0.67,1.68c0-0.65-0.22-1.21-0.68-1.68 C6.57,16.46,6.02,16.21,5.37,16.18z M7.7,8.98c1.26-0.06,2.33-0.55,3.21-1.47c0.88-0.92,1.32-2.01,1.32-3.28 c0,1.27,0.44,2.36,1.32,3.28s1.95,1.4,3.22,1.47c-0.83,0.04-1.59,0.27-2.29,0.71c-0.69,0.43-1.24,1.01-1.65,1.73 c-0.4,0.72-0.6,1.49-0.6,2.33c0-1.27-0.44-2.37-1.32-3.29C10.03,9.53,8.96,9.04,7.7,8.98z M11.02,19.75 c0.95-0.04,1.76-0.41,2.42-1.1c0.66-0.69,0.99-1.51,0.99-2.47c0,0.96,0.33,1.78,0.99,2.47c0.66,0.69,1.46,1.06,2.41,1.1 c-0.95,0.04-1.75,0.41-2.41,1.1c-0.66,0.69-0.99,1.51-0.99,2.47c0-0.96-0.33-1.78-0.99-2.47C12.77,20.16,11.97,19.8,11.02,19.75z M17.83,15.01c0.95-0.04,1.75-0.41,2.41-1.1c0.66-0.69,0.98-1.51,0.98-2.48c0,0.96,0.33,1.78,0.99,2.47s1.47,1.06,2.42,1.1 c-0.95,0.04-1.76,0.41-2.42,1.1c-0.66,0.69-0.99,1.51-0.99,2.47c0-0.96-0.33-1.78-0.98-2.47C19.58,15.42,18.78,15.05,17.83,15.01z",
-  "strong-wind": "M3.1,16.97c0,0.24,0.09,0.45,0.28,0.62c0.16,0.19,0.37,0.28,0.63,0.28H18.7c0.29,0,0.53,0.1,0.73,0.3 c0.2,0.2,0.3,0.45,0.3,0.74c0,0.29-0.1,0.53-0.3,0.72c-0.2,0.19-0.44,0.29-0.74,0.29c-0.29,0-0.54-0.1-0.73-0.29 c-0.16-0.18-0.36-0.26-0.6-0.26c-0.25,0-0.46,0.09-0.64,0.26s-0.27,0.38-0.27,0.61c0,0.25,0.09,0.46,0.28,0.63 c0.56,0.55,1.22,0.83,1.96,0.83c0.78,0,1.45-0.27,2.01-0.81c0.56-0.54,0.83-1.19,0.83-1.97s-0.28-1.44-0.84-2 c-0.56-0.56-1.23-0.84-2-0.84H4.01c-0.25,0-0.46,0.09-0.64,0.26C3.19,16.51,3.1,16.72,3.1,16.97z M3.1,13.69 c0,0.23,0.09,0.43,0.28,0.61c0.17,0.18,0.38,0.26,0.63,0.26h20.04c0.78,0,1.45-0.27,2.01-0.82c0.56-0.54,0.84-1.2,0.84-1.97 c0-0.77-0.28-1.44-0.84-1.99s-1.23-0.83-2.01-0.83c-0.77,0-1.42,0.27-1.95,0.8c-0.18,0.16-0.27,0.38-0.27,0.67 c0,0.26,0.09,0.47,0.26,0.63c0.17,0.16,0.38,0.24,0.63,0.24c0.24,0,0.45-0.08,0.63-0.24c0.19-0.21,0.42-0.31,0.7-0.31 c0.29,0,0.53,0.1,0.73,0.3c0.2,0.2,0.3,0.44,0.3,0.73c0,0.29-0.1,0.53-0.3,0.72c-0.2,0.19-0.44,0.29-0.73,0.29H4.01 c-0.25,0-0.46,0.09-0.64,0.26C3.19,13.23,3.1,13.44,3.1,13.69z",
-  "sunrise": "M2.75,15.36c0-0.25,0.1-0.48,0.3-0.69c0.22-0.19,0.46-0.29,0.7-0.29h2.33c0.27,0,0.49,0.1,0.67,0.29 c0.18,0.19,0.27,0.43,0.27,0.69c0,0.29-0.09,0.53-0.27,0.72c-0.18,0.19-0.41,0.29-0.67,0.29H3.75c-0.27,0-0.5-0.1-0.7-0.3 C2.85,15.86,2.75,15.62,2.75,15.36z M6.08,7.38c0-0.27,0.09-0.5,0.26-0.68C6.57,6.5,6.8,6.4,7.05,6.4c0.26,0,0.49,0.1,0.68,0.29 l1.64,1.65c0.19,0.22,0.28,0.45,0.28,0.69c0,0.28-0.09,0.52-0.27,0.7s-0.4,0.28-0.66,0.28c-0.24,0-0.48-0.1-0.7-0.29L6.34,8.11 C6.17,7.9,6.08,7.65,6.08,7.38z M8.08,20.88c0-0.28,0.1-0.51,0.29-0.68c0.18-0.17,0.4-0.26,0.68-0.26h2.63l3.11-2.92 c0.1-0.08,0.21-0.08,0.34,0l3.16,2.92h2.77c0.27,0,0.5,0.09,0.69,0.28c0.19,0.18,0.29,0.41,0.29,0.67c0,0.27-0.1,0.5-0.29,0.69 c-0.19,0.19-0.42,0.29-0.69,0.29h-3.38c-0.1,0-0.2-0.02-0.29-0.07l-2.41-2.27l-2.39,2.27c-0.08,0.05-0.17,0.07-0.28,0.07H9.05 c-0.27,0-0.5-0.1-0.69-0.29C8.17,21.38,8.08,21.15,8.08,20.88z M9,15.36c0,0.97,0.21,1.85,0.62,2.64c0.02,0.12,0.11,0.18,0.25,0.18 h1.88c0.07,0,0.12-0.03,0.15-0.08c0.03-0.06,0.02-0.12-0.02-0.19c-0.64-0.77-0.96-1.62-0.96-2.55c0-1.12,0.4-2.08,1.2-2.87 c0.8-0.79,1.76-1.18,2.89-1.18c1.12,0,2.07,0.39,2.86,1.18c0.79,0.79,1.19,1.74,1.19,2.87c0,0.94-0.32,1.79-0.95,2.55 c-0.04,0.07-0.05,0.13-0.03,0.19s0.07,0.08,0.15,0.08h1.9c0.13,0,0.21-0.06,0.23-0.18C20.8,17.23,21,16.35,21,15.36 c0-0.81-0.16-1.59-0.48-2.32c-0.32-0.74-0.75-1.37-1.28-1.91c-0.53-0.53-1.17-0.96-1.91-1.28c-0.74-0.32-1.51-0.47-2.32-0.47 c-0.81,0-1.59,0.16-2.33,0.47c-0.74,0.32-1.38,0.74-1.92,1.28c-0.54,0.53-0.97,1.17-1.29,1.91C9.16,13.77,9,14.54,9,15.36z M14.03,6.4v-2.3c0-0.29,0.09-0.52,0.28-0.71s0.43-0.28,0.71-0.28c0.28,0,0.51,0.09,0.7,0.28S16,3.83,16,4.11v2.3 c0,0.29-0.09,0.52-0.28,0.71c-0.18,0.18-0.42,0.28-0.7,0.28c-0.29,0-0.52-0.09-0.71-0.28C14.12,6.93,14.03,6.69,14.03,6.4z M20.38,9.04c0-0.25,0.09-0.48,0.27-0.69l1.62-1.65c0.19-0.19,0.43-0.29,0.7-0.29c0.27,0,0.51,0.1,0.69,0.29 c0.19,0.19,0.28,0.42,0.28,0.69c0,0.29-0.09,0.53-0.26,0.73L22,9.73c-0.21,0.19-0.45,0.29-0.7,0.29c-0.27,0-0.49-0.09-0.66-0.28 S20.38,9.32,20.38,9.04z M22.99,15.36c0-0.27,0.09-0.5,0.27-0.69c0.18-0.19,0.4-0.29,0.66-0.29h2.35c0.27,0,0.5,0.1,0.69,0.29 c0.19,0.19,0.29,0.43,0.29,0.69c0,0.28-0.1,0.51-0.29,0.71s-0.42,0.3-0.69,0.3h-2.35c-0.27,0-0.49-0.1-0.67-0.29 C23.08,15.88,22.99,15.64,22.99,15.36z",
-  "sunset": "M2.88,15.47c0-0.28,0.1-0.5,0.3-0.68c0.17-0.18,0.4-0.26,0.68-0.26h2.31c0.27,0,0.49,0.09,0.67,0.27 c0.17,0.18,0.26,0.4,0.26,0.67c0,0.28-0.09,0.52-0.27,0.71c-0.18,0.19-0.4,0.29-0.66,0.29H3.87c-0.27,0-0.5-0.1-0.69-0.3 C2.98,15.97,2.88,15.74,2.88,15.47z M6.17,7.61c0-0.28,0.08-0.51,0.25-0.68c0.2-0.2,0.43-0.3,0.7-0.3c0.29,0,0.51,0.1,0.68,0.3 l1.62,1.63c0.46,0.44,0.46,0.89,0,1.35c-0.19,0.19-0.4,0.28-0.65,0.28c-0.22,0-0.44-0.09-0.68-0.28L6.43,8.29 C6.26,8.11,6.17,7.88,6.17,7.61z M8.14,20.89c0-0.26,0.1-0.49,0.3-0.69c0.18-0.18,0.41-0.27,0.68-0.27h3.22 c0.11,0,0.2,0.02,0.28,0.08l2.35,2.22L17.36,20c0.07-0.05,0.17-0.08,0.29-0.08h3.3c0.27,0,0.5,0.09,0.69,0.28 c0.19,0.19,0.29,0.42,0.29,0.68c0,0.27-0.1,0.5-0.29,0.69c-0.19,0.19-0.42,0.29-0.69,0.29h-2.68l-3.14,2.84 c-0.12,0.09-0.23,0.09-0.33,0l-3.08-2.84h-2.6c-0.27,0-0.5-0.1-0.69-0.29C8.24,21.39,8.14,21.16,8.14,20.89z M9.08,15.47 c0,0.99,0.19,1.87,0.58,2.62c0.06,0.11,0.15,0.16,0.27,0.16h1.87c0.08,0,0.13-0.02,0.15-0.07c0.02-0.05-0.01-0.11-0.07-0.18 c-0.59-0.74-0.89-1.59-0.89-2.53c0-1.1,0.39-2.04,1.18-2.81c0.79-0.77,1.74-1.16,2.85-1.16c1.1,0,2.04,0.39,2.83,1.16 c0.78,0.78,1.18,1.71,1.18,2.8c0,0.94-0.3,1.79-0.89,2.53c-0.07,0.07-0.09,0.13-0.07,0.18c0.02,0.05,0.07,0.07,0.15,0.07h1.88 c0.13,0,0.21-0.05,0.24-0.16c0.41-0.78,0.62-1.66,0.62-2.62c0-0.79-0.16-1.56-0.47-2.29s-0.74-1.37-1.27-1.9s-1.16-0.95-1.89-1.27 c-0.73-0.32-1.5-0.47-2.3-0.47c-0.8,0-1.57,0.16-2.3,0.47c-0.73,0.32-1.36,0.74-1.89,1.27s-0.95,1.16-1.27,1.9 S9.08,14.68,9.08,15.47z M14.04,6.66V4.33c0-0.27,0.1-0.5,0.29-0.69s0.42-0.29,0.69-0.29c0.27,0,0.5,0.1,0.69,0.29 s0.29,0.42,0.29,0.69v2.32c0,0.27-0.1,0.5-0.29,0.69c-0.19,0.19-0.42,0.29-0.69,0.29c-0.27,0-0.5-0.1-0.69-0.29 C14.13,7.15,14.04,6.93,14.04,6.66z M20.31,9.24c0-0.28,0.09-0.51,0.26-0.67l1.63-1.63c0.16-0.2,0.39-0.3,0.68-0.3 c0.27,0,0.5,0.1,0.68,0.29c0.18,0.19,0.27,0.42,0.27,0.69c0,0.28-0.08,0.51-0.25,0.68l-1.66,1.63c-0.23,0.19-0.46,0.28-0.69,0.28 c-0.26,0-0.48-0.09-0.66-0.28C20.4,9.74,20.31,9.51,20.31,9.24z M22.9,15.47c0-0.27,0.09-0.49,0.26-0.67 c0.17-0.18,0.4-0.27,0.67-0.27h2.32c0.27,0,0.5,0.09,0.69,0.27c0.19,0.18,0.29,0.4,0.29,0.67c0,0.27-0.1,0.5-0.29,0.7 c-0.19,0.2-0.42,0.3-0.69,0.3h-2.32c-0.26,0-0.48-0.1-0.66-0.29C22.99,15.99,22.9,15.75,22.9,15.47z",
-  "thermometer": "M9.91,19.56c0-0.85,0.2-1.64,0.59-2.38s0.94-1.35,1.65-1.84V5.42c0-0.8,0.27-1.48,0.82-2.03S14.2,2.55,15,2.55 c0.81,0,1.49,0.28,2.04,0.83c0.55,0.56,0.83,1.23,0.83,2.03v9.92c0.71,0.49,1.25,1.11,1.64,1.84s0.58,1.53,0.58,2.38 c0,0.92-0.23,1.78-0.68,2.56s-1.07,1.4-1.85,1.85s-1.63,0.68-2.56,0.68c-0.92,0-1.77-0.23-2.55-0.68s-1.4-1.07-1.86-1.85 S9.91,20.48,9.91,19.56z M11.67,19.56c0,0.93,0.33,1.73,0.98,2.39c0.65,0.66,1.44,0.99,2.36,0.99c0.93,0,1.73-0.33,2.4-1 s1.01-1.46,1.01-2.37c0-0.62-0.16-1.2-0.48-1.73c-0.32-0.53-0.76-0.94-1.32-1.23l-0.28-0.14c-0.1-0.04-0.15-0.14-0.15-0.29V5.42 c0-0.32-0.11-0.59-0.34-0.81C15.62,4.4,15.34,4.29,15,4.29c-0.32,0-0.6,0.11-0.83,0.32c-0.23,0.21-0.34,0.48-0.34,0.81v10.74 c0,0.15-0.05,0.25-0.14,0.29l-0.27,0.14c-0.55,0.29-0.98,0.7-1.29,1.23C11.82,18.35,11.67,18.92,11.67,19.56z M12.45,19.56 c0,0.71,0.24,1.32,0.73,1.82s1.07,0.75,1.76,0.75s1.28-0.25,1.79-0.75c0.51-0.5,0.76-1.11,0.76-1.81c0-0.63-0.22-1.19-0.65-1.67 c-0.43-0.48-0.96-0.77-1.58-0.85V9.69c0-0.06-0.03-0.13-0.1-0.19c-0.07-0.07-0.14-0.1-0.22-0.1c-0.09,0-0.16,0.03-0.21,0.08 c-0.05,0.06-0.08,0.12-0.08,0.21v7.34c-0.61,0.09-1.13,0.37-1.56,0.85C12.66,18.37,12.45,18.92,12.45,19.56z",
-  "thunderstorm": "M4.63,16.91c0,1.11,0.33,2.1,0.99,2.97s1.52,1.47,2.58,1.79l-0.66,1.68c-0.03,0.14,0.02,0.22,0.14,0.22h2.13l-0.98,4.3h0.28 l3.92-5.75c0.04-0.04,0.04-0.09,0.01-0.14c-0.03-0.05-0.08-0.07-0.15-0.07h-2.18l2.48-4.64c0.07-0.14,0.02-0.22-0.14-0.22h-2.94 c-0.09,0-0.17,0.05-0.23,0.15l-1.07,2.87c-0.71-0.18-1.3-0.57-1.77-1.16c-0.47-0.59-0.7-1.26-0.7-2.01c0-0.83,0.28-1.55,0.85-2.17 c0.57-0.61,1.27-0.97,2.1-1.07l0.53-0.07c0.13,0,0.2-0.06,0.2-0.18l0.07-0.51c0.11-1.08,0.56-1.99,1.37-2.72 c0.81-0.73,1.76-1.1,2.85-1.1c1.09,0,2.04,0.37,2.85,1.1c0.82,0.73,1.28,1.64,1.4,2.72l0.07,0.58c0,0.11,0.06,0.17,0.18,0.17h1.6 c0.91,0,1.68,0.32,2.32,0.95c0.64,0.63,0.97,1.4,0.97,2.28c0,0.85-0.3,1.59-0.89,2.21c-0.59,0.62-1.33,0.97-2.2,1.04 c-0.13,0-0.2,0.06-0.2,0.18v1.37c0,0.11,0.07,0.17,0.2,0.17c1.33-0.04,2.46-0.55,3.39-1.51s1.39-2.11,1.39-3.45 c0-0.9-0.22-1.73-0.67-2.49c-0.44-0.76-1.05-1.36-1.81-1.8c-0.77-0.44-1.6-0.66-2.5-0.66H20.1c-0.33-1.33-1.04-2.42-2.11-3.26 s-2.3-1.27-3.68-1.27c-1.41,0-2.67,0.44-3.76,1.31s-1.79,1.99-2.1,3.36c-1.11,0.26-2.02,0.83-2.74,1.73S4.63,15.76,4.63,16.91z M12.77,26.62c0,0.39,0.19,0.65,0.58,0.77c0.01,0,0.05,0,0.11,0.01c0.06,0.01,0.11,0.01,0.14,0.01c0.17,0,0.33-0.05,0.49-0.15 c0.16-0.1,0.27-0.26,0.32-0.48l2.25-8.69c0.06-0.24,0.04-0.45-0.07-0.65c-0.11-0.19-0.27-0.32-0.5-0.39 c-0.17-0.02-0.26-0.03-0.26-0.03c-0.16,0-0.32,0.05-0.47,0.15c-0.15,0.1-0.26,0.25-0.31,0.45l-2.26,8.72 C12.78,26.44,12.77,26.53,12.77,26.62z M16.93,23.56c0,0.13,0.03,0.26,0.1,0.38c0.14,0.22,0.31,0.37,0.51,0.44 c0.11,0.03,0.21,0.05,0.3,0.05s0.2-0.02,0.32-0.08c0.21-0.09,0.35-0.28,0.42-0.57l1.44-5.67c0.03-0.14,0.05-0.23,0.05-0.27 c0-0.15-0.05-0.3-0.16-0.45s-0.26-0.26-0.46-0.32c-0.17-0.02-0.26-0.03-0.26-0.03c-0.17,0-0.33,0.05-0.47,0.15 c-0.14,0.1-0.24,0.25-0.3,0.45l-1.46,5.7c0,0.02,0,0.05-0.01,0.11C16.93,23.5,16.93,23.53,16.93,23.56z",
-  "wind-deg": "M3.74,14.5c0-2.04,0.51-3.93,1.52-5.66s2.38-3.1,4.11-4.11s3.61-1.51,5.64-1.51c1.52,0,2.98,0.3,4.37,0.89 s2.58,1.4,3.59,2.4s1.81,2.2,2.4,3.6s0.89,2.85,0.89,4.39c0,1.52-0.3,2.98-0.89,4.37s-1.4,2.59-2.4,3.59s-2.2,1.8-3.59,2.39 s-2.84,0.89-4.37,0.89c-1.53,0-3-0.3-4.39-0.89s-2.59-1.4-3.6-2.4s-1.8-2.2-2.4-3.58S3.74,16.03,3.74,14.5z M6.22,14.5 c0,2.37,0.86,4.43,2.59,6.18c1.73,1.73,3.79,2.59,6.2,2.59c1.58,0,3.05-0.39,4.39-1.18s2.42-1.85,3.21-3.2s1.19-2.81,1.19-4.39 s-0.4-3.05-1.19-4.4s-1.86-2.42-3.21-3.21s-2.81-1.18-4.39-1.18s-3.05,0.39-4.39,1.18S8.2,8.75,7.4,10.1S6.22,12.92,6.22,14.5z M11.11,20.35l3.75-13.11c0.01-0.1,0.06-0.15,0.15-0.15s0.14,0.05,0.15,0.15l3.74,13.11c0.04,0.11,0.03,0.19-0.02,0.25 s-0.13,0.06-0.24,0l-3.47-1.3c-0.1-0.04-0.2-0.04-0.29,0l-3.5,1.3c-0.1,0.06-0.17,0.06-0.21,0S11.09,20.45,11.11,20.35z",
-  "windy": "M4.65,15.5c0-0.22,0.08-0.41,0.23-0.56c0.16-0.15,0.35-0.22,0.57-0.22h12.08c0.22,0,0.4,0.07,0.54,0.22 c0.14,0.15,0.22,0.34,0.22,0.57c0,0.22-0.07,0.4-0.22,0.54c-0.14,0.14-0.32,0.22-0.54,0.22H5.45c-0.22,0-0.42-0.07-0.57-0.22 C4.72,15.9,4.65,15.72,4.65,15.5z M7.06,12.6c0-0.22,0.08-0.4,0.23-0.55c0.15-0.15,0.34-0.23,0.56-0.23h12.09 c0.21,0,0.39,0.08,0.54,0.23c0.15,0.15,0.22,0.33,0.22,0.55c0,0.22-0.07,0.4-0.22,0.56c-0.15,0.15-0.33,0.23-0.54,0.23H7.86 c-0.22,0-0.41-0.08-0.56-0.23S7.06,12.82,7.06,12.6z M8.68,18.34c0-0.21,0.08-0.39,0.24-0.54c0.14-0.14,0.32-0.22,0.54-0.22h12.1 c0.22,0,0.41,0.07,0.56,0.22c0.15,0.14,0.22,0.32,0.22,0.54s-0.08,0.41-0.23,0.56s-0.34,0.23-0.56,0.23H9.46 c-0.22,0-0.4-0.08-0.56-0.23S8.68,18.56,8.68,18.34z M19.26,15.5c0-0.23,0.07-0.42,0.22-0.57c0.15-0.15,0.34-0.22,0.57-0.22h4.52 c0.23,0,0.42,0.07,0.57,0.22c0.15,0.15,0.22,0.34,0.22,0.56c0,0.22-0.07,0.4-0.22,0.54c-0.15,0.14-0.34,0.22-0.56,0.22h-4.52 c-0.23,0-0.42-0.07-0.57-0.22C19.33,15.9,19.26,15.72,19.26,15.5z",
-}
-
-function WiIcon({ name, size = 19, style }: { name: string; size?: number; style?: React.CSSProperties }) {
-  const d = WI_PATHS[name]
-  if (!d) return null
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width={size} height={size} viewBox="0 0 30 30" fill="currentColor" style={{ flexShrink: 0, ...style }}>
-      <path d={d} />
-    </svg>
-  )
-}
-
-// Sky-state icon from cloud cover % alone — shared by the WMO-code path (codes 0-3, where
-// cloud_cover_pct is more precise than the coarse API code) and the no-code fallback path
-// (providers, like 7Timer, that never supply a WMO code at all). Only the clear-sky tier
-// distinguishes moon visibility (wi-stars vs wi-night-clear) — once there's cloud cover,
-// the clouds already block star/moon visibility, so partly-cloudy/cloudy don't need a
-// separate moon-aware variant.
-function skyIconFromCloudCover(cloudCover: number, moonUp: boolean, size: number) {
-  if (cloudCover <= 25) return <WiIcon name={moonUp ? 'night-clear' : 'stars'} size={size} />
-  if (cloudCover <= 65) return <WiIcon name="cloud" size={size} />
-  return <WiIcon name="cloudy" size={size} />
-}
-
-// WMO code → weather-icons name, for codes >= 4 (codes 0-3 use skyIconFromCloudCover
-// instead, since cloud_cover_pct is more precise there). Fog (45/48) is deliberately
-// absent — it renders as bold text in WmoIcon, not an icon.
-const WMO_ICON_NAMES: Record<number, string> = {
-  51: 'showers', 53: 'showers', 55: 'showers',
-  56: 'rain-mix', 57: 'rain-mix',
-  61: 'rain', 63: 'rain', 65: 'rain-wind',
-  66: 'rain-mix', 67: 'rain-mix',
-  71: 'snow', 73: 'snow', 75: 'snow', 77: 'sleet',
-  80: 'showers', 81: 'showers', 82: 'rain-wind',
-  85: 'snow', 86: 'snow',
-  95: 'thunderstorm', 96: 'thunderstorm', 99: 'thunderstorm',
-}
-
-// precip_type → weather-icons name, for providers that supply precip_type but no numeric
-// WMO code (7Timer). "fog" is absent here too, same reason as WMO_ICON_NAMES above.
-const PRECIP_TYPE_ICON_NAMES: Record<string, string> = {
-  frzr: 'rain-mix',
-  icep: 'sleet',
-  snow: 'snow',
-  rain: 'rain',
-  tstorm: 'thunderstorm',
-}
-
-// Transparency label for the combined Seeing/Transp column — the standard label as-is
-// (not an abbreviation). Explicit map, not a passthrough, so any unexpected value from
-// the provider (transparency is a generic string, not a strict union) still falls back
-// gracefully via the `?? p.transparency` at the call site.
-const TRANSP_ABBR: Record<string, string> = {
-  Excellent: 'Optimal',
-  Good: 'Good',
-  Fair: 'Fair',
-  Poor: 'Poor',
-}
-
-// Seeing (arcsec, lower = better) → the same 4-tier scale as TRANSP_ABBR, so the two
-// halves of the combined column read consistently ("Optimal / Good", not "Exc / Good").
-function seeingTier(arcsec: number): 'Optimal' | 'Good' | 'Fair' | 'Poor' {
-  if (arcsec < 1.0) return 'Optimal'
-  if (arcsec <= 1.5) return 'Good'
-  if (arcsec <= 2.0) return 'Fair'
-  return 'Poor'
-}
-
-function WmoIcon({ code, cloudCover, moonUp = false, size = 32, aod, pm25, visibilityM, precipType, windSpeedMs, windGustMs, transparency }: {
-  code: number | null; cloudCover?: number | null; moonUp?: boolean; size?: number
-  aod?: number | null; pm25?: number | null; visibilityM?: number | null; precipType?: string | null
-  windSpeedMs?: number | null; windGustMs?: number | null; transparency?: string | null
-}) {
-  // Fog and haze render as bold text, not an icon — both are visibility hazards easy to
-  // misjudge from a small glyph at a glance; the word is unambiguous.
-  if (precipType === 'fog') return <span className="wx-cond-word">Fog</span>
-
-  const isNoPrecip = precipType == null || precipType === 'none'
-  // Thresholds mirror rate_conditions()'s limiter curves: AOD 0.3 and PM2.5 35 µg/m³ are
-  // both where the gentle taper gives way to the steep power-curve penalty. Checked
-  // independently (not aod-priority) — a shallow, trapped surface smoke layer can read
-  // hazardous on PM2.5 while satellite column AOD still looks moderate.
-  const aodHazy = isNoPrecip && aod != null && aod > 0.3
-  const pmHazy  = isNoPrecip && pm25 != null && pm25 > 35
-  const visHazy = isNoPrecip && visibilityM != null && visibilityM < 10000 && visibilityM >= 1000
-  // AOD is the column (satellite/model) measurement — an elevated-but-surface-clean
-  // reading means the aerosol is aloft, decoupled from the ground. PM2.5 and horizontal
-  // visibility are both near-surface measurements, so either one without AOD means the
-  // haze is surface-coupled. Both together is the unqualified, worst-case "Haze".
-  const surfaceHazy = pmHazy || visHazy
-  // Cross-check: an AOD-only ("Aloft") reading is discarded when the independently
-  // sourced (7Timer) transparency assessment already says Good or better — the two
-  // would otherwise contradict each other in the same cell (AOD says hazy, transparency
-  // says clear). Below that (Fair/Poor), the signals agree rather than conflict, so AOD
-  // still gets to show. If transparency is unavailable at all (date outside 7Timer's
-  // range, or the API is down), there's nothing to cross-check against — fall back to
-  // the AOD signal alone rather than silently discarding a real reading for no reason.
-  const transpContradicts = transparency === 'Excellent' || transparency === 'Good'
-  const aloftConfirmed = aodHazy && !transpContradicts
-  // "Hazy" describes a visible-but-degraded sky — once cloud cover crosses into the same
-  // "cloudy" tier skyIconFromCloudCover uses for overcast, the sky isn't visible at all,
-  // so the claim is moot regardless of what AOD/PM2.5 say underneath or above the deck.
-  // Unlike the wind check below (which combines with cloud state instead of gating on
-  // it), there's no meaningful "hazy AND overcast" icon — it's just overcast.
-  const skyObscured = cloudCover != null && cloudCover > 25
-  if (!skyObscured && (aloftConfirmed || surfaceHazy)) {
-    const label = aloftConfirmed && surfaceHazy ? 'Haze' : aloftConfirmed ? 'Haze: Aloft' : 'Haze: Surface'
-    const readings = [
-      aod != null ? `AOD ${aod.toFixed(2)} (satellite column)` : null,
-      pm25 != null ? `PM2.5 ${Math.round(pm25)} µg/m³ (surface)` : null,
-      visibilityM != null ? `visibility ${Math.round(visibilityM / 1000)} km` : null,
-    ].filter(Boolean).join(' · ')
-    return (
-      <InfoTip tip={<>Smoke / aerosol haze — dims stars and flattens contrast even under a cloudless sky. {readings}. Penalized in the hourly rating past AOD 0.3 or PM2.5 35.</>}>
-        <span className="wx-cond-word">{label}</span>
-      </InfoTip>
-    )
-  }
-
-  // Wind/gust: only when there's no active precip (a rain/snow/fog icon already implies
-  // wind may be part of it, e.g. wi-rain-wind). Two severities:
-  //   "gusty"  — sustained OR gust crosses 20 mph (~8.94 m/s). This threshold is specific
-  //              to the Conditions icon, deliberately decoupled from the Wind cell's own
-  //              highlight threshold (6.3 m/s / ~14 mph, see windSevere/gustSevere below),
-  //              which stays unchanged by design.
-  //   "gate"   — the worse of sustained/gust reaches 17 m/s, the same cap where
-  //              rate_conditions()'s wind limiter already hits zero (a de facto hard gate).
-  // Gate takes priority over gusty when both are true. Needs cloudCover to pick between
-  // the clear-sky and cloudy variants; skip entirely if cloud state is unknown.
-  if (isNoPrecip && cloudCover != null) {
-    const windValues = [windSpeedMs, windGustMs].filter((v): v is number => v != null)
-    const effectiveWind = windValues.length ? Math.max(...windValues) : null
-    const isGate  = effectiveWind != null && effectiveWind >= 17
-    const isGusty = !isGate && effectiveWind != null && effectiveWind >= 8.94 // 20 mph
-    const isClear = cloudCover <= 25
-    if (isGate) return <WiIcon name={isClear ? 'strong-wind' : 'cloudy-windy'} size={size} />
-    if (isGusty) return <WiIcon name={isClear ? 'strong-wind' : 'cloudy-gusts'} size={size} />
-  }
-
-  if (code != null) {
-    // For pure sky-state codes (0–3), cloud cover % is more precise than the API code
-    if (code <= 3 && cloudCover != null) return skyIconFromCloudCover(cloudCover, moonUp, size)
-    const name = WMO_ICON_NAMES[code]
-    if (name) return <WiIcon name={name} size={size} />
-  }
-
-  // No numeric WMO code available — e.g. 7Timer serving as full primary because the
-  // Open-Meteo provider is down. It still supplies cloud_cover_pct and precip_type
-  // directly, so fall back to those instead of rendering nothing.
-  if (!isNoPrecip) {
-    const name = PRECIP_TYPE_ICON_NAMES[precipType!]
-    if (name) return <WiIcon name={name} size={size} />
-  }
-  if (cloudCover != null) return skyIconFromCloudCover(cloudCover, moonUp, size)
-  return null
-}
-
-// Horizontal coordinates in the conventional order: azimuth (with compass point)
-// first, then altitude — e.g. "Az 195° S · Alt 42°".
-// "Mon, Jul 13" — date-only strings are report-local, render without TZ shift
-function fmtNightDate(iso: string): string {
-  return new Date(iso + 'T00:00:00').toLocaleDateString('en-US', {
-    weekday: 'short', month: 'short', day: 'numeric',
-  })
-}
-
-function fmtPos(altDeg: number, azDeg: number): string {
-  return `Az ${Math.round(azDeg)}° ${cardinal(azDeg)} · Alt ${Math.round(altDeg)}°`
-}
-
-// Phase A of the "View Details" terminal-recalculation sequence: while a
-// date-only fetch is in flight, value cells render this placeholder instead of
-// their (now-stale) content — row/column structure never changes, so there's
-// no layout shift to guard against.
-function cell(isFetching: boolean, value: React.ReactNode): React.ReactNode {
-  return isFetching ? '—' : value
-}
-
-// ── Weather table ────────────────────────────────────────────────────────────
-
-function WeatherTable({ points, events = [], tz, imperial, moonrise, moonset, isFetching = false, cathodeSnap = false, wxSource = null, wxFetchedAt = null }: {
-  points: WeatherPoint[]
-  events?: SkyEvent[]
-  tz: string
-  imperial: boolean
-  moonrise?: string | null
-  moonset?: string | null
-  isFetching?: boolean
-  cathodeSnap?: boolean
-  wxSource?: string | null
-  wxFetchedAt?: string | null
-}) {
-  // Clip the table to the sunset→sunrise window. Events/points outside this
-  // range are daytime and not useful once the astro-night band conveys darkness.
-  // sunriseTs must be the sunrise AFTER sunset — for western-US summer dates the
-  // events array can start with an early-morning sunrise (end of the previous night,
-  // UTC date = target date) before the actual sunset that starts the target night.
-  const sunsetTs  = (() => { const e = events.find(e => e.label.toLowerCase().includes('sunset'));  return e ? new Date(e.time).getTime() : -Infinity })()
-  const sunriseTs = (() => {
-    const afterSunset = events.find(e => e.label.toLowerCase().includes('sunrise') && new Date(e.time).getTime() > sunsetTs)
-    if (afterSunset) return new Date(afterSunset.time).getTime()
-    const first = events.find(e => e.label.toLowerCase().includes('sunrise'))
-    return first ? new Date(first.time).getTime() : Infinity
-  })()
-  const visiblePoints = points.filter(p => { const t = new Date(p.time).getTime(); return t >= sunsetTs && t <= sunriseTs })
-  const visibleEvents = events.filter(e => {
-    const t = new Date(e.time).getTime()
-    return t >= sunsetTs && t <= sunriseTs
-  })
-
-  const hasTemp   = visiblePoints.some(p => p.temperature_c   != null)
-  const hasDew    = visiblePoints.some(p => p.dew_point_c     != null)
-  const hasSeeing = visiblePoints.some(p => p.seeing_arcsec   != null)
-  const hasTransp = visiblePoints.some(p => p.transparency    != null)
-  const hasAtmos  = hasSeeing || hasTransp
-  const hasWx     = visiblePoints.length > 0
-  const totalCols = 5 + (hasAtmos ? 1 : 0) + ((hasTemp || hasDew) ? 1 : 0)
-
-  // Astronomical night window (sun-based, matches the "ASTRO DARK BEGINS/ENDS"
-  // divider rows below) — NOT darkIntervals/dark_intervals, which is a different
-  // concept (astro night intersected with moon-below-horizon) that goes empty
-  // whenever the moon is up past the crescent exemption, even on a clear night.
-  const astroStartTs = (() => { const e = events.find(e => e.label.toLowerCase().includes('astronomical night') && e.label.toLowerCase().includes('begins')); return e ? new Date(e.time).getTime() : null })()
-  const astroEndTs   = (() => { const e = events.find(e => e.label.toLowerCase().includes('astronomical night') && e.label.toLowerCase().includes('ends'));   return e ? new Date(e.time).getTime() : null })()
-
-  // Moon up/down at an arbitrary timestamp, from the full sky-events list — used to
-  // pick between the full-dark and moonlit-dark shades within the astro window.
-  // Generalizes to a full moonrise→moonset→moonrise cycle in one night (long winter
-  // nights), unlike moonUpAt() in format.ts, which only handles one rise/set pair.
-  // A transition event's own timestamp takes the state it introduces (inclusive).
-  const sortedMoonEvents = events
-    .filter(e => { const l = e.label.toLowerCase(); return l.includes('moonrise') || l.includes('moonset') })
-    .sort((a, b) => new Date(a.time).getTime() - new Date(b.time).getTime())
-  function moonUpAtTs(ts: number): boolean {
-    const lastAtOrBefore = [...sortedMoonEvents].reverse().find(e => new Date(e.time).getTime() <= ts)
-    if (lastAtOrBefore) return lastAtOrBefore.label.toLowerCase().includes('moonrise')
-    const firstAfter = sortedMoonEvents.find(e => new Date(e.time).getTime() > ts)
-    if (!firstAfter) return false
-    return !firstAfter.label.toLowerCase().includes('moonrise')
-  }
-
-  // Determine moon state at sunset so the Sunset row can always orient the user.
-  // Uses the full events array (moon events aren't window-clipped) and falls back to
-  // inferring state from the first post-sunset event when no pre-sunset event exists.
-  const moonStateAtSunset: 'above' | 'below' | null = (() => {
-    if (sunsetTs === -Infinity) return null
-    const allMoonEvents = events
-      .filter(e => { const l = e.label.toLowerCase(); return l.includes('moonrise') || l.includes('moonset') })
-      .sort((a, b) => new Date(a.time).getTime() - new Date(b.time).getTime())
-    if (allMoonEvents.length === 0) return null
-    const lastBefore = [...allMoonEvents].reverse().find(e => new Date(e.time).getTime() <= sunsetTs)
-    if (lastBefore) return lastBefore.label.toLowerCase().includes('moonrise') ? 'above' : 'below'
-    const firstAfter = allMoonEvents.find(e => new Date(e.time).getTime() > sunsetTs)
-    if (!firstAfter) return null
-    return firstAfter.label.toLowerCase().includes('moonrise') ? 'below' : 'above'
-  })()
-
-  // Live "now" marker — updates every 30s so it stays accurate if the page is
-  // left open. Only injected when the current moment is inside the night window.
-  const [nowTs, setNowTs] = useState(Date.now)
-  useEffect(() => {
-    const id = setInterval(() => {
-      // Skip update when the tab is hidden — avoids re-sorting hidden content.
-      if (document.visibilityState === 'visible') setNowTs(Date.now())
-    }, 30_000)
-    return () => clearInterval(id)
-  }, [])
-  const isLive = sunsetTs !== -Infinity && sunriseTs !== Infinity
-               && nowTs > sunsetTs && nowTs < sunriseTs
-
-  type Row = { kind: 'event'; ev: SkyEvent; ts: number } | { kind: 'wx'; pt: WeatherPoint; ts: number } | { kind: 'now'; ts: number }
-
-  // Memoize the static sorted rows (events + weather points). These only change when
-  // the underlying data changes, not every 30s when the "now" marker ticks.
-  const baseRows = useMemo<Row[]>(() => [
-    ...visibleEvents.map(ev => ({ kind: 'event' as const, ev, ts: new Date(ev.time).getTime() })),
-    ...visiblePoints.map(pt => ({ kind: 'wx'    as const, pt, ts: new Date(pt.time).getTime() })),
-  ].sort((a, b) => a.ts - b.ts), [visibleEvents, visiblePoints])
-
-  // Insert the live "now" row at the correct sorted position on each tick.
-  // Suppressed while recalculating — showing "now" against stale, about-to-be-
-  // replaced rows is misleading mid-fetch.
-  const rows = useMemo<Row[]>(() => {
-    if (!isLive || isFetching) return baseRows
-    const nowRow: Row = { kind: 'now', ts: nowTs }
-    const insertAt = baseRows.findIndex(r => r.ts > nowTs)
-    if (insertAt === -1) return [...baseRows, nowRow]
-    return [...baseRows.slice(0, insertAt), nowRow, ...baseRows.slice(insertAt)]
-  }, [baseRows, isLive, isFetching, nowTs])
-
-  const ip = { size: 12, strokeWidth: 1.5, style: { flexShrink: 0 } } as const
-  // Sunrise/sunset/moonrise/moonset/astronomical-night all render as text dividers
-  // (see isDivider below) — twilight is the only event type that still shows an icon.
-  function evIcon(label: string) {
-    const l = label.toLowerCase()
-    if (l.includes('twilight')) return <Star {...ip} />
-    return null
-  }
-  function evClass(label: string): string {
-    const l = label.toLowerCase()
-    if (l.includes('sunrise') || l.includes('sunset')) return 'wx-ev-sun'
-    if (l.includes('moonrise') || l.includes('moonset')) return 'wx-ev-moon'
-    if (l.includes('astronomical night')) return 'wx-ev-astro'
-    return ''
-  }
-
-  // Collapsed-summary takeaway: cloud range + worst wind across the night.
-  const wxBrief = (() => {
-    if (!visiblePoints.length) return null
-    const clouds = visiblePoints.map(p => p.cloud_cover_pct).filter((v): v is number => v != null)
-    const parts: string[] = []
-    if (clouds.length) {
-      const minC = Math.round(Math.min(...clouds))
-      const maxC = Math.round(Math.max(...clouds))
-      parts.push(maxC <= 15 ? 'Clear' : `Clouds ${minC}–${maxC}%`)
-    }
-    const winds = visiblePoints.map(p => Math.max(p.wind_speed_ms ?? 0, p.wind_gust_ms ?? 0))
-    const maxWind = winds.length ? Math.max(...winds) : 0
-    parts.push(maxWind < 3 ? 'wind calm' : `wind to ${fmtWindSpeed(maxWind, imperial)} ${windUnitLabel(imperial)}`)
-    return parts.join(' · ')
-  })()
-
-  return (
-    <details id="report-timeline" className="wx-details" open>
-      <summary>
-        Night Timeline
-        {sunsetTs !== -Infinity && sunriseTs !== Infinity &&
-          `: ${formatDayTime(new Date(sunsetTs).toISOString(), tz)} → ${formatDayTime(new Date(sunriseTs).toISOString(), tz)}`}
-        {wxBrief && <span className="sum-brief"> · {wxBrief}</span>}
-      </summary>
-      <div className={`wx-table-wrap${isFetching ? ' is-recalculating' : ''}${cathodeSnap ? ' cathode-snap' : ''}`}>
-        {isFetching && (
-          <div className="recalc-overlay" role="status" aria-live="polite">
-            RECALCULATING...<span className="recalc-cursor" aria-hidden="true">█</span>
-          </div>
-        )}
-        <table className="wx-table">
-          {hasWx && (
-            <thead>
-              <tr>
-                <th>Time</th>
-                <th className="wx-cond-col"></th>
-                <th className="wx-cloud-col wx-cloud-hdr">Cloud %<br />Low/Med/High</th>
-                {hasAtmos  && (
-                  <th className="wx-atmos-col wx-atmos-hdr">
-                    <InfoTip tip={<>Seeing — turbulence blur in arcseconds: under 1.5″ is pin-sharp, over 2.5″ smears stars (matters most at long focal lengths). Transparency — how cleanly light passes the air column: Excellent → Poor. Both from 7Timer's astro forecast.</>}>
-                      Seeing /<br />Transp.
-                    </InfoTip>
-                  </th>
-                )}
-                {(hasTemp || hasDew) && <th className="wx-temp-col wx-temp-hdr">TEMP/<br />DEW PT</th>}
-                <th className="wx-wind-col">Wind</th>
-              </tr>
-            </thead>
-          )}
-          <tbody>
-            {rows.map((row, i) => {
-              if (row.kind === 'now') {
-                return (
-                  <tr key="now-marker" className="wx-now-row">
-                    <td className="wx-time wx-now-time">{formatTime(new Date(row.ts).toISOString(), tz)}</td>
-                    <td colSpan={hasWx ? totalCols - 1 : 1} className="wx-now-content">
-                      <span className="wx-now-inner">▶ Now</span>
-                    </td>
-                  </tr>
-                )
-              }
-              if (row.kind === 'event') {
-                const icon    = evIcon(row.ev.label)
-                const cls     = evClass(row.ev.label)
-                const l       = row.ev.label.toLowerCase()
-                const isSunset   = l.includes('sunset')
-                const isSunrise  = l.includes('sunrise')
-                const isMoonrise = l.includes('moonrise')
-                const isMoonset  = l.includes('moonset')
-                const isAstro    = l.includes('astronomical night')
-                // Sunrise/sunset/moonrise/moonset/astro-night-begin/end all render as
-                // the same plain, centered text divider — no icon, no italic label.
-                const isDivider  = isSunrise || isSunset || isMoonrise || isMoonset || isAstro
-                // Shade any event row landing inside the dark window (e.g. a moonrise
-                // or moonset mid-window), not just the begin/end markers themselves —
-                // lighter tone while the moon is up, full tone while it's down.
-                const inDarkWindow  = astroStartTs != null && astroEndTs != null && row.ts >= astroStartTs && row.ts <= astroEndTs
-                const astroShadeCls = inDarkWindow ? (moonUpAtTs(row.ts) ? 'wx-row-astro-lit' : 'wx-row-astro') : ''
-
-                const dividerText = isAstro
-                  ? (l.includes('begins') ? 'ASTRO DARK BEGINS' : 'ASTRO DARK ENDS')
-                  : isSunset
-                  ? (moonStateAtSunset === 'above' ? 'SUNSET - MOON UP' : moonStateAtSunset === 'below' ? 'SUNSET - MOON DOWN' : 'SUNSET')
-                  : isSunrise
-                  ? 'SUNRISE'
-                  : isMoonrise
-                  ? 'MOONRISE'
-                  : 'MOONSET'
-
-                return (
-                  <tr key={`ev-${i}`} className={`wx-ev-row${cls ? ` ${cls}` : ''}${astroShadeCls ? ` ${astroShadeCls}` : ''}`}>
-                  <td className="wx-time wx-ev-time">{formatTime(row.ev.time, tz)}</td>
-                  <td colSpan={hasWx ? totalCols - 1 : 1} className="wx-ev-content">
-                  <span className="wx-ev-inner">
-                  {icon && !isDivider && <span className="wx-ev-icon">{icon}</span>}
-
-                  {isDivider ? (
-                    <span className="wx-ev-divider">{dividerText}</span>
-                  ) : (
-                    <span className="wx-ev-label">{row.ev.label}</span>
-                  )}
-            </span>
-          </td>
-        </tr>
-  )
-}
-              const p = row.pt
-              const inAstroWindow = astroStartTs != null && astroEndTs != null && row.ts >= astroStartTs && row.ts <= astroEndTs
-              const astroShadeCls = inAstroWindow ? (moonUpAtTs(row.ts) ? 'wx-row-astro-lit' : 'wx-row-astro') : undefined
-              const windSevere = p.wind_speed_ms != null && p.wind_speed_ms >= 6.3
-              const gustSevere = p.wind_gust_ms  != null && p.wind_gust_ms  >= 6.3
-              const dewSpread = p.temperature_c != null && p.dew_point_c != null ? p.temperature_c - p.dew_point_c : null
-              const dewGate  = dewSpread != null && dewSpread <= 5
-              return (
-                <tr key={`wx-${i}`} className={astroShadeCls}>
-                  <td className="wx-time">{formatTime(p.time, tz)}</td>
-                  <td className="wx-num wx-rating">
-                    {cell(isFetching, <WmoIcon code={p.weather_code} cloudCover={p.cloud_cover_pct}
-                      moonUp={moonUpAt(p.time, moonrise ?? null, moonset ?? null)}
-                      aod={p.aerosol_optical_depth} pm25={p.pm2_5} visibilityM={p.visibility_m} precipType={p.precip_type}
-                      windSpeedMs={p.wind_speed_ms} windGustMs={p.wind_gust_ms} transparency={p.transparency} />)}
-                  </td>
-                  <td className="wx-num wx-cloud-col">
-                    {cell(isFetching, (() => {
-                      if (p.cloud_cover_low_pct == null && p.cloud_cover_mid_pct == null && p.cloud_cover_high_pct == null) {
-                        return p.cloud_cover_pct != null ? `${p.cloud_cover_pct}%` : '—'
-                      }
-                      const low = p.cloud_cover_low_pct ?? 0
-                      const mid = p.cloud_cover_mid_pct ?? 0
-                      const high = p.cloud_cover_high_pct ?? 0
-                      return (
-                        <div className="wx-cloud-readout" title={`Low: ${low}% | Mid: ${mid}% | High: ${high}%`}>
-                          <span className="wx-cloud-val">{p.cloud_cover_pct != null ? `${p.cloud_cover_pct}%` : '—'}</span>
-                          <div className="wx-cloud-eq">
-                            <div className="wx-eq-track"><div className="wx-eq-fill" style={{ height: `${low}%` }} /></div>
-                            <div className="wx-eq-track"><div className="wx-eq-fill" style={{ height: `${mid}%` }} /></div>
-                            <div className="wx-eq-track"><div className="wx-eq-fill" style={{ height: `${high}%` }} /></div>
-                          </div>
-                        </div>
-                      )
-                    })())}
-                  </td>
-                  {hasAtmos && (
-                    <td className="wx-num wx-atmos-col">
-                      {cell(isFetching, <span className="wx-atmos-inner">
-                        {p.seeing_arcsec != null && (() => {
-                          const tier = seeingTier(p.seeing_arcsec!)
-                          return (
-                            <span
-                              className={`wx-atmos-seeing${(tier === 'Fair' || tier === 'Poor') ? ' wx-gate-warn' : ''}`}
-                              title={`Seeing: ${p.seeing_arcsec!.toFixed(2)}"`}
-                            >
-                              {tier}
-                            </span>
-                          )
-                        })()}
-                        {p.seeing_arcsec != null && p.transparency != null && (
-                          <span className="wx-atmos-sep">/</span>
-                        )}
-                        {p.transparency != null && (
-                          <span className={`wx-atmos-transp${(p.transparency === 'Fair' || p.transparency === 'Poor') ? ' wx-gate-warn' : ''}`}>
-                            {TRANSP_ABBR[p.transparency] ?? p.transparency}
-                          </span>
-                        )}
-                        {p.seeing_arcsec == null && p.transparency == null && '—'}
-                      </span>)}
-                    </td>
-                  )}
-                  {(hasTemp || hasDew) && (
-                    <td className="wx-num wx-temp-col">
-                      {cell(isFetching, <>
-                        {p.temperature_c != null && <span className="wx-temp-val">{fmtTempValue(p.temperature_c, imperial)}</span>}
-                        {p.temperature_c != null && p.dew_point_c != null && <span className="wx-temp-sep">/</span>}
-                        {p.dew_point_c != null && (
-                          <span className={`wx-temp-val wx-dew-val${dewGate ? ' wx-dew-warn' : ''}`}>
-                            {fmtTempValue(p.dew_point_c, imperial)}
-                          </span>
-                        )}
-                        {(p.temperature_c != null || p.dew_point_c != null) && (
-                          <span className="wx-temp-unit">{tempUnitLabel(imperial)}</span>
-                        )}
-                      </>)}
-                    </td>
-                  )}
-                  <td className={`wx-num wx-wind-col${windSevere ? ' wx-gate-warn' : ''}`}>
-                {cell(isFetching, <>
-                {p.wind_speed_ms != null ? (
-                  <>
-                    <span className="wx-wind-val">{fmtWindSpeed(p.wind_speed_ms, imperial)}</span>
-                    {p.wind_gust_ms != null && (
-                      <>
-                        <span className="wx-wind-sep">/</span>
-                        <span className={`wx-wind-val wx-wind-gust${gustSevere ? ' wx-wind-gust-severe' : ''}`}>
-                          {fmtWindSpeed(p.wind_gust_ms, imperial)}
-                        </span>
-                      </>
-                    )}
-                    <span className="wx-wind-unit">{windUnitLabel(imperial)}</span>
-                  </>
-                ) : '—'}
-
-                {p.wind_direction_deg != null && (
-                  // wi-wind-deg's default orientation + weather-icons' own "towards-*"
-                  // rotation classes assume the "towards" convention, but our data is
-                  // "from" (standard meteorological convention — see wind.py). Rotate
-                  // by +180° so the arrow points where the wind is heading, matching
-                  // typical consumer weather-app wind arrows.
-                  <WiIcon name="wind-deg" size={24}
-                  style={{ marginLeft: 4, verticalAlign: 'middle', transform: `rotate(${(p.wind_direction_deg + 180) % 360}deg)` }} />
-                )}
-            </>)}
-                </td>
-                </tr>
-              )
-            })}
-          </tbody>
-        </table>
-        {hasWx && <WxProvenanceBadge source={wxSource} fetchedAt={wxFetchedAt} />}
-      </div>
-    </details>
-  )
-}
-
-// Dense, monospace "instrument readout" footer for the Night Timeline table — shows
-// which weather source served this forecast and how fresh the data is.
-function WxProvenanceBadge({ source, fetchedAt }: { source: string | null; fetchedAt: string | null }) {
-  if (!source) return null
-  return (
-    <div className="wx-provenance">
-      <span className="wx-prov-item">SOURCE: {source.toUpperCase()}</span>
-      <span className="wx-prov-sep">·</span>
-      <span className="wx-prov-item">ISSUED: {formatIssuedUtc(fetchedAt)}</span>
-      <span className="wx-prov-sep">·</span>
-      <span className="wx-prov-item">UPDATED: {formatAge(fetchedAt)}</span>
-    </div>
-  )
-}
-
-// ── Metadata row ─────────────────────────────────────────────────────────────
-
-function MetaRow({ k, v, icon, tip }: { k: string; v: string; icon?: React.ReactNode; tip?: React.ReactNode }) {
-  return (
-    <div className="meta-row">
-      <span className="meta-k">{tip ? <InfoTip tip={tip}>{k}</InfoTip> : k}:</span>
-      <span className="meta-v" style={icon ? { display: 'inline-flex', alignItems: 'center', gap: 6 } : undefined}>
-        {icon}{v}
-      </span>
-    </div>
-  )
-}
-
-// ── Satellite passes ─────────────────────────────────────────────────────────
-
-function SatellitePasses({ report, isFetching = false, cathodeSnap = false }: { report: NightReport; isFetching?: boolean; cathodeSnap?: boolean }) {
-  const tz = report.tz_name
-
-  // Unavailability notices
-  if (report.sat_network_error) {
-    return <p className="sat-notice">Satellite data unavailable: could not reach Celestrak and no cached TLE exists.</p>
-  }
-  if (report.sat_stale) {
-    return <p className="sat-notice">Satellite pass predictions require a current TLE; historical dates are not supported.</p>
-  }
-  if (report.sat_future_stale) {
-    return <p className="sat-notice">Satellite pass predictions longer than 7 days are highly inaccurate and TLE data cannot accurately predict this date.</p>
-  }
-
-  const notes: string[] = []
-  if (report.sat_tle_stale)   notes.push('Note: Using cached TLE data (Celestrak unreachable). Pass times may be slightly off.')
-  if (report.sat_future_warn) notes.push('Note: Pass times are approximate. TLE accuracy degrades beyond 3 days.')
-
-  const visible  = report.sat_passes.filter(p => p.in_sunlight && p.sky_dark)
-  const twilight = report.sat_passes.filter(p => p.in_sunlight && !p.sky_dark)
-  const shadow   = report.sat_passes.filter(p => !p.in_sunlight)
-  const display  = [...visible, ...twilight].sort((a, b) => a.rise_time.localeCompare(b.rise_time))
-
-  const trains = report.starlink_trains ?? []
-
-  const hasAny = display.length > 0 || trains.length > 0
-
-  if (!hasAny) {
-    const shadowMsg = shadow.length > 0
-      ? `${shadow.length} transit${shadow.length > 1 ? 's' : ''} tonight but in Earth's shadow: not visible.`
-      : 'No notable orbital transits this night.'
-    return (
-      <>
-        {notes.map((n, i) => <p key={i} className="sat-notice sat-note">{n}</p>)}
-        <p className="sat-notice">{shadowMsg}</p>
-        </>
-    )
-  }
-
-  const az = (deg: number) => `${deg.toFixed(0)}° ${cardinal(deg)}`
-
-  return (
-    <>
-      {notes.map((n, i) => <p key={i} className="sat-notice sat-note">{n}</p>)}
-
-      {trains.length > 0 && (
-        <div className="sat-trains">
-          <div className="sat-trains-label">Starlink Train{trains.length > 1 ? 's' : ''}:</div>
-          {trains.map((tr, i) => {
-            const moonStr = tr.moon_sep_deg != null ? `  ·  Moon Sep: ${tr.moon_sep_deg.toFixed(1)}°` : ''
-            const skyTag  = !tr.sky_dark ? '  [civil twilight]' : ''
-            const daysAgo = tr.launch_date
-              ? `  ·  launched ${Math.round((new Date(report.date).getTime() - new Date(tr.launch_date).getTime()) / 86400000)}d ago`
-              : ''
-            return (
-              <div key={i} className="sat-train-row">
-                {tr.satellite_count} satellites{daysAgo}{skyTag}  ·  {formatTime(tr.first_rise, tz)} – {formatTime(tr.last_rise, tz)}  ·  Peak {tr.peak_alt_deg.toFixed(0)}°  ·  from {az(tr.lead_az_deg)}{moonStr}
-              </div>
-            )
-          })}
-        </div>
-      )}
-
-      {display.length > 0 && (
-        <div className={`wx-table-wrap${isFetching ? ' is-recalculating' : ''}${cathodeSnap ? ' cathode-snap' : ''}`}>
-          {isFetching && (
-            <div className="recalc-overlay" role="status" aria-live="polite">
-              RECALCULATING...<span className="recalc-cursor" aria-hidden="true">█</span>
-            </div>
-          )}
-          <table className="wx-table sat-table">
-            <thead>
-              <tr>
-                <th>Satellite</th>
-                <th colSpan={3}>Rise</th>
-                <th colSpan={3}>Peak</th>
-                <th className="sat-set-col" colSpan={3}>Set</th>
-                <th className="sat-dur-col">Dur</th>
-                <th>Moon Sep</th>
-                {report.light_dome && <th>Glow</th>}
-              </tr>
-              <tr className="sat-subhdr">
-                <th></th>
-                <th>Time</th><th>Alt</th><th>Az</th>
-                <th>Time</th><th>Alt</th><th className="sat-peak-az-col">Az</th>
-                <th className="sat-set-col">Time</th><th className="sat-set-col">Alt</th><th className="sat-set-col">Az</th>
-                <th className="sat-dur-col"></th><th></th>
-                {report.light_dome && <th></th>}
-              </tr>
-            </thead>
-            <tbody>
-              {display.map((p, i) => {
-                const label    = p.satellite_name + (!p.sky_dark ? ' †' : '')
-                const setAlt   = `${p.set_alt_deg.toFixed(0)}°${p.ends_in_shadow ? '*' : ''}`
-                const moonSepLow = !p.moon_transit && p.moon_sep_deg != null && p.moon_sep_deg < 5
-                const moonStr  = p.moon_transit
-                  ? `TRANSIT ${p.moon_transit_sep_deg?.toFixed(3)}°`
-                  : p.moon_sep_deg != null
-                    ? `${p.moon_sep_deg.toFixed(1)}°`
-                    : '—'
-                const satGlow = report.light_dome
-                  ? glowToward(report.light_dome, p.peak_az_deg, p.peak_alt_deg)
-                  : null
-                const wxAtPeak = wxAtTime(report.weather_points || [], p.peak_time)
-                const satCloudy = wxAtPeak != null && wxAtPeak.cloud_cover_pct != null && wxAtPeak.cloud_cover_pct > 70
-                if (satCloudy) return (
-                  <tr key={i} className="tg-row-blocked">
-                    <td>{cell(isFetching, label)}</td>
-                    <td className="wx-num" colSpan={11 + (satGlow != null ? 1 : 0)}>
-                      {cell(isFetching, <span className="mw-moon-badge badge-poor">Clouded out</span>)}
-                    </td>
-                  </tr>
-                )
-                return (
-                  <tr key={i}>
-                    <td>{cell(isFetching, label)}</td>
-                    <td className="wx-num">{cell(isFetching, formatTime(p.rise_time, tz))}</td>
-                    <td className="wx-num">{cell(isFetching, `${p.rise_alt_deg.toFixed(0)}°`)}</td>
-                    <td className="wx-num">{cell(isFetching, az(p.rise_az_deg))}</td>
-                    <td className="wx-num">
-                      {cell(isFetching, <>
-                        {formatTime(p.peak_time, tz)}
-                        {wxAtPeak && (
-                          <span className={`tg-wx-inline wx-rating-${scoreBand(rateConditions(wxAtPeak))}`}>
-                            <WmoIcon code={wxAtPeak.weather_code} size={12} />
-                          </span>
-                        )}
-                      </>)}
-                    </td>
-                    <td className="wx-num">{cell(isFetching, `${p.peak_alt_deg.toFixed(0)}°`)}</td>
-                    <td className="wx-num sat-peak-az-col">{cell(isFetching, az(p.peak_az_deg))}</td>
-                    <td className="wx-num sat-set-col">{cell(isFetching, formatTime(p.set_time, tz))}</td>
-                    <td className="wx-num sat-set-col">{cell(isFetching, setAlt)}</td>
-                    <td className="wx-num sat-set-col">{cell(isFetching, az(p.set_az_deg))}</td>
-                    <td className="wx-num sat-dur-col">{cell(isFetching, `${p.duration_min.toFixed(0)}m`)}</td>
-                    <td className="wx-num" style={moonSepLow ? {color: 'var(--excellent)', fontWeight: 700, fontSize: '1rem'} : undefined}>{cell(isFetching, moonStr)}</td>
-                    {report.light_dome && (
-                      <td className="wx-num cond-glow" style={satGlow != null && satGlow >= 0.03 ? glowStyle(satGlow) : undefined}>
-                        {cell(isFetching, satGlow != null && satGlow >= 0.03 ? glowLabel(satGlow) : '—')}
-                      </td>
-                    )}
-                  </tr>
-                )
-              })}
-            </tbody>
-          </table>
-        </div>
-      )}
-
-      {(display.some(p => p.ends_in_shadow) || twilight.length > 0 || shadow.length > 0) && (
-        <div className="sat-footnotes">
-          {display.some(p => p.ends_in_shadow) && <div>* Set alt &gt; 10°: satellite entered Earth's shadow before geometric set</div>}
-          {twilight.length > 0 && <div>† Pass during civil twilight: sky too bright to observe</div>}
-          {shadow.length > 0 && <div>+{shadow.length} pass{shadow.length > 1 ? 'es' : ''} in Earth's shadow (not visible)</div>}
-        </div>
-      )}
-    </>
-  )
-}
-
-// ── Targets helpers ──────────────────────────────────────────────────────────
-
-// nebula / galaxy / cluster are collapsed into a single display group ("dso")
-// so the targets table shows a clean unlabeled DSO block followed by Planets.
-const DSO_TYPES = new Set(['nebula', 'galaxy', 'cluster'])
-
-const TYPE_ORDER: Record<string, number> = {
-  meteor_shower: 0,
-  dso:           1,  // nebula + galaxy + cluster
-  planet:        2,
-}
-const TYPE_LABELS: Record<string, string> = {
-  meteor_shower: 'Meteor Showers',
-  // 'dso' has no label — after prominence filtering the list is short enough
-  planet: 'Planets',
-}
-
-const MOON_ARCMIN = 30
-
-function moonScaleLabel(arcmin: number | null | undefined): string | null {
-  if (arcmin == null) return null
-  const ratio = arcmin / MOON_ARCMIN
-  if (ratio >= 1.5) return `${Math.round(ratio)}x Moon`
-  if (ratio >= 1.0) return '1x Moon'
-  if (ratio >= 0.5) return '½ Moon'
-  if (ratio >= 0.3) return '⅓ Moon'
-  return null
-}
-
-// ── Milky Way card ───────────────────────────────────────────────────────────
-
-// Waypoints disclosure — closed by default with Phase 3 density reductions applied inside.
-function WaypointsAccordion({ waypoints, summary, report }: {
-  waypoints: VisibleTarget[]
-  summary: MilkyWaySummary
-  report: NightReport
-}) {
-  const tz = report.tz_name
-  const detailsRef = useRef<HTMLDetailsElement>(null)
-
-  useEffect(() => {
-    const el = detailsRef.current
-    if (!el) return
-    const handler = () => {
-      if (el.open) el.scrollIntoView({ behavior: 'smooth', block: 'nearest' })
-    }
-    el.addEventListener('toggle', handler)
-    return () => el.removeEventListener('toggle', handler)
-  }, [])
-
-  return (
-    <details ref={detailsRef} className="mw-waypoints-detail">
-      <summary className="mw-waypoints-summary">
-        Galactic Plane Waypoints ({summary.n_visible})
-      </summary>
-      <div className="tg-table-wrap mw-waypoints-table-wrap">
-        <table className="tg-table">
-          <thead>
-            <tr>
-              <th>Waypoint</th>
-              <th>Best</th>
-              <th></th>
-              <th>Window</th>
-            </tr>
-          </thead>
-          <tbody>
-            {waypoints.map(t => {
-              const w = bestWindow(t)
-              if (!w.peak_time || w.peak_alt_deg == null) {
-                return (
-                  <tr key={t.name}>
-                    <td>{t.name}</td>
-                    <td className="wx-num">—</td>
-                    <td className="wx-num">—</td>
-                  </tr>
-                )
-              }
-              const archAngle = w.arch_angle_deg
-              const archBadge = archAngle != null && (archAngle < 35 || archAngle >= 60)
-                ? <span className="tg-note"> · {archAngle.toFixed(0)}° {archAngle >= 60 ? 'steep' : 'flat'}</span>
-                : null
-              const glow = report.light_dome
-                ? glowToward(report.light_dome, w.peak_az_deg, w.peak_alt_deg)
-                : null
-              const showGlow = glow != null && glow >= 0.03
-              const bestT = w.best_time ?? w.peak_time
-              const wxPt = !report.wx_no_data && !report.wx_pending
-                ? wxAtTime(report.weather_points || [], bestT)
-                : null
-              const waypointCloudy = wxPt != null && wxPt.cloud_cover_pct != null && wxPt.cloud_cover_pct > 70
-              if (waypointCloudy) return (
-                <tr key={t.name} className="tg-row-blocked">
-                  <td>{t.name}</td>
-                  <td className="wx-num" colSpan={3} style={{textAlign: 'center'}}>
-                    <span className="mw-moon-badge badge-poor">Clouded out</span>
-                  </td>
-                </tr>
-              )
-              const sky = skyCondition(
-                bestT, report.dark_intervals, report.night_start, report.night_end,
-                report.illumination_pct, report.moonrise, report.moonset,
-                w.moon_sep_at_peak_deg, w.moon_alt_at_peak_deg,
-              )
-              const moonBadge = sky.startsWith('Moon')
-                ? <span className={`tg-sky-inline ${skyClass(sky)}`}>{' '}{sky}</span>
-                : null
-              return (
-                <tr key={t.name}>
-                  <td>
-                    {t.name}
-                    {showGlow && (
-                      <span className="tg-glow-inline cond-glow" style={glowStyle(glow!)}>
-                        {` · glow ${glowLabel(glow!)}`}
-                      </span>
-                    )}
-                  </td>
-                  <td className="wx-num">
-                    <span className="tg-t">{formatTime(bestT, tz)}</span>
-                    <span className="tg-p"> · Alt </span>
-                    <span className="tg-alt">{Math.round(w.peak_alt_deg)}°</span>
-                    <span className="tg-p"> · Az </span>
-                    <span className="tg-az">{Math.round(w.peak_az_deg)}°</span>
-                    <span className="tg-p"> </span>
-                    <span className="tg-dir">{cardinal(w.peak_az_deg)}</span>
-                    {archBadge}
-                    {moonBadge}
-                  </td>
-                  <td className="wx-num tg-cond-col">
-                    {wxPt && (
-                      <span className={`tg-wx-inline wx-rating-${scoreBand(rateConditions(wxPt))}`}>
-                        <WmoIcon code={wxPt.weather_code} size={12} />
-                      </span>
-                    )}
-                  </td>
-                  <td className="wx-num wp-window-td">
-                    <span className="tg-t">{formatTime(w.start, tz)}</span>
-                    <span className="tg-p"> – </span>
-                    <span className="tg-t">{formatTime(w.end, tz)}</span>
-                  </td>
-                </tr>
-              )
-            })}
-          </tbody>
-        </table>
-      </div>
-    </details>
-  )
-}
-
-function MoonBadge({ type, severity }: { type: 'penalty' | 'limited'; severity?: string | null }) {
-  const base = type === 'penalty' ? 'Moon interference' : 'Moon limited'
-  const text = severity ? `${base}: ${severity}` : base
-  return (
-    <InfoTip tip={<>Moon wash — scattered moonlight brightening the sky along this line of sight (Krisciunas &amp; Schaefer 1991). Severity comes from phase, moon altitude, and angular separation, not illumination % alone.</>}>
-      <span className="mw-moon-badge">{text}</span>
-    </InfoTip>
-  )
-}
-
-
-// IAU (1958) galactic → ICRS rotation matrix (mirrors milky_way.py exactly).
-const _GAL_TO_ICRS = [
-  [-0.0548755604, +0.4941094279, -0.8676661490],
-  [-0.8734370902, -0.4448296300, -0.1980763734],
-  [-0.4838350155, +0.7469822445, +0.4559837762],
-]
-
-
-
-
-// Full galactic → horizontal (Alt/Az) transformation at a given UTC instant.
-// l_deg, b_deg   — galactic coordinates
-// lat_deg, lon_deg — observer position
-// utcMs          — UTC milliseconds since Unix epoch
-function galToAltAz(l_deg: number, b_deg: number, lat_deg: number, lon_deg: number, utcMs: number) {
-  const toRad = (d: number) => d * Math.PI / 180
-  const l = toRad(l_deg), b = toRad(b_deg)
-  const xg = Math.cos(b) * Math.cos(l)
-  const yg = Math.cos(b) * Math.sin(l)
-  const zg = Math.sin(b)
-  const R  = _GAL_TO_ICRS
-  const xi = R[0][0]*xg + R[0][1]*yg + R[0][2]*zg
-  const yi = R[1][0]*xg + R[1][1]*yg + R[1][2]*zg
-  const zi = R[2][0]*xg + R[2][1]*yg + R[2][2]*zg
-  const ra_rad  = Math.atan2(yi, xi)
-  const dec_rad = Math.asin(Math.max(-1, Math.min(1, zi)))
-  // GMST (degrees) via Meeus Ch.12 — valid at any time of day, not just 0h UT
-  const jd      = utcMs / 86_400_000 + 2_440_587.5
-  const D       = jd - 2_451_545.0                    // days from J2000.0
-  const T       = D / 36_525.0
-  const gmst_deg = ((280.46061837 + 360.98564736629 * D + 0.000387933 * T * T - T * T * T / 38_710_000) % 360 + 360) % 360
-  const lst_rad  = toRad(gmst_deg + lon_deg)
-  const ha_rad  = lst_rad - ra_rad
-  const lat_rad = toRad(lat_deg)
-  const alt = Math.asin(
-    Math.sin(dec_rad) * Math.sin(lat_rad) +
-    Math.cos(dec_rad) * Math.cos(lat_rad) * Math.cos(ha_rad)
-  )
-  const az = Math.atan2(
-    -Math.cos(dec_rad) * Math.sin(ha_rad),
-    Math.sin(dec_rad) * Math.cos(lat_rad) - Math.cos(dec_rad) * Math.sin(lat_rad) * Math.cos(ha_rad)
-  )
-  return {
-    alt: alt * 180 / Math.PI,
-    az:  ((az  * 180 / Math.PI) + 360) % 360,
-  }
-}
-
-// Simplified moon position (Meeus Ch.47, largest perturbation terms).
-// Accurate to ~0.3° — sufficient for the dome visualization glow blob.
-function moonAltAz(lat: number, lon: number, utcMs: number): { alt: number; az: number } {
-  const r   = (d: number) => d * Math.PI / 180
-  const mod = (x: number) => ((x % 360) + 360) % 360
-  const JD  = utcMs / 86_400_000 + 2_440_587.5
-  const D   = JD - 2_451_545.0
-  const T   = D / 36_525.0
-
-  // Fundamental arguments (degrees)
-  const Lp = mod(218.3164477 + 481267.88123421 * T)
-  const Mp = mod(134.9633964 + 477198.8675055  * T)
-  const M  = mod(357.5291092 + 35999.0502909   * T)
-  const Dg = mod(297.8501921 + 445267.1114034  * T)
-  const F  = mod(93.2720950  + 483202.0175233  * T)
-
-  // Ecliptic longitude (10 largest terms, coefficients ×1e-6 degrees)
-  const ΣL = (
-    + 6288774 * Math.sin(r(Mp))
-    + 1274027 * Math.sin(r(2*Dg - Mp))
-    +  658314 * Math.sin(r(2*Dg))
-    +  213618 * Math.sin(r(2*Mp))
-    -  185116 * Math.sin(r(M))
-    -  114332 * Math.sin(r(2*F))
-    +   58793 * Math.sin(r(2*Dg - 2*Mp))
-    +   57066 * Math.sin(r(2*Dg - M - Mp))
-    +   53322 * Math.sin(r(2*Dg + Mp))
-    +   45758 * Math.sin(r(2*Dg - M))
-  ) / 1e6
-
-  // Ecliptic latitude (6 largest terms)
-  const ΣB = (
-    + 5128122 * Math.sin(r(F))
-    +  280602 * Math.sin(r(Mp + F))
-    +  277693 * Math.sin(r(Mp - F))
-    +  173237 * Math.sin(r(2*Dg - F))
-    +   55413 * Math.sin(r(2*Dg - Mp + F))
-    +   46271 * Math.sin(r(2*Dg - Mp - F))
-  ) / 1e6
-
-  const λ = r(mod(Lp + ΣL))
-  const β = r(ΣB)
-  const ε = r(23.4393 - 0.013004 * T)  // obliquity of ecliptic
-
-  // Ecliptic → equatorial
-  const ra  = Math.atan2(Math.sin(λ) * Math.cos(ε) - Math.tan(β) * Math.sin(ε), Math.cos(λ))
-  const dec = Math.asin(Math.sin(β) * Math.cos(ε) + Math.cos(β) * Math.sin(ε) * Math.sin(λ))
-
-  // GMST (Meeus Ch.12) → HA — same formula as galToAltAz
-  const gmst = r(mod(280.46061837 + 360.98564736629 * D + 0.000387933 * T * T - T * T * T / 38_710_000))
-  const ha   = gmst + r(lon) - ra
-  const latR = r(lat)
-
-  const sinAlt = Math.sin(dec) * Math.sin(latR) + Math.cos(dec) * Math.cos(latR) * Math.cos(ha)
-  const alt    = Math.asin(Math.max(-1, Math.min(1, sinAlt))) * 180 / Math.PI
-  const az     = (Math.atan2(
-    -Math.cos(dec) * Math.sin(ha),
-    Math.sin(dec) * Math.cos(latR) - Math.cos(dec) * Math.sin(latR) * Math.cos(ha),
-  ) * 180 / Math.PI + 360) % 360
-  return { alt, az }
-}
-
-// ── Light dome direction/glow utilities ──────────────────────────────────────
-// Defined before MilkyWayDome so archSegmentBrightness (below) can call glowToward.
-
-const LD_DIRS: Direction[] = ['N', 'NE', 'E', 'SE', 'S', 'SW', 'W', 'NW']
-const LD_DIR_AZ: Record<Direction, number> = { N:0, NE:45, E:90, SE:135, S:180, SW:225, W:270, NW:315 }
-// Thresholds mirror light_dome.py (minor_threshold / major_threshold) so the colour
-// scale is comparable across sites: green→amber at MINOR, →red by MAJOR.
-const LD_MINOR = 0.25
-const LD_MAJOR = 3.0
-// Log-interpolated colour stops (RGB), keyed on the thresholds, using the app's own
-// quality ramp (excellent → good → fair → poor). The darkest end fades to BLACK — the
-// "excellent of excellent" = true darkness. Glow rises: black → green → blue → lilac → rose.
-const LD_STOPS: [number, [number, number, number]][] = [
-  [0,        [0, 0, 0]],        // darkness — the best (excellent-of-excellent)
-  [0.03,     [74, 94, 168]],    // --excellent blue
-  [0.12,     [92, 184, 92]],    // --good green
-  [LD_MINOR, [240, 173, 78]],   // --fair amber — a minor dome
-  [0.9,      [217, 83, 79]],    // --poor red
-  [LD_MAJOR, [150, 40, 40]],    // deep --poor — a major dome
-]
-// Bloom-legibility transform: real dome heights are ~1° (a sub-pixel rim sliver), so
-// for *display* we scale them up and floor them. This shapes the on-screen bloom, not
-// the physics. Directions with no flagged dome get a small default height.
-const LD_THETA_K = 5
-const LD_THETA_FLOOR_DEG = 6
-const LD_THETA_DEFAULT_DEG = 4
-const LD_SIZE = 300            // CSS px; disk + room for N/E/S/W labels
-
-function ldColor(v: number): [number, number, number] {
-  if (v <= LD_STOPS[0][0]) return LD_STOPS[0][1]
-  for (let i = 1; i < LD_STOPS.length; i++) {
-    const [hv, hc] = LD_STOPS[i]
-    if (v <= hv) {
-      const [lv, lc] = LD_STOPS[i - 1]
-      const t = (Math.log(Math.max(v, 1e-4)) - Math.log(Math.max(lv, 1e-4))) /
-                (Math.log(hv) - Math.log(Math.max(lv, 1e-4)))
-      const k = Math.max(0, Math.min(1, t))
-      return [lc[0] + (hc[0] - lc[0]) * k, lc[1] + (hc[1] - lc[1]) * k, lc[2] + (hc[2] - lc[2]) * k]
-    }
-  }
-  return LD_STOPS[LD_STOPS.length - 1][1]
-}
-
-// Tent-interpolate a per-direction array at an arbitrary azimuth (partition of unity
-// across the two nearest cardinals — same scheme as light_dome.glow_toward).
-function ldTent(arr: number[], azDeg: number): number {
-  const p = (((azDeg % 360) + 360) % 360) / 45
-  const lo = Math.floor(p) % 8
-  const hi = (lo + 1) % 8
-  const f = p - Math.floor(p)
-  return arr[lo] * (1 - f) + arr[hi] * f
-}
-
-// Mirrors light_dome.glow_toward(): score(az) / (1 + (alt/θ(az))²)
-function glowToward(summary: LightDomeSummary, azDeg: number, altDeg: number): number {
-  const scores8  = LD_DIRS.map(d => summary.scores[d] ?? 0)
-  const heights8 = LD_DIRS.map(d => summary.dome_heights[d] ?? 0)
-  const score    = ldTent(scores8, azDeg)
-  const theta    = ldTent(heights8, azDeg)
-  const alt      = Math.max(0, altDeg)
-  if (theta <= 0) return alt === 0 ? score : 0
-  return score / (1 + (alt / theta) ** 2)
-}
-
-function glowLabel(g: number): string {
-  if (g < 0.03)     return 'negligible'
-  if (g < LD_MINOR) return 'minor'
-  if (g < LD_MAJOR) return 'moderate'
-  return 'major'
-}
-
-// CSS colour for a glow value, reusing the LD stop palette.
-// Negligible glow returns {} so the element inherits var(--text-dim) from CSS;
-// the LD ramp starts at black (the zero-glow "excellent" end) which is invisible
-// on dark backgrounds, so we only apply inline colour once the glow is meaningful.
-function glowStyle(g: number): React.CSSProperties {
-  if (g < 0.03) return {}
-  const [r, gr, b] = ldColor(g)
-  return { color: `rgb(${Math.round(r)},${Math.round(gr)},${Math.round(b)})` }
-}
-
-// Sky background brightness for MW arch visibility — distinct from glowToward().
-// dome_heights are geometric angles (1-3° for a city 30+ mi away) but city glow
-// scatters through the atmosphere and degrades sky brightness well above that.
-// A 40° characteristic altitude matches observer perception: score at the horizon
-// falls to ~50% at 40°, ~20% at 80° (zenith). Uses direction scores only, no heights.
-function archGlowAt(summary: LightDomeSummary, azDeg: number, altDeg: number): number {
-  const scores8 = LD_DIRS.map(d => summary.scores[d] ?? 0)
-  const score   = ldTent(scores8, azDeg)
-  const alt     = Math.max(0, altDeg)
-  return score / (1 + (alt / 40) ** 2)
-}
-
-// ── Milky Way brightness model ────────────────────────────────────────────────
-// Two-component empirical model: compact Gaussian bulge + linear disk + floor.
-// sigma=0.28 chosen to match known bright regions (core→scutum→cygnus→anticenter).
-function intrinsicBrightness(l_deg: number): number {
-  const norm  = ((l_deg % 360) + 360) % 360
-  const delta = Math.min(norm, 360 - norm)    // 0° at core, 180° at anticenter
-  const x     = delta / 180                    // normalised [0, 1]
-  const bulge = 0.70 * Math.exp(-x * x / (2 * 0.28 * 0.28))
-  const disk  = 0.18 * (1 - x * 0.7)
-  return Math.max(0, Math.min(1, 0.12 + bulge + disk))
-}
-
-// Exponential attenuation from a light-dome glow index → brightness multiplier [0,1].
-// glow=0→1.0, glow=0.25(LD_MINOR)→0.82, glow=1.0→0.45, glow=3.0(LD_MAJOR)→0.09
-function washoutFactor(glow: number): number {
-  return Math.exp(-0.8 * glow)
-}
-
-// Per-segment arch brightness: intrinsic galactic profile × light-dome washout.
-// Returns {glowOpacity, coreOpacity} for the two rendering layers.
-function archSegmentBrightness(
-  l: number, alt: number, az: number,
-  lightDome: LightDomeSummary | null
-): { glowOpacity: number; coreOpacity: number } {
-  const I    = intrinsicBrightness(l)
-  const glow = lightDome ? archGlowAt(lightDome, az, alt) : 0
-  const W    = washoutFactor(glow)
-  const B    = I * W
-  return { glowOpacity: B * 0.35, coreOpacity: B * 0.85 }
-}
-
-// Orthographic projection of the sky dome, camera centered on the galactic core azimuth.
-// The arch traces the galactic equator (b=0), sampled every 5° of galactic longitude.
-function MilkyWayDome({ summary, waypoints, report }: { summary: MilkyWaySummary; waypoints: VisibleTarget[]; report: NightReport }) {
-  const [heading, setHeading] = useState<number>(
-    summary.core_peak_az_deg != null ? Math.round(summary.core_peak_az_deg) : 180
-  );
-  const [tilt, setTilt] = useState<number>(0);
-  const pointerRef = useRef<{ x: number; y: number } | null>(null);
-  const svgRef = useRef<SVGSVGElement>(null);
-
-  // 1. ADDED HERE: State to track which dot is currently being hovered
-  const [hoveredDot, setHoveredDot] = useState<{name: string, x: number, y: number} | null>(null);
-
-  // Native touch handlers for iOS Safari — PointerEvents + setPointerCapture can be
-  // unreliable on iOS when touch-action isn't applied before the first touch. Native
-  // listeners with passive:false let us call preventDefault() to block page scroll.
-  useEffect(() => {
-    const svg = svgRef.current;
-    if (!svg) return;
-    const fConst = 100 / Math.tan(60 * Math.PI / 180);
-    let lastX = 0, lastY = 0;
-    const onTouchStart = (e: TouchEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      lastX = e.touches[0].clientX;
-      lastY = e.touches[0].clientY;
-    };
-    const onTouchMove = (e: TouchEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      if (!e.touches[0]) return;
-      const dx = e.touches[0].clientX - lastX;
-      const dy = e.touches[0].clientY - lastY;
-      lastX = e.touches[0].clientX;
-      lastY = e.touches[0].clientY;
-      const rect = svg.getBoundingClientRect();
-      const sens = (180 / Math.PI) / fConst;
-      setHeading(h => ((h + dx * (180 / rect.width) * sens) % 360 + 360) % 360);
-      setTilt(t => Math.max(0, Math.min(45, t - dy * (120 / rect.height) * sens)));
-    };
-    svg.addEventListener('touchstart', onTouchStart, { passive: false });
-    svg.addEventListener('touchmove', onTouchMove, { passive: false });
-    return () => {
-      svg.removeEventListener('touchstart', onTouchStart);
-      svg.removeEventListener('touchmove', onTouchMove);
-    };
-  }, []);
-
-  if (summary.core_peak_alt_deg == null || summary.core_peak_alt_deg <= 0) {
-    return (
-      <div className="mw-dome-absent">
-        Arch below horizon tonight
-        <div className="mw-absent-reason">The Galactic Core is currently out of view.</div>
-      </div>
-    );
-  }
-
-  const toRad = (deg: number) => (deg * Math.PI) / 180;
-  const peakTimeMs = new Date(summary.core_peak_time).getTime();
-
-  type EqSample = { l: number; alt: number; az: number; }
-  type ArchSegment = { x1: number; y1: number; x2: number; y2: number; glowOpacity: number; coreOpacity: number }
-
-  // Galactic-to-horizontal conversion depends only on location and peak time —
-  // memoised so it does NOT recompute on every heading/tilt drag frame.
-  const allEquatorSamples = useMemo<EqSample[]>(() =>
-    Array.from({ length: 72 }, (_, i) => {
-      const l = i * 5;
-      const { alt, az } = galToAltAz(l, 0, report.lat, report.lon, peakTimeMs);
-      return { l, alt, az };
-    }),
-    [report.lat, report.lon, peakTimeMs]
-  );
-
-  // Arch segment brightness only depends on sky position + light dome — memoised
-  // separately so it survives heading/tilt state changes.
-  const equatorBrightness = useMemo(() =>
-    allEquatorSamples.map(s =>
-      archSegmentBrightness(s.l, s.alt, s.az, report.light_dome ?? null)
-    ),
-    [allEquatorSamples, report.light_dome]
-  );
-
-  // Gnomonic (rectilinear/perspective) projection — models a wide-angle camera
-  // pointed at the horizon along `heading`. Great circles project to straight lines.
-  const FOV_HALF_DEG = 60;                        // 120° total horizontal FoV
-  const f = 100 / Math.tan(toRad(FOV_HALF_DEG)); // focal length ~57.7 SVG units
-
-  const project = (alt: number, az: number) => {
-    const altR = toRad(alt);
-    const azR  = toRad(az - heading);
-    const tiltR = toRad(tilt);
-    const cosAlt = Math.cos(altR), sinAlt = Math.sin(altR);
-    const cosAzR = Math.cos(azR), sinAzR = Math.sin(azR);
-    const cosT = Math.cos(tiltR), sinT = Math.sin(tiltR);
-    const dx = cosAlt * sinAzR;
-    const dy = sinAlt * cosT - cosAlt * cosAzR * sinT;
-    const dz = cosAlt * cosAzR * cosT + sinAlt * sinT;
-    if (dz <= 0) return { x: 100, y: 100, isFront: false };
-    return {
-      x: 100 + f * (dx / dz),
-      y: 100 - f * (dy / dz),
-      isFront: true,
-    };
-  };
-
-  // Horizon y-position in SVG coords — moves down as camera tilts up; clips to frame bottom.
-  const horizonY = Math.min(120, 100 + f * Math.tan(toRad(tilt)));
-
-  // Map the known Milky Way waypoint names to their exact Galactic Longitude (l)
-  const WAYPOINT_L: Record<string, number> = {
-    'Galactic Anticenter': 180,
-    'Cassiopeia/Perseus': 135,
-    'Cepheus Cloud': 105,
-    'Cygnus Star Cloud': 80,
-    'Aquila Rift': 45,
-    'Scutum Star Cloud': 27,
-    'Galactic Core': 0,
-    'Scorpius Star Cloud': 347,
-    'Norma Star Cloud': 330,
-    'Crux & Coalsack': 302,
-    'Carina Nebula & Cloud': 287,
-    'Vela Supernova Region': 265,
-    'Puppis Star Cloud': 245,
-    'Monoceros': 210,
-  };
-
-  // Waypoint alt/az is also stable wrt heading/tilt — memoised separately.
-  const wpDotsRaw = useMemo(() =>
-    waypoints
-      .map(wp => {
-        const l = WAYPOINT_L[wp.name];
-        if (l != null) {
-          const { alt, az } = galToAltAz(l, 0, report.lat, report.lon, peakTimeMs);
-          return { name: wp.name, alt, az };
-        }
-        const w = bestWindow(wp);
-        return { name: wp.name, alt: w.peak_alt_deg ?? -1, az: w.peak_az_deg ?? 0 };
-      })
-      .filter(p => p.alt > 0),
-    [waypoints, report.lat, report.lon, peakTimeMs]
-  );
-
-  // Project waypoints using current heading/tilt (view-dependent).
-  const wpDots = wpDotsRaw.map(p => ({ ...p, proj: project(p.alt, p.az) }));
-
-  const doubled = [...allEquatorSamples, ...allEquatorSamples];
-  let longestStreak: EqSample[] = [];
-  let currentStreak: EqSample[] = [];
-  for (const s of doubled) {
-    if (s.alt > -2) currentStreak.push(s);
-    else {
-      if (currentStreak.length > longestStreak.length) longestStreak = currentStreak;
-      currentStreak = [];
-    }
-  }
-  const visibleArc = longestStreak.slice(0, 73).map(s => ({ ...s, proj: project(s.alt, s.az) }))
-
-  // Build per-segment brightness using the pre-computed equatorBrightness cache.
-  // l is always a multiple of 5 in [0,355], so index = (l/5) % 72.
-  const visibleSegments: ArchSegment[] = []
-  for (let i = 0; i < visibleArc.length - 1; i++) {
-    const a = visibleArc[i], b = visibleArc[i + 1]
-    if (!a.proj.isFront || !b.proj.isFront) continue
-    const bA = equatorBrightness[Math.round(a.l / 5) % 72]
-    const bB = equatorBrightness[Math.round(b.l / 5) % 72]
-    visibleSegments.push({
-      x1: a.proj.x, y1: a.proj.y,
-      x2: b.proj.x, y2: b.proj.y,
-      glowOpacity: (bA.glowOpacity + bB.glowOpacity) / 2,
-      coreOpacity: (bA.coreOpacity + bB.coreOpacity) / 2,
-    })
-  }
-  const corePos = project(summary.core_peak_alt_deg, summary.core_peak_az_deg ?? 0);
-
-  // Altitude reference rings for photographer framing (20° and 40°).
-  // In gnomonic, constant-altitude loci curve upward at the edges — rendered as polylines.
-  const ALT_RINGS = [20, 40] as const;
-  const ringPolylines = ALT_RINGS.map(alt => {
-    const pts: string[] = [];
-    for (let dAz = -FOV_HALF_DEG; dAz <= FOV_HALF_DEG; dAz += 1) {
-      const p = project(alt, heading + dAz);
-      if (p.isFront) pts.push(`${p.x.toFixed(1)},${p.y.toFixed(1)}`);
-    }
-    return { alt, points: pts.join(' ') };
-  });
-
-  // Left-edge label anchor for each ring — sampled ~54° left of center so x≈20 near frame edge.
-  const ringLabels = ALT_RINGS.map(alt => {
-    const p = project(alt, heading - 54);
-    if (!p.isFront || p.x < 11 || p.x > 100) return null;
-    return { alt, x: p.x, y: p.y };
-  });
-
-  // Sky glow blobs: one per dome direction, anchored 5° above horizon so the gradient
-  // peak sits just inside the dome rather than exactly on the arc edge.
-  type DomeGlow = { dir: Direction; x: number; y: number; r: number; op: number }
-  const domeGlows: DomeGlow[] = report.light_dome
-    ? LD_DIRS.flatMap(d => {
-        const score = report.light_dome!.scores[d] ?? 0
-        if (score < LD_MINOR) return []
-        const pos = project(5, LD_DIR_AZ[d])
-        if (!pos.isFront) return []
-        const r  = Math.min(85, 30 + 18 * Math.log1p(score))
-        const op = Math.min(0.50, 0.12 + 0.15 * Math.log1p(score))
-        return [{ dir: d, x: pos.x, y: pos.y, r, op }]
-      })
-    : []
-
-  // Moon glow blob: position at arch peak time, brightness scales with illumination.
-  // Only shown when moon is above the horizon and meaningfully illuminated (≥5%).
-  const moonGlowPos = (() => {
-    if (report.illumination_pct < 5) return null
-    const { alt, az } = moonAltAz(report.lat, report.lon, peakTimeMs)
-    if (alt <= 0) return null
-    const pos = project(alt, az)
-    if (!pos.isFront) return null
-    const illumFrac = report.illumination_pct / 100
-    return {
-      x: pos.x, y: pos.y,
-      r:  Math.min(70, 15 + 55 * illumFrac),
-      op: Math.min(0.45, 0.05 + 0.40 * illumFrac),
-    }
-  })()
-
-  const cardinals = [
-    { deg: 0, label: 'N' }, { deg: 45, label: 'NE' }, { deg: 90, label: 'E' },
-    { deg: 135, label: 'SE' }, { deg: 180, label: 'S' }, { deg: 225, label: 'SW' },
-    { deg: 270, label: 'W' }, { deg: 315, label: 'NW' }
-  ];
-
-  // 1px drag → (180/displayWidth) SVG units → (180/π)/f degrees of heading/tilt change.
-  const handlePointerDown = (e: React.PointerEvent<SVGSVGElement>) => {
-    pointerRef.current = { x: e.clientX, y: e.clientY };
-    (e.currentTarget as SVGSVGElement).setPointerCapture(e.pointerId);
-  };
-  const handlePointerMove = (e: React.PointerEvent<SVGSVGElement>) => {
-    if (!pointerRef.current) return;
-    const dx = e.clientX - pointerRef.current.x;
-    const dy = e.clientY - pointerRef.current.y;
-    pointerRef.current = { x: e.clientX, y: e.clientY };
-    const rect = (e.currentTarget as SVGSVGElement).getBoundingClientRect();
-    const sens = (180 / Math.PI) / f;                    // ~1°/SVG unit
-    const dxSvg = dx * (180 / rect.width) * sens;        // heading degrees
-    const dySvg = dy * (120 / rect.height) * sens;        // tilt degrees
-    setHeading(h => ((h + dxSvg) % 360 + 360) % 360);
-    setTilt(t => Math.max(0, Math.min(45, t - dySvg)));
-  };
-  const handlePointerUp = () => { pointerRef.current = null; };
-
-  return (
-    <div className="mw-dome-wrap">
-      <div className="mw-dome-title">360° Sky Dome</div>
-      <svg
-        ref={svgRef}
-        viewBox="10 0 180 120"
-        xmlns="http://www.w3.org/2000/svg"
-        className="mw-dome-svg"
-        onPointerDown={handlePointerDown}
-        onPointerMove={handlePointerMove}
-        onPointerUp={handlePointerUp}
-        onPointerLeave={handlePointerUp}
-      >
-        <defs>
-          <clipPath id="mw-half-dome-clip">
-            <rect x="0" y="0" width="200" height={horizonY} />
-          </clipPath>
-          <filter id="mw-f-band" x="-50%" y="-50%" width="200%" height="200%">
-            <feGaussianBlur in="SourceGraphic" stdDeviation="3" />
-          </filter>
-          {domeGlows.map(g => (
-            <radialGradient key={g.dir} id={`mw-ldg-${g.dir}`}
-              cx={g.x} cy={g.y} r={g.r} gradientUnits="userSpaceOnUse">
-              <stop offset="0%"   stopColor="currentColor" stopOpacity={g.op} />
-              <stop offset="100%" stopColor="currentColor" stopOpacity="0" />
-            </radialGradient>
-          ))}
-          {moonGlowPos && (
-            <radialGradient id="mw-moon-g"
-              cx={moonGlowPos.x} cy={moonGlowPos.y} r={moonGlowPos.r} gradientUnits="userSpaceOnUse">
-              <stop offset="0%"   stopColor="currentColor" stopOpacity={moonGlowPos.op} />
-              <stop offset="100%" stopColor="currentColor" stopOpacity="0" />
-            </radialGradient>
-          )}
-        </defs>
-
-        <rect x="0" y="0" width="200" height="120" fill="rgba(10, 15, 30, 0.4)" />
-        {horizonY < 120 && (
-          <rect className="mw-dome-ground" x="0" y={horizonY} width="200" height={120 - horizonY} />
-        )}
-
-        {/* Sky glow from light domes — color controlled via .mw-dome-glow for red-mode compliance */}
-        {domeGlows.length > 0 && (
-          <g className="mw-dome-glow" clipPath="url(#mw-half-dome-clip)">
-            {domeGlows.map(g => (
-              <circle key={g.dir} cx={g.x} cy={g.y} r={g.r}
-                fill={`url(#mw-ldg-${g.dir})`} />
-            ))}
-            {domeGlows.map(g => (
-              <circle key={`${g.dir}-hit`} cx={g.x} cy={g.y} r="20"
-                fill="transparent" pointerEvents="all" style={{ cursor: 'pointer' }}
-                onPointerDown={(e) => e.stopPropagation()}
-                onMouseEnter={() => setHoveredDot({ name: `${g.dir} sky glow`, x: g.x, y: g.y })}
-                onMouseLeave={() => setHoveredDot(null)} />
-            ))}
-          </g>
-        )}
-
-        {/* Moon glow — color controlled via .mw-moon-glow for red-mode compliance */}
-        {moonGlowPos && (
-          <g className="mw-moon-glow" clipPath="url(#mw-half-dome-clip)">
-            <circle cx={moonGlowPos.x} cy={moonGlowPos.y} r={moonGlowPos.r} fill="url(#mw-moon-g)" />
-            <circle cx={moonGlowPos.x} cy={moonGlowPos.y} r="2" fill="currentColor" opacity="0.75" />
-            <circle cx={moonGlowPos.x} cy={moonGlowPos.y} r="18"
-              fill="transparent" pointerEvents="all" style={{ cursor: 'pointer' }}
-              onPointerDown={(e) => e.stopPropagation()}
-              onMouseEnter={() => setHoveredDot({ name: 'Moon', x: moonGlowPos.x, y: moonGlowPos.y })}
-              onMouseLeave={() => setHoveredDot(null)} />
-          </g>
-        )}
-
-        <g clipPath="url(#mw-half-dome-clip)">
-          {/* GLOW LAYER: blurred wide band; filter on the group blends joints between segments */}
-          <g filter="url(#mw-f-band)" className="mw-arch-glow">
-            {visibleSegments.map((seg, i) => (
-              <line
-                key={i}
-                x1={seg.x1} y1={seg.y1} x2={seg.x2} y2={seg.y2}
-                stroke="currentColor"
-                strokeWidth="14"
-                strokeOpacity={seg.glowOpacity}
-              />
-            ))}
-          </g>
-          {/* CORE LAYER: thin bright stripe, variable opacity per segment */}
-          <g className="mw-arch-core">
-            {visibleSegments.map((seg, i) => (
-              <line
-                key={i}
-                x1={seg.x1} y1={seg.y1} x2={seg.x2} y2={seg.y2}
-                stroke="currentColor"
-                strokeWidth="1.5"
-                strokeOpacity={seg.coreOpacity}
-              />
-            ))}
-          </g>
-          {/* Altitude reference rings — perspective curves at 20° and 40° */}
-          {ringPolylines.map(ring => ring.points && (
-            <polyline key={ring.alt} className="mw-dome-ring" points={ring.points} fill="none" />
-          ))}
-          {ringLabels.map(lbl => lbl && (
-            <text key={`rl-${lbl.alt}`} className="mw-dome-ring-label"
-              x={lbl.x + 2} y={lbl.y - 2} textAnchor="start">{lbl.alt}°</text>
-          ))}
-        </g>
-
-        {/* 2. REPLACED HERE: The Galactic Core */}
-        {corePos.isFront && (
-          <g>
-            <circle className="mw-dome-core" cx={corePos.x} cy={corePos.y} r="3.5" pointerEvents="none" />
-            <circle
-              cx={corePos.x} cy={corePos.y} r="12" fill="transparent" pointerEvents="all" style={{ cursor: 'pointer' }}
-              onPointerDown={(e) => e.stopPropagation()}
-              onMouseEnter={() => setHoveredDot({ name: 'Galactic Core', x: corePos.x, y: corePos.y })}
-              onMouseLeave={() => setHoveredDot(null)}
-            />
-          </g>
-        )}
-
-        {/* 3. REPLACED HERE: Waypoints Array */}
-        {wpDots.map((wp, i) => (
-          wp.proj.isFront && (
-            <g key={i}>
-              <circle cx={wp.proj.x} cy={wp.proj.y} r="2" className="mw-dome-waypoint" pointerEvents="none" />
-              <circle
-                cx={wp.proj.x} cy={wp.proj.y} r="12" fill="transparent" pointerEvents="all" style={{ cursor: 'pointer' }}
-                onPointerDown={(e) => e.stopPropagation()}
-                onMouseEnter={() => setHoveredDot({ name: wp.name, x: wp.proj.x, y: wp.proj.y })}
-                onMouseLeave={() => setHoveredDot(null)}
-              />
-            </g>
-          )
-        ))}
-
-        {horizonY < 120 && (
-          <line className="mw-dome-horizon" x1="0" y1={horizonY} x2="200" y2={horizonY} />
-        )}
-        <rect className="mw-dome-frame" x="10.5" y="0.5" width="179" height="119" fill="none" />
-
-        {horizonY < 120 && cardinals.map(c => {
-          let relAz = c.deg - heading;
-          while (relAz <= -180) relAz += 360;
-          while (relAz > 180) relAz -= 360;
-
-          if (Math.abs(relAz) < FOV_HALF_DEG) {
-            // Gnomonic + tilt: horizon objects at azimuth relAz project to:
-            // x = 100 + f * tan(relAz) / cos(tilt)
-            const x = 100 + f * Math.tan(toRad(relAz)) / Math.cos(toRad(tilt));
-            const labelY = horizonY + 14;
-            if (x < 11 || x > 189 || labelY > 119) return null;
-            return (
-              <g key={c.label}>
-                <line className="mw-dome-tick" x1={x} y1={horizonY} x2={x} y2={horizonY + 3} />
-                <text className="mw-dome-label" x={x} y={labelY} textAnchor="middle">{c.label}</text>
-              </g>
-            );
-          }
-          return null;
-        })}
-
-        {/* 4. ADDED HERE: The Custom Tooltip directly inside the SVG */}
-        {hoveredDot && (
-          <foreignObject x={hoveredDot.x - 75} y={hoveredDot.y - 35} width="150" height="30" pointerEvents="none">
-            <div style={{ display: 'flex', justifyContent: 'center', alignItems: 'flex-end', height: '100%' }}>
-              <div style={{
-                background: 'rgba(8, 13, 28, 0.95)',
-                color: 'rgba(200, 212, 238, 0.95)',
-                border: '1px solid rgba(90, 120, 200, 0.25)',
-                borderRadius: '2px',
-                padding: '3px 7px',
-                fontSize: '7px',
-                fontFamily: 'var(--mono)',
-                letterSpacing: '0.05em',
-                textTransform: 'uppercase',
-                whiteSpace: 'nowrap',
-                boxShadow: '0 2px 8px rgba(0,0,0,0.8)'
-              }}>
-                {hoveredDot.name}
-              </div>
-            </div>
-          </foreignObject>
-        )}
-      </svg>
-      <div className="mw-dome-subtitle">Light pollution vs. Milky Way plane</div>
-    </div>
-  );
-}
-
-function MilkyWayAbsent({ report: r }: { report: NightReport }) {
-  const coreMaxAlt = Math.max(0, 90 - Math.abs(r.lat - (-28.9)))
-  const bortle     = r.light_pollution?.bortle_class ?? 0
-  const bortleDesc = r.light_pollution?.bortle_desc  ?? 'bright sky'
-
-  let reason: string
-  if (coreMaxAlt < 10) {
-    reason = `Galactic core never rises above 10° from this latitude (max ${coreMaxAlt.toFixed(0)}° altitude)`
-  } else if (bortle >= 6) {
-    reason = `${bortleDesc} (Bortle ${bortle}) — light pollution prevents Milky Way visibility here`
-  } else if (bortle >= 4) {
-    reason = `Suburban skies (Bortle ${bortle}) are too bright for Milky Way visibility here`
-  } else if (r.dark_intervals.length === 0) {
-    reason = `Bright moon (${r.illumination_pct.toFixed(0)}%) is up all night — no dark sky window`
-  } else {
-    reason = 'Galactic core is below the horizon during tonight\'s dark window'
-  }
-
-  return <p className="mw-absent-reason">{reason}</p>
-}
-
-export function MilkyWayCard({ summary, waypoints, report }: {
-  summary: MilkyWaySummary
-  waypoints: VisibleTarget[]
-  report: NightReport
-}) {
-  const tz = report.tz_name
-  const s  = summary
-
-  const archQuality = s.arch_angle_deg != null
-    ? (s.arch_angle_deg >= 60 ? 'steep' : s.arch_angle_deg >= 35 ? 'moderate' : 'flat')
-    : null
-
-  // Directions where dome glow visibly dims the arch at peak time.
-  // Uses archGlowAt (40° characteristic alt) at each dome direction within ±90° of the
-  // core az; flags it when the resulting glow ≥ LD_MINOR (≥18% brightness reduction).
-  const domeSections: { dir: Direction; glow: number }[] = (() => {
-    if (!report.light_dome || s.core_peak_az_deg == null) return []
-    const coreAz   = s.core_peak_az_deg
-    const proxyAlt = Math.max(5, s.core_peak_alt_deg ?? 25)
-    return LD_DIRS.flatMap(d => {
-      const score = report.light_dome!.scores[d] ?? 0
-      if (score < LD_MINOR) return []
-      const dirAz = LD_DIR_AZ[d]
-      let delta = ((dirAz - coreAz) + 360) % 360
-      if (delta > 180) delta = 360 - delta
-      if (delta > 90) return []
-      const glow = archGlowAt(report.light_dome!, dirAz, proxyAlt)
-      if (glow < LD_MINOR) return []
-      return [{ dir: d, glow }]
-    })
-  })()
-
-  const bestLabel = 'Best time'
-  const bestTime  = s.best_viewing_time ?? (s.core_peak_in_window ? s.core_peak_time : s.arch_end)
-
-  const moonSeverity = moonWashSeverity(
-    report.illumination_pct,
-    s.core_moon_sep_deg ?? null,
-    s.core_moon_alt_deg ?? null,
-  )
-
-      return (
-    <div className="mw-card">
-      <div className="mw-meta-block">
-        {/* Unified Score Row */}
-        <div className="meta-row">
-          <span className="meta-k">Score</span>
-            {/* Using the standard meta-v class for uniformity */}
-          <span className={`meta-v mw-score mw-score-band-${scoreBand(s.local_score)}`}>
-            {s.local_score.toFixed(1)}
-          <span className="mw-score-denom">{scoreLabel(s.local_score)}</span>
-            {s.weather_blocked && <span className="mw-moon-badge badge-poor" style={{marginLeft: 8}}>Clouded out</span>}
-            {!s.weather_blocked && s.weather_limited && <span className="mw-moon-badge" style={{marginLeft: 8}}>Partly cloudy</span>}
-  </span>
-</div>
-        {/* Unified Metadata List */}
-        <div className="meta-row">
-          <span className="meta-k">Arch window</span>
-          <span className="meta-v">
-            {formatTime(s.arch_start, tz)} – {formatTime(s.arch_end, tz)}
-            {'  ·  '}{Math.floor(s.arch_hours)}h {Math.round((s.arch_hours % 1) * 60).toString().padStart(2,'0')}m
-            {s.moon_limited && !s.arch_moon_washout && <MoonBadge type="limited" severity={moonSeverity} />}
-            {s.weather_limited && !s.weather_blocked && <span className="mw-moon-badge">{`${s.clear_arch_hours.toFixed(1)}h clear`}</span>}
-          </span>
-        </div>
-        <div className="meta-row">
-          <span className="meta-k">Galactic core</span>
-          <span className="meta-v">
-            {fmtPos(s.core_peak_alt_deg, s.core_peak_az_deg)} (max {s.core_max_alt_deg}° alt)
-            {archQuality && s.arch_angle_deg != null && `  ·  arch ${s.arch_angle_deg.toFixed(0)}° (${archQuality})`}
-          </span>
-        </div>
-        <div className="meta-row">
-          <span className="meta-k">{bestLabel}</span>
-          <span className="meta-v">
-            {formatTime(bestTime, tz)} — core @ {fmtPos(s.core_peak_alt_deg, s.core_peak_az_deg)}
-          </span>
-        </div>
-      </div>
-
-      {/* New Title Centered Above Both */}
-      <div className="mw-group-title">360° Sky View</div>
-
-      {/* 3 & 4: Skydome (Left) and Notes (Right) */}
-      <div className="mw-mid-section">
-
-        <div className="mw-dome-container">
-          <MilkyWayDome summary={s} waypoints={waypoints} report={report} />
-        </div>
-
-        <div className="mw-notes-container">
-          {s.moon_penalised && !s.arch_moon_washout && <MoonBadge type="penalty" severity={moonSeverity} />}
-          {s.arch_moon_washout && <span className="mw-moon-badge">Moon washout</span>}
-          {domeSections.length > 0 && (() => {
-            const maxGlow  = Math.max(...domeSections.map(ds => ds.glow))
-            return (
-              <span className="mw-moon-badge cond-glow" style={glowStyle(maxGlow)}>
-                {`Dome glow: ${glowLabel(maxGlow)}`}
-              </span>
-            )
-          })()}
-        </div>
-      </div>
-
-      <div className="mw-bars-section telemetry-mini-bars">
-        <ScoreBar label="Altitude" value={s.alt_score} />
-        <ScoreBar label="Coverage" value={s.cov_score} />
-        <ScoreBar label="Window" value={s.win_score} />
-      </div>
-
-      {waypoints.length > 0 && (
-        <WaypointsAccordion waypoints={waypoints} summary={s} report={report} />
-      )}
-    </div>
-  )
-}
-
-function bestWindow(t: VisibleTarget): TargetWindow {
-  const clean = t.windows.filter(w => !w.moon_interference)
-  const pool  = clean.length ? clean : t.windows
-  return pool.reduce((best, w) => (w.peak_alt_deg ?? 0) > (best.peak_alt_deg ?? 0) ? w : best)
-}
-
-// Sky condition at a given ISO time, incorporating K&S moon wash (mirrors CLI)
-function skyCondition(
-  peakIso: string,
-  darkIntervals: [string, string][],
-  nightStart: string | null,
-  nightEnd: string | null,
-  illuminationPct: number,
-  moonrise: string | null,
-  moonset:  string | null,
-  moonSepAtPeak:  number | null,
-  moonAltAtPeak:  number | null,
-): string {
-  const pt = new Date(peakIso).getTime()
-
-  let base = 'Twilight'
-  for (const [s, e] of darkIntervals) {
-    if (pt >= new Date(s).getTime() && pt <= new Date(e).getTime()) { base = 'Dark sky'; break }
-  }
-  if (base === 'Twilight' && nightStart && nightEnd) {
-    const ns = new Date(nightStart).getTime(), ne = new Date(nightEnd).getTime()
-    if (pt >= ns && pt <= ne) base = 'Astro night'
-  }
-
-  if (moonUpAt(peakIso, moonrise, moonset)) {
-    const sev = moonWashSeverity(illuminationPct, moonSepAtPeak, moonAltAtPeak)
-    if (sev) return `Moon wash: ${sev}`
-  }
-  return base
-}
-
-// Interpolate altitude at a clipped time (mirrors _alt_at in render_report.py)
-function altAt(cutoffIso: string, w: TargetWindow): number {
-  if (!w.peak_time || w.peak_alt_deg == null) return w.start_alt_deg
-  const t     = new Date(cutoffIso).getTime()
-  const tPeak = new Date(w.peak_time).getTime()
-  const t0    = t <= tPeak ? new Date(w.start).getTime()    : tPeak
-  const a0    = t <= tPeak ? w.start_alt_deg                : w.peak_alt_deg
-  const t1    = t <= tPeak ? tPeak                          : new Date(w.end).getTime()
-  const a1    = t <= tPeak ? w.peak_alt_deg                 : w.end_alt_deg
-  const frac  = t1 > t0 ? (t - t0) / (t1 - t0) : 0.5
-  return Math.round(a0 + Math.max(0, Math.min(1, frac)) * (a1 - a0))
-}
-
-// Mirrors targets.is_prime(): DSOs need ≥40° peak alt, planets ≥20°, all need ≥1h window.
-// Falls back to dark-interval overlap when all windows are moon-interfered.
-function isPrime(t: VisibleTarget, darkIntervals: [string, string][]): boolean {
-  const MIN_ALT = 40, PLANET_MIN_ALT = 20, MIN_HRS = 1.0
-  const effAlt = t.type === 'planet' ? PLANET_MIN_ALT : MIN_ALT
-  const clean = t.windows.filter(w => !w.moon_interference)
-
-  if (clean.length === 0) {
-    if (darkIntervals.length === 0) return false
-    for (const w of t.windows) {
-      const ws = new Date(w.start).getTime(), we = new Date(w.end).getTime()
-      for (const [ds, de] of darkIntervals) {
-        const oS = Math.max(ws, new Date(ds).getTime())
-        const oE = Math.min(we, new Date(de).getTime())
-        if ((oE - oS) / 3_600_000 >= MIN_HRS) {
-          return t.type === 'milky_way' || (w.peak_alt_deg ?? 0) >= effAlt
-        }
-      }
-    }
-    return false
-  }
-
-  const best = clean.reduce((a, b) => (b.peak_alt_deg ?? 0) > (a.peak_alt_deg ?? 0) ? b : a)
-  const durH = (new Date(best.end).getTime() - new Date(best.start).getTime()) / 3_600_000
-  return t.type === 'milky_way'
-    ? durH >= MIN_HRS
-    : (best.peak_alt_deg ?? 0) >= effAlt && durH >= MIN_HRS
-}
-
-// ── Meteor shower card ───────────────────────────────────────────────────────
-
-function skyClass(sky: string): string {
-  const moonWashMatch = sky.match(/^Moon wash: (minor|moderate|severe)$/)
-  if (moonWashMatch) return `tg-sky-moon-wash-${moonWashMatch[1]}`
-  return `tg-sky-${sky.replace(/ /g, '-').toLowerCase()}`
-}
-
-function MeteorShowerCard({ target, zhr, report }: {
-  target: VisibleTarget
-  zhr: number
-  report: NightReport
-}) {
-  const tz = report.tz_name
-  const w  = bestWindow(target)
-
-  const sky = w.peak_time
-    ? skyCondition(
-        w.peak_time, report.dark_intervals, report.night_start, report.night_end,
-        report.illumination_pct, report.moonrise, report.moonset,
-        w.moon_sep_at_peak_deg, w.moon_alt_at_peak_deg,
-      )
-    : null
-  const skyCls = sky ? skyClass(sky) : ''
-
-  return (
-    <div className="ms-card">
-      <div className="ms-header-row">
-        <span className="ms-name">{target.name} Meteor Shower</span>
-        <span className="ms-zhr">
-          <InfoTip tip={<>ZHR — zenithal hourly rate: meteors per hour for a single observer under a perfectly dark sky with the radiant overhead. Field counts run well below it.</>}>
-            Peak ZHR {zhr}
-          </InfoTip>
-        </span>
-      </div>
-      {target.note && <div className="ms-note">{target.note}</div>}
-      {w.peak_time && w.peak_alt_deg != null && (
-        <>
-          <div className="mw-row">
-            <span className="mw-label">Best viewing</span>
-            <span>{formatTime(w.peak_time, tz)} @ {fmtPos(w.peak_alt_deg, w.peak_az_deg)}</span>
-          </div>
-          <div className="mw-row">
-            <span className="mw-label">Window</span>
-            <span>
-              {formatTime(w.start, tz)} – {formatTime(w.end, tz)}
-              {'  ·  '}{w.start_alt_deg.toFixed(0)}° → {w.end_alt_deg.toFixed(0)}°
-            </span>
-          </div>
-          {sky && (
-            <div className="mw-row">
-              <span className="mw-label">Sky</span>
-              <span className={`tg-sky ${skyCls}`}>{sky}</span>
-            </div>
-          )}
-        </>
-      )}
-    </div>
-  )
-}
-
-// ── Blocker badge (Phase 2) ───────────────────────────────────────────────────
-
-function blockerLabel(blockers: string[]): string {
-  if (blockers.includes('cloud') || blockers.includes('transparency')) return 'Clouded out'
-  if (blockers.includes('moon_washout')) return 'Moon washout'
-  if (blockers.includes('light_dome')) return 'Lost in light dome'
-  return 'Unavailable Tonight'
-}
-
-function BlockerBadge({ blockers }: { blockers: string[] }) {
-  return <span className="tg-blocker-badge">{blockerLabel(blockers)}</span>
-}
-
-function clipTooltip(w: TargetWindow, tz: string): string {
-  const b = w.blockers ?? []
-  const end = w.effective_end
-  if (b.includes('cloud') || b.includes('transparency'))
-    return `Partly cloudy${end ? ` after ${formatTime(end, tz)}` : ''}`
-  if (b.includes('moon_washout')) return 'Moon washout'
-  if (b.includes('light_dome'))   return 'Viewing constrained by horizon glow'
-  return 'Window clipped by conditions'
-}
-
-/*function clipReasonShort(w: TargetWindow): string {
-  const b = w.blockers ?? []
-  if (b.includes('cloud') || b.includes('transparency')) return 'cloud'
-  if (b.includes('moon_washout')) return 'moon'
-  if (b.includes('light_dome'))   return 'dome'
-  return 'conditions'
-}*/
-
-// ── TargetsTable ──────────────────────────────────────────────────────────────
-
-function TargetsTable({ targets, report }: { targets: VisibleTarget[]; report: NightReport }) {
-  const tz = report.tz_name
-  // Blocked targets collapse to one summary row by default — on a bad night a
-  // wall of identical "Clouded out" rows is noise, not signal.
-  const [showBlocked, setShowBlocked] = useState(false)
-
-  // Milky Way + meteor showers rendered separately as cards; rest filtered to prime prominent
-  const allPrime = targets
-    .filter(t => t.type !== 'milky_way' && t.type !== 'meteor_shower')
-    .filter(t => isPrime(t, report.dark_intervals))
-    .filter(t => (t.landscape_suitability ?? 'prominent') === 'prominent')
-
-  const viable   = allPrime.filter(t => t.viability !== 'blocked')
-  const unviable = allPrime.filter(t => t.viability === 'blocked')
-
-  if (allPrime.length === 0) return null
-
-  // Shared sort + group logic.
-  // nebula / galaxy / cluster are normalized to 'dso' so they render as one unlabeled block.
-  type RowItem =
-    | { kind: 'header'; type: string; key: string }
-    | { kind: 'target'; target: VisibleTarget; key: string; blocked?: boolean }
-
-  const displayType = (t: VisibleTarget) => DSO_TYPES.has(t.type) ? 'dso' : t.type
-
-  function sortAndGroup(list: VisibleTarget[], blocked = false): RowItem[] {
-    // Pre-compute bestWindow once per target so the sort comparator doesn't
-    // call it O(n log n) times — sort only reads from this map.
-    const bwMap = new Map(list.map(t => [t, bestWindow(t)]))
-    const sorted = [...list].sort((a, b) => {
-      const ao = TYPE_ORDER[displayType(a)] ?? 99
-      const bo = TYPE_ORDER[displayType(b)] ?? 99
-      if (ao !== bo) return ao - bo
-      const at = bwMap.get(a)!.peak_time ?? ''
-      const bt = bwMap.get(b)!.peak_time ?? ''
-      return at.localeCompare(bt)
-    })
-    const groups: { type: string; targets: VisibleTarget[] }[] = []
-    for (const t of sorted) {
-      const dt   = displayType(t)
-      const last = groups[groups.length - 1]
-      if (last && last.type === dt) last.targets.push(t)
-      else groups.push({ type: dt, targets: [t] })
-    }
-    const rows: RowItem[] = []
-    for (const g of groups) {
-      if (TYPE_LABELS[g.type]) {
-        rows.push({ kind: 'header', type: g.type, key: `hdr-${blocked ? 'blocked-' : ''}${g.type}` })
-      }
-      for (const t of g.targets) {
-        rows.push({ kind: 'target', target: t, key: `${g.type}-${t.name}`, blocked })
-      }
-    }
-    return rows
-  }
-
-  const viableRows   = sortAndGroup(viable, false)
-  const unviableRows = sortAndGroup(unviable, true)
-
-  function renderTargetRow(t: VisibleTarget, key: string, blocked: boolean) {
-    const w         = bestWindow(t)
-    const name      = t.type === 'meteor_shower' ? `${t.name} Meteor Shower` : t.name
-    const sizeLabel = moonScaleLabel(t.angular_size_arcmin)
-
-    // photo_cutoff clips both Peak and Window (mirrors CLI)
-    const hasClip = !!(w.photo_cutoff
-      && new Date(w.photo_cutoff) > new Date(w.start)
-      && new Date(w.photo_cutoff) < new Date(w.end))
-
-    // Fixed-width helpers — each piece gets a min-width span so columns
-    // stay consistent across rows regardless of digit count or direction length.
-    const Tt  = ({ t }: { t: string })    => <span className="tg-t">{formatTime(t, tz)}</span>
-    const Alt = ({ deg }: { deg: number }) => <span className="tg-alt">{Math.round(deg)}°</span>
-    const Az  = ({ az }: { az: number })   => <span className="tg-az">{Math.round(az)}°</span>
-    const Dir = ({ az }: { az: number })   => <span className="tg-dir">{cardinal(az)}</span>
-    const Sep = () => <span className="tg-p"> – </span>
-
-    const peakForSky = blocked
-      ? null
-      : (w.best_time ?? (hasClip ? w.photo_cutoff! : w.peak_time))
-
-    // Conditions: weather icon and glow inline in Target cell
-    const wxPt = peakForSky && !report.wx_no_data && !report.wx_pending
-      ? wxAtTime(report.weather_points || [], peakForSky)
-      : null
-    const glow = report.light_dome && w.peak_alt_deg != null
-      ? glowToward(report.light_dome, w.peak_az_deg, w.peak_alt_deg)
-      : null
-
-    const targetCell = (
-      <td>
-        {name}
-        {t.note    && <span className="tg-note"> · {t.note}</span>}
-        {sizeLabel && <span className="tg-note"> · Size: {sizeLabel}</span>}
-        {glow != null && glow >= 0.03 && (
-          <span className="tg-glow-inline cond-glow" style={glowStyle(glow)}>
-            {` · glow ${glowLabel(glow)}`}
-          </span>
-        )}
-      </td>
-    )
-
-    if (blocked) {
-      return (
-        <tr key={key} className="tg-row-blocked">
-          {targetCell}
-          <td><BlockerBadge blockers={t.windows[0]?.blockers ?? []} /></td>
-          <td></td>
-          <td className="wx-num">—</td>
-        </tr>
-      )
-    }
-
-    // Peak cell: time · Alt · Az · Dir + sky badge when non-dark
-    let peakJsx: React.ReactNode = '—'
-    if (w.peak_time && w.peak_alt_deg != null) {
-      const effectiveBestTime = w.best_time ?? (hasClip ? w.photo_cutoff! : w.peak_time)
-      const effectiveBestAlt  = altAt(effectiveBestTime, w)
-
-      const sky = peakForSky
-        ? skyCondition(
-            peakForSky, report.dark_intervals, report.night_start, report.night_end,
-            report.illumination_pct, report.moonrise, report.moonset,
-            w.moon_sep_at_peak_deg, w.moon_alt_at_peak_deg,
-          )
-        : '—'
-      const skyCls = skyClass(sky)
-      const moonNote = w.moon_interference && !sky.startsWith('Moon')
-      const moonIsUpAtPeak = peakForSky
-        ? moonUpAt(peakForSky, report.moonrise, report.moonset)
-        : false
-      const moonNoteText = moonNote
-        ? (moonIsUpAtPeak ? ' · moon wash minimal' : ' · pre-moonrise')
-        : null
-      const showSkyBadge = sky !== 'Dark sky' && sky !== '—'
-
-      peakJsx = (
-        <>
-          <Tt t={effectiveBestTime} />
-          <span className="tg-p"> · Alt </span><Alt deg={effectiveBestAlt} />
-          <span className="tg-p"> Az </span><Az az={w.peak_az_deg} /><span className="tg-p"> </span><Dir az={w.peak_az_deg} />
-          {showSkyBadge && (
-            <span className={`tg-sky-inline ${skyCls}`}>
-              {' '}{sky}{moonNoteText ? <span className="tg-moon-note">{moonNoteText}</span> : null}
-            </span>
-          )}
-          {!showSkyBadge && moonNoteText && (
-            <span className="tg-moon-note">{' '}{moonNoteText}</span>
-          )}
-        </>
-      )
-    }
-
-    // Window: use effective_start/effective_end from condition vectors when available
-    let winJsx: React.ReactNode = '—'
-    if (w.peak_time) {
-      const rawEnd    = hasClip ? w.photo_cutoff! : w.end
-      const effStart  = w.effective_start ?? w.start
-      const effEnd    = w.effective_end   ?? rawEnd
-      const effStartAlt = altAt(effStart, w)
-      const effEndAlt   = altAt(effEnd, w)
-      const isClipped = effStart > w.start || effEnd < rawEnd
-
-      // Visual-observation extension beyond photo_cutoff (preserved from original logic)
-      const visualExtJsx = hasClip && w.visual_cutoff ? (() => {
-        const extraMin = Math.round(
-          (new Date(w.visual_cutoff).getTime() - new Date(w.photo_cutoff!).getTime()) / 60000
-        )
-        return extraMin >= 10 ? <span className="tg-p"> +{extraMin}m visual</span> : null
-      })() : null
-
-      winJsx = (
-        <>
-          <Tt t={effStart} /><span className="tg-p"> @ </span><Alt deg={effStartAlt} />
-          <Sep />
-          <Tt t={effEnd} /><span className="tg-p"> @ </span><Alt deg={effEndAlt} />
-          {isClipped && (
-            <span
-              className="tg-clip-indicator"
-              title={clipTooltip(w, tz)}
-            >*</span>
-          )}
-          {visualExtJsx}
-        </>
-      )
-    }
-
-    return (
-      <tr key={key}>
-        {targetCell}
-        <td className="wx-num">{peakJsx}</td>
-        <td className="wx-num tg-cond-col">
-          {wxPt && (
-            <span className={`tg-wx-inline wx-rating-${scoreBand(rateConditions(wxPt))}`}>
-              <WmoIcon code={wxPt.weather_code} size={12} />
-            </span>
-          )}
-        </td>
-        <td className="wx-num">{winJsx}</td>
-      </tr>
-    )
-  }
-
-  return (
-    <div className="tg-table-wrap">
-      <table className="tg-table">
-        <thead>
-          <tr>
-            <th>Target</th>
-            <th>Peak</th>
-            <th></th>
-            <th>Window</th>
-          </tr>
-        </thead>
-        <tbody>
-          {viableRows.map(row => {
-            if (row.kind === 'header') {
-              return (
-                <tr key={row.key} className="tg-group-hdr">
-                  <td colSpan={4}>{TYPE_LABELS[row.type] ?? row.type}</td>
-                </tr>
-              )
-            }
-            return renderTargetRow(row.target, row.key, false)
-          })}
-
-          {unviable.length > 0 && (
-            <>
-              <tr className="tg-unviable-hdr">
-                <td colSpan={4}>
-                  <button
-                    type="button"
-                    className="tg-blocked-toggle"
-                    aria-expanded={showBlocked}
-                    onClick={() => setShowBlocked(v => !v)}
-                  >
-                    <span className="tg-blocked-caret" aria-hidden="true">{showBlocked ? '▾' : '▸'}</span>
-                    {`Unavailable Tonight (${unviable.length})`}
-                    <span className="tg-blocked-counts">
-                      {' — '}
-                      {(() => {
-                        const counts = new Map<string, number>()
-                        for (const t of unviable) {
-                          const label = blockerLabel(t.windows[0]?.blockers ?? [])
-                          counts.set(label, (counts.get(label) ?? 0) + 1)
-                        }
-                        return [...counts.entries()]
-                          .sort((a, b) => b[1] - a[1])
-                          .map(([label, n]) => (n > 1 ? `${label} ×${n}` : label))
-                          .join(' · ')
-                      })()}
-                    </span>
-                  </button>
-                </td>
-              </tr>
-              {showBlocked && unviableRows.map(row => {
-                if (row.kind === 'header') {
-                  return (
-                    <tr key={row.key} className="tg-group-hdr tg-row-blocked">
-                      <td colSpan={4}>{TYPE_LABELS[row.type] ?? row.type}</td>
-                    </tr>
-                  )
-                }
-                return renderTargetRow(row.target, row.key, true)
-              })}
-            </>
-          )}
-        </tbody>
-      </table>
-    </div>
-  )
-}
-
-// ── Nearby dark-sky results ──────────────────────────────────────────────────
-
-function nearbyBortleClass(bortleClass: number | null): string {
-  if (bortleClass == null) return 'nearby-bortle'
-  let colorClass = 'nearby-bortle-excellent'
-  if (bortleClass <= 2) colorClass = 'nearby-bortle-excellent'
-  else if (bortleClass <= 4) colorClass = 'nearby-bortle-good'
-  else if (bortleClass <= 6) colorClass = 'nearby-bortle-fair'
-  else colorClass = 'nearby-bortle-poor'
-  return `nearby-bortle ${colorClass}`
-}
-
-function NearbyResults(
-  { data, imperial, originLat, originLon }:
-  { data: NearbyResult; imperial: boolean; originLat: number; originLon: number },
-) {
-  const { origin_bortle, origin_sqm, radius_miles, results, light_domes, best_available } = data
-  const sqmStr = origin_sqm != null ? ` (SQM ${origin_sqm.toFixed(1)})` : ''
-
-  // Convert stored miles to the active unit system
-  const fmtMi = (mi: number) => fmtDist(mi * 1.60934, imperial)
-  // Prefer actual road distance from the routing API; fall back to straight-line when a
-  // candidate wasn't routed (raw "Remote" fallbacks).
-  const distOf = (p: NearbyPlace) => fmtMi(p.drive_miles ?? p.distance_miles)
-  // Google Maps driving directions, origin → location (falls back to a place pin).
-  const dirLink = (p: NearbyPlace) =>
-    `https://www.google.com/maps/dir/?api=1&origin=${originLat},${originLon}` +
-    `&destination=${p.lat},${p.lon}&travelmode=driving`
-
-  const placeStr = (p: NearbyPlace) =>
-    p.name ?? `${p.lat.toFixed(2)}°, ${p.lon.toFixed(2)}°`
-  const formatDriveTime = (minutes: number | null): string | null => {
-    if (minutes == null) return null
-    const hrs = Math.floor(minutes / 60)
-    const mins = minutes % 60
-    return hrs > 0 ? `${hrs} hr ${mins} min` : `${mins} min`
-  }
-  const POI_TYPE_LABEL: Record<string, string> = {
-    parking: 'Parking', viewpoint: 'Viewpoint', camp_site: 'Campsite', rest_area: 'Rest area',
-    caravan_site: 'RV park', picnic_site: 'Picnic area', ranger_station: 'Ranger station',
-    observatory: 'Observatory', attraction: 'Attraction', information: 'Info point',
-    tourism: 'Tourism', pier: 'Pier', lighthouse: 'Lighthouse', tower: 'Observation tower',
-    summer_camp: 'Summer camp', firepit: 'Fire pit', beach_resort: 'Beach resort',
-    historic: 'Historic site',
-  }
-  // Render a place name with a category badge (routable POIs) or a "Remote" tag (off-road
-  // fallbacks), plus a Google Maps driving-directions link on every result.
-  const placeNode = (p: NearbyPlace) => {
-    const appLink = `?lat=${p.lat.toFixed(5)}&lon=${p.lon.toFixed(5)}`
-    return (
-      <>
-        <a className="poi-namelink" href={appLink}>{placeStr(p)}</a>
-        {p.area_name && <span className="poi-area">{p.area_name}</span>}
-        <span className="poi-type-link">
-          {p.is_poi
-            ? (p.poi_type && <span className="poi-badge">{POI_TYPE_LABEL[p.poi_type] ?? p.poi_type}</span>)
-            : <span className="poi-remote">Remote</span>}
-          <a className="poi-maplink" href={dirLink(p)} target="_blank" rel="noopener noreferrer" aria-label="Directions"><Navigation size={12} strokeWidth={2} /></a>
-        </span>
-      </>
-    )
-  }
-
-  return (
-    <>
-      <div className="meta-row">
-        <span className="meta-k">Origin:</span>
-        <span className="meta-v"><span className={nearbyBortleClass(origin_bortle)}>Bortle {origin_bortle}</span>{sqmStr}  ·  {fmtMi(radius_miles)} radius</span>
-      </div>
-
-      {/* 1. Note when already at Bortle 1 — results still shown below */}
-      {origin_bortle <= 1 && results.length > 0 && (
-        <p className="sat-notice">
-          Already at Bortle {origin_bortle}{sqmStr} — showing other Bortle 1 sites within {fmtMi(radius_miles)}.
-        </p>
-      )}
-
-      {/* 2. Empty state */}
-      {results.length === 0 && (
-        <p className="sat-notice">
-          {origin_bortle <= 1
-            ? `No other Bortle 1 sites found within ${fmtMi(radius_miles)}.`
-            : `No significantly darker sky found within ${fmtMi(radius_miles)}.`
-          }
-          {best_available && origin_bortle > 1 && (
-            <> Closest darker spot: <span className={nearbyBortleClass(best_available.bortle_class)}>Bortle {best_available.bortle_class}</span>, {distOf(best_available)} {best_available.direction}{formatDriveTime(best_available.drive_minutes) ? ` · ${formatDriveTime(best_available.drive_minutes)} drive` : ''}  ({placeNode(best_available)})</>
-          )}
-        </p>
-      )}
-
-      {/* 3. Results table */}
-      {results.length > 0 && (() => {
-        const hasDrive = results.some(p => p.drive_minutes != null)
-
-        // New Tiered Drive-Time Sort
-        const sortedByDriveTime = [...results].sort((a, b) => {
-          const bothHaveDrive = a.drive_minutes != null && b.drive_minutes != null;
-
-          // Group pristine skies (Bortle 1 & 2) at the top
-          const aIsTopTier = a.bortle_class <= 2;
-          const bIsTopTier = b.bortle_class <= 2;
-
-          if (aIsTopTier !== bIsTopTier) {
-            return aIsTopTier ? -1 : 1;
-          }
-
-          // Sort by drive time (or distance fallback) within tiers
-          if (bothHaveDrive) {
-            if (a.drive_minutes !== b.drive_minutes) {
-              return a.drive_minutes! - b.drive_minutes!;
-            }
-          } else {
-            if (a.distance_miles !== b.distance_miles) {
-              return a.distance_miles - b.distance_miles;
-            }
-          }
-
-          // Tie-breaker: Darkest sky
-          return a.bortle_class - b.bortle_class;
-        });
-
-        const nearest = sortedByDriveTime[0];
-
-        // Keep darkest calculation as-is (strict Bortle-first sort)
-        const darkest = [...results].sort((a, b) =>
-          a.bortle_class !== b.bortle_class ? a.bortle_class - b.bortle_class : a.distance_miles - b.distance_miles
-        )[0]
-
-        const showDarkest = darkest !== nearest && darkest.bortle_class < nearest.bortle_class
-
-        return (
-          <>
-            <div className="nearby-highlights">
-              <div className="nearby-highlight-row">
-                <span className="nearby-highlight-label">Nearest</span>
-                <span><span className={nearbyBortleClass(nearest.bortle_class)}>Bortle {nearest.bortle_class}</span>  ·  {distOf(nearest)} {nearest.direction}{formatDriveTime(nearest.drive_minutes) ? `  ·  ${formatDriveTime(nearest.drive_minutes)} drive` : ''}  ({placeNode(nearest)})</span>
-              </div>
-              {showDarkest && (
-                <div className="nearby-highlight-row">
-                  <span className="nearby-highlight-label">Darkest</span>
-                  <span><span className={nearbyBortleClass(darkest.bortle_class)}>Bortle {darkest.bortle_class}</span>  ·  {distOf(darkest)} {darkest.direction}{formatDriveTime(darkest.drive_minutes) ? `  ·  ${formatDriveTime(darkest.drive_minutes)} drive` : ''}  ({placeNode(darkest)})</span>
-                </div>
-              )}
-            </div>
-            <div className="wx-table-wrap">
-              <table className="wx-table nearby-table">
-                <thead>
-                  <tr>
-                    <th className="nearby-area-th">Area</th>
-                    <th>Bortle</th>
-                    <th>SQM</th>
-                    <th>Dist</th>
-                    {hasDrive && <th>Drive</th>}
-                    <th>Dir</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {[...results]
-                    // Order by drive time, lowest first; unrouted (no ETA) last, then by distance.
-                    .sort((a, b) => {
-                      const ad = a.drive_minutes, bd = b.drive_minutes
-                      if (ad == null && bd == null) return a.distance_miles - b.distance_miles
-                      if (ad == null) return 1
-                      if (bd == null) return -1
-                      return ad - bd || a.bortle_class - b.bortle_class
-                    })
-                    .map((p, i) => (
-                      <tr key={i}>
-                        <td className={`${nearbyBortleClass(p.bortle_class)} nearby-area-td`}>
-                          <div className="nearby-area-inner">{placeNode(p)}</div>
-                        </td>
-                        <td className={`wx-num nearby-bortle-col ${nearbyBortleClass(p.bortle_class)}`}>{p.bortle_class}</td>
-                        <td className="wx-num nearby-sqm-col">{p.sqm != null ? p.sqm.toFixed(1) : '—'}</td>
-                        <td className="wx-num nearby-dist-col">{distOf(p)}</td>
-                        {hasDrive && <td className="wx-num nearby-drive-col">{formatDriveTime(p.drive_minutes) ?? '—'}</td>}
-                        <td className="wx-num nearby-dir-col">{p.direction}</td>
-                      </tr>
-                    ))}
-                </tbody>
-              </table>
-            </div>
-          </>
-        )
-      })()}
-
-      {/* 4. ALWAYS show domes if they exist, regardless of origin Bortle */}
-      {light_domes.length > 0 && (
-        <div className="nearby-domes">
-          <div className="nearby-domes-label">Light domes</div>
-          {light_domes.map((d, i) => (
-            <div key={i} className="nearby-dome-row">
-              <span className="nearby-dome-name">{placeStr(d)}</span>
-              <span className={`nearby-dome-bortle ${nearbyBortleClass(d.bortle_class)}`}>Bortle {d.bortle_class}</span>
-              <span className="nearby-dome-dist">{fmtMi(d.distance_miles)} {d.direction}</span>
-            </div>
-          ))}
-        </div>
-      )}
-    </>
-  )
-}
-
-// ── Light dome (all-sky fisheye) ─────────────────────────────────────────────
-// An all-sky heatmap of horizon light pollution: centre = zenith (dark), rim = the
-// 360° horizon, N at top. Each direction's horizon glow blooms upward by that
-// dome's apparent height, so a distant low metro dome hugs the rim while a near one
-// reaches higher. Mirrors the engine: glow(az,alt) = score(az)/(1+(alt/θ(az))²)
-// (PyNightSkyPredictor/light_dome.py glow_toward).
-// LD constants and utility functions (LD_DIRS, ldTent, glowToward, etc.) are defined
-// earlier in this file, before MilkyWayDome, so archSegmentBrightness can use them.
-
-// Nearest hourly weather point to a given ISO time.
-function wxAtTime(points: WeatherPoint[], isoTime: string): WeatherPoint | null {
-  if (!points.length) return null
-  const t = new Date(isoTime).getTime()
-  return points.reduce((a, b) =>
-    Math.abs(new Date(a.time).getTime() - t) <= Math.abs(new Date(b.time).getTime() - t) ? a : b)
-}
-
-// Commenting out.
-/* Compact conditions badge: WMO weather icon + rating score + horizon glow label.
-// Used in both the prime targets table and the Milky Way waypoints table.
-function CondBadges({ wxPt, glow }: { wxPt: WeatherPoint | null; glow: number | null }) {
-  const rating = wxPt != null ? rateConditions(wxPt) : null
-  return (
-    <span className="cond-badges">
-      {wxPt != null && rating != null && (
-        <span className={`cond-badge wx-rating-${scoreBand(rating)}`}>
-          <WmoIcon code={wxPt.weather_code} size={13} />
-        </span>
-      )}
-      {glow != null && glow > 0 && (
-        <span className="cond-badge cond-glow" style={glowStyle(glow)}>
-          {glow >= 0.03 ? `Light dome ${glowLabel(glow)}` : null}
-        </span>
-      )}
-    </span>
-  )
-}*/
-
-function LightDomePanel({ summary, imperial }: { summary: LightDomeSummary; imperial: boolean }) {
-  const canvasRef = useRef<HTMLCanvasElement>(null)
-  const panelRef = useRef<HTMLDivElement>(null)
-  // Disk size; matched to the score card's content (the meta block) so the panel
-  // doesn't make the card taller. Capped at LD_SIZE; falls back to it pre-measure.
-  const [size, setSize] = useState(LD_SIZE)
-  const { sky_state, scores, darkest_direction, domes } = summary
-  // The darkest horizon is only a meaningful "point here" call when a darker side exists.
-  const showBest = sky_state === 'dark' || sky_state === 'domed'
-
-  useEffect(() => {
-    const panel = panelRef.current
-    if (!panel) return
-    // ResizeObserver on the panel itself (flex: 0 0 100% — outer size is CSS-controlled,
-    // not affected by canvas content changes, so no feedback loop).
-    let rafId = 0
-    const measure = () => {
-      const body = panel.querySelector('.ld-body') as HTMLElement | null
-      const isRow = !!body && getComputedStyle(body).flexDirection === 'row'
-      const caption = panel.querySelector('.ld-caption') as HTMLElement | null
-      const captionW = isRow ? (caption?.offsetWidth ?? 130) : 0
-      const gap     = isRow ? 18 : 0
-      const avail   = panel.clientWidth - captionW - gap
-      setSize(Math.max(88, Math.min(LD_SIZE, Math.round(avail))))
-    }
-    // Debounce ResizeObserver: skip intermediate firings during layout reflow,
-    // coalescing them into one rAF-deferred measurement per animation frame.
-    const onResize = () => {
-      cancelAnimationFrame(rafId)
-      rafId = requestAnimationFrame(measure)
-    }
-    const ro = new ResizeObserver(onResize)
-    ro.observe(panel)
-    return () => { ro.disconnect(); cancelAnimationFrame(rafId) }
-  }, [])
-
-  useEffect(() => {
-    const canvas = canvasRef.current
-    if (!canvas) return
-    const ctx = canvas.getContext('2d')
-    if (!ctx) return
-
-    const dpr = window.devicePixelRatio || 1
-    const W = Math.round(size * dpr)
-    canvas.width = W
-    canvas.height = W
-    canvas.style.width = `${size}px`
-    canvas.style.height = `${size}px`
-
-    // Per-direction glow and apparent dome height (real height for flagged domes).
-    const domeH: Partial<Record<Direction, number>> = {}
-    for (const d of domes) domeH[d.direction] = d.dome_height_deg
-    const scoreArr = LD_DIRS.map(d => scores[d] ?? 0)
-    const thetaArr = LD_DIRS.map(d =>
-      domeH[d] != null ? Math.max(domeH[d]! * LD_THETA_K, LD_THETA_FLOOR_DEG) : LD_THETA_DEFAULT_DEG)
-
-    // Pixel pass over the disk (device resolution; transform-independent putImageData).
-    const margin = Math.max(11, size * 0.095)   // room for the N/E/S/W labels
-    const cx = W / 2, cy = W / 2
-    const R = (size / 2 - margin) * dpr
-    const img = ctx.createImageData(W, W)
-    const buf = img.data
-    for (let y = 0; y < W; y++) {
-      for (let x = 0; x < W; x++) {
-        const i = (y * W + x) * 4
-        const dx = x - cx, dy = y - cy
-        const rr = Math.sqrt(dx * dx + dy * dy)
-        if (rr > R) { buf[i + 3] = 0; continue }
-        const alt = 90 * (1 - rr / R)                       // centre = zenith, rim = horizon
-        const az = (Math.atan2(dx, -dy) * 180 / Math.PI + 360) % 360  // N up, clockwise
-        const sc = ldTent(scoreArr, az)
-        const th = ldTent(thetaArr, az)
-        const g = sc / (1 + (alt / th) ** 2)
-        const [r, gg, b] = ldColor(g)
-        const edge = Math.max(0, Math.min(1, (R - rr) / (1.5 * dpr)))  // soft rim AA
-        buf[i] = r; buf[i + 1] = gg; buf[i + 2] = b; buf[i + 3] = 255 * edge
-      }
-    }
-    ctx.putImageData(img, 0, 0)
-
-    // Decorations in CSS px.
-    ctx.scale(dpr, dpr)
-    const c = size / 2
-    const cssR = size / 2 - margin
-    ctx.lineWidth = 1
-    for (const rad of [cssR * 0.5, cssR * 0.83]) {
-      ctx.beginPath(); ctx.arc(c, c, rad, 0, Math.PI * 2)
-      ctx.strokeStyle = 'rgba(148,163,184,0.12)'; ctx.stroke()
-    }
-    ctx.beginPath(); ctx.arc(c, c, cssR, 0, Math.PI * 2)
-    ctx.strokeStyle = 'rgba(148,163,184,0.22)'; ctx.stroke()
-
-    const lab = margin * 0.45
-    ctx.fillStyle = '#94A3B8'
-    ctx.font = '600 10px Poppins, system-ui, sans-serif'
-    ctx.textAlign = 'center'; ctx.textBaseline = 'middle'
-    ctx.fillText('N', c, lab); ctx.fillText('S', c, size - lab)
-    ctx.fillText('E', size - lab, c); ctx.fillText('W', lab, c)
-
-    if (showBest) {
-      const a = (LD_DIRS.indexOf(darkest_direction) * 45) * Math.PI / 180
-      ctx.fillStyle = '#34D399'
-      ctx.font = '700 12px Poppins, system-ui, sans-serif'
-      ctx.fillText('★', c + Math.sin(a) * (cssR - 8), c - Math.cos(a) * (cssR - 8))
-    }
-  }, [summary, scores, darkest_direction, domes, sky_state, showBest, size])
-
-  const fmtMi = (mi: number) => fmtDist(mi * 1.60934, imperial)
-  const top = domes[0]
-  const topDist = top?.mean_distance_mi != null ? `  ·  ${fmtMi(top.mean_distance_mi)}` : ''
-
-  const aria =
-    sky_state === 'urban' ? 'Urban sky: horizon washed out in all directions.'
-    : sky_state === 'bright' ? `Bright sky: uniform glow, darkest horizon to the ${darkest_direction}.`
-    : sky_state === 'domed' ? `${top?.label ?? 'Light dome'}. Darkest horizon to the ${darkest_direction}.`
-    : `Dark sky. Darkest horizon to the ${darkest_direction}.`
-
-  return (
-    <div className="lightdome-panel" ref={panelRef}>
-      <div className="ld-title">Horizon Glow</div>
-      <div className="ld-body">
-      <canvas ref={canvasRef} className="ld-canvas" role="img" aria-label={aria} />
-      <div className="ld-caption">
-        <span className={`ld-state ld-state-${sky_state}`}>
-          {sky_state === 'dark' ? 'Dark sky'
-            : sky_state === 'domed' ? 'Light dome'
-            : sky_state === 'bright' ? 'Bright sky'
-            : 'Urban sky'}
-        </span>
-        {sky_state === 'domed' && (
-          <>
-            <span className={`ld-line${top?.severity ? ` ld-dome-${top.severity}` : ''}`}>{top?.label}{topDist}</span>
-            <span className="ld-sub">Best view <b>{darkest_direction}</b></span>
-          </>
-        )}
-        {sky_state === 'dark' && (
-          <span className="ld-line">Darkest horizon <b>{darkest_direction}</b></span>
-        )}
-        {sky_state === 'bright' && (
-          <>
-            <span className="ld-line">Uniform glow, no single dome</span>
-            <span className="ld-sub">Darkest <b>{darkest_direction}</b>, still washed</span>
-          </>
-        )}
-        {sky_state === 'urban' && (
-          <span className="ld-line">Washed out in all directions</span>
-        )}
-        <span className="ld-legend" aria-hidden="true">
-          <span>dark</span><span className="ld-legend-bar" /><span>dome</span>
-        </span>
-      </div>
-      </div>
-    </div>
-  )
 }
 
 // ── Main report card ─────────────────────────────────────────────────────────
