+++ apps/web/src/styles/09-empty-state.css
@@ -1,0 +1,170 @@
+/* ── empty state ──────────────────────────────────────────── */
+.empty-state {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  animation: report-in 0.35s cubic-bezier(0.16, 1, 0.3, 1) both;
+}
+@media (prefers-reduced-motion: reduce) {
+  .empty-state { animation: none; }
+}
+
+.es-copy {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 8px;
+  text-align: center;
+}
+.es-headline {
+  margin: 0;
+  font-size: 15px;
+  font-weight: 600;
+  color: var(--text-h);
+  letter-spacing: -0.01em;
+}
+.es-body {
+  margin: 0;
+  font-size: 13px;
+  color: var(--text-dim);
+  line-height: 1.65;
+  max-width: 56ch;
+}
+
+.es-score-scale {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+}
+.es-score-bar {
+  height: 5px;
+  border-radius: 3px;
+  background: linear-gradient(
+    to right,
+    var(--poor),
+    var(--fair) 33%,
+    var(--good) 66%,
+    var(--excellent)
+  );
+  opacity: 0.72;
+}
+.es-score-labels {
+  display: flex;
+  justify-content: space-between;
+  font-size: 10px;
+  font-weight: 500;
+  letter-spacing: 0.02em;
+  color: var(--text-dim);
+  opacity: 0.6;
+}
+
+.es-divider {
+  height: 0;
+  border: none;
+  border-top: 1px solid rgba(0, 0, 0, 0.14);
+  border-bottom: 1px solid rgba(255, 255, 255, 0.80);
+}
+
+.es-quickstart {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+.es-quickstart-label {
+  font-size: 11px;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.07em;
+  color: var(--text-dim);
+  opacity: 0.65;
+}
+.es-quickstart-btns {
+  display: flex;
+  gap: 8px;
+  flex-wrap: wrap;
+}
+.es-qs-btn {
+  flex: 1 1 180px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 3px;
+  padding: 10px 14px;
+  border: 1px solid var(--card-border);
+  border-radius: 2px;
+  background: rgba(255, 255, 255, 0.025);
+  color: var(--text);
+  font: inherit;
+  font-size: 13px;
+  text-align: left;
+  cursor: pointer;
+  transition: color 0.15s, border-color 0.15s, background 0.15s, box-shadow 0.15s;
+}
+.es-qs-btn:hover {
+  color: var(--text-h);
+  border-color: var(--accent);
+  background: rgba(var(--tint-rgb), 0.07);
+  box-shadow: 0 0 0 1px rgba(var(--tint-rgb), 0.08);
+}
+.es-qs-btn:active {
+  transform: scale(0.98);
+}
+.es-qs-name {
+  font-weight: 600;
+  color: var(--text-h);
+}
+.es-qs-hook {
+  font-size: 11px;
+  line-height: 1.5;
+  color: var(--text-dim);
+}
+
+/* ── capability strip (empty state) ─────────────────────────── */
+.es-caps-label {
+  display: block;
+  text-align: center;
+  font-size: 11px;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.07em;
+  color: var(--text-dim);
+  opacity: 0.65;
+  margin-bottom: -6px;
+}
+html.red-mode .es-caps-label { color: var(--text-dim); }
+.es-caps {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 10px 22px;
+}
+.es-cap {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  min-width: 0;
+}
+.es-cap-k {
+  font-family: var(--mono);
+  font-size: 10px;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.07em;
+  color: var(--accent);
+}
+.es-cap-v {
+  font-size: 12px;
+  line-height: 1.55;
+  color: var(--text-dim);
+}
+/* Table-stakes sweep: one quiet line under the differentiator tiles. */
+.es-caps-more {
+  margin: -6px auto 0;
+  font-size: 11.5px;
+  line-height: 1.6;
+  color: var(--text-dim);
+  opacity: 0.85;
+  max-width: 72ch;
+  text-align: center;
+}
+html.red-mode .es-cap-k { color: var(--accent); }
+html.red-mode .es-qs-hook, html.red-mode .es-cap-v, html.red-mode .es-caps-more { color: var(--text-dim); }
